Manually Add Roles

add_role can add a role definition to an existing variable in the recipe.

add_role(recipe, vars, role = "predictor", data = NULL)

Arguments

recipe

An existing recipe.

vars

A character string of variable names or simple formula that defines one or more variable that will be assigned to the role.

role

A character string for a single role.

data

Am optional data template that may be needed to resolve the formula (in case there is a .).

Value

An updated recipe object.

Details

If vars is a formula, it should not include any functions, negative signs, or . (since there is no data element to evaluate it).

Examples

data(biomass)

# Create the recipe manually
rec <- recipe(x = biomass)
rec

#> Error in `colnames<-`(`*tmp*`, value = c("role", "#variables")): 'names' attribute [2] must be the same length as the vector [1]

rec$var_info

#> # A tibble: 8 × 4
#>   variable    type  role   source
#>      <chr>   <chr> <lgl>    <chr>
#> 1   sample nominal    NA original
#> 2  dataset nominal    NA original
#> 3   carbon numeric    NA original
#> 4 hydrogen numeric    NA original
#> 5   oxygen numeric    NA original
#> 6 nitrogen numeric    NA original
#> 7   sulfur numeric    NA original
#> 8      HHV numeric    NA original

rec <- add_role(rec, vars = colnames(biomass)[3:7], role = "predictor")
rec <- add_role(rec, vars = "sample", role = "id variable")
rec <- add_role(rec, vars = "dataset", role = "splitting variable")
rec <- add_role(rec, vars = "HHV", role = "outcome")
rec

#> Data Recipe
#> 
#> Inputs:
#> 
#>                role #variables
#>         id variable          1
#>             outcome          1
#>           predictor          5
#>  splitting variable          1

# or, if you really want the formula method, you can change roles

rec2 <- recipe(HHV ~ ., data = biomass)
# Now change inappropriate predictor roles
rec2 <- add_role(rec2, vars = "sample", role = "id variable")

#> Warning: Overwriting role(s) forsample

rec2 <- add_role(rec2, vars = "dataset", role = "splitting variable")

#> Warning: Overwriting role(s) fordataset

rec2$var_info

#> # A tibble: 8 × 4
#>   variable    type               role   source
#>      <chr>   <chr>              <chr>    <chr>
#> 1   sample nominal        id variable original
#> 2  dataset nominal splitting variable original
#> 3   carbon numeric          predictor original
#> 4 hydrogen numeric          predictor original
#> 5   oxygen numeric          predictor original
#> 6 nitrogen numeric          predictor original
#> 7   sulfur numeric          predictor original
#> 8      HHV numeric            outcome original