check_cols {recipes}	R Documentation

Check if all Columns are Present

Description

check_cols creates a specification of a recipe step that will check if all the columns of the training frame are present in the new data.

Usage

check_cols(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  skip = FALSE,
  id = rand_id("cols")
)

## S3 method for class 'check_cols'
tidy(x, ...)

Arguments

recipe
A recipe object. The check will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are checked in the check See selections() for more details. For the tidy method, these are not currently used.

role
Not used by this check since no new variables are created.

trained
A logical for whether the selectors in ... have been resolved by prep().

skip
A logical. Should the check be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id
A character string that is unique to this step to identify it.

x
A check_cols object.

Details

This check will break the bake function if any of the specified columns is not present in the data. If the check passes, nothing is changed to the data.

Examples

library(modeldata)
data(biomass)

biomass_rec <- recipe(HHV ~ ., data = biomass) %>%
   step_rm(sample, dataset) %>%
   check_cols(contains("gen")) %>%
   step_center(all_predictors())

## Not run: 
bake(biomass_rec, biomass[, c("carbon", "HHV")])

## End(Not run)

[Package recipes version 0.1.15 Index]