check_name {recipes}	R Documentation

check that newly created variable names don't overlap

Description

check_name is to be used in the bake function to ensure that newly created variable names don't overlap with existing names. Throws an error if check fails.

Usage

check_name(res, new_data, object, newname = NULL, names = FALSE)

Arguments

res
A data frame or tibble of the newly created variables.

new_data
A data frame or tibble passed to the bake function.

object
A trained object passed to the bake function.

newname
A string of variable names if prefix isn't specified in the trained object.

names
A logical determining if the names should be set using the names function (TRUE) or colnames function (FALSE).

[Package recipes version 0.1.15 Index]