check_range {recipes}	R Documentation

Check Range Consistency

Description

check_range creates a specification of a recipe check that will check if the range of a numeric variable changed in the new data.

Usage

check_range(
  recipe,
  ...,
  role = NA,
  skip = FALSE,
  trained = FALSE,
  slack_prop = 0.05,
  warn = FALSE,
  lower = NULL,
  upper = NULL,
  id = rand_id("range_check_")
)

## S3 method for class 'check_range'
tidy(x, ...)

Arguments

recipe
A recipe object. The check will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the check. See selections() for more details. For the tidy method, these are not currently used.

role
Not used by this check since no new variables are created.

skip
A logical. Should the check be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

slack_prop
The allowed slack as a proportion of the range of the variable in the train set.

warn
If TRUE the check will throw a warning instead of an error when failing.

lower
A named numeric vector of minimum values in the train set. This is NULL until computed by prep.recipe().

upper
A named numeric vector of maximum values in the train set. This is NULL until computed by prep.recipe().

id
A character string that is unique to this step to identify it.

x
A check_range object.

Details

The amount of slack that is allowed is determined by the slack_prop. This is a numeric of length one or two. If of length one, the same proportion will be used at both ends of the train set range. If of length two, its first value is used to compute the allowed slack at the lower end, the second to compute the allowed slack at the upper end.

Value

An updated version of recipe with the new check added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms (the selectors or variables selected) and value (the means).

See Also

recipe() prep.recipe() bake.recipe()

Examples

  slack_df <- data_frame(x = 0:100)
  slack_new_data <- data_frame(x = -10:110)

  # this will fail the check both ends
## Not run: 
  recipe(slack_df) %>%
    check_range(x) %>%
    prep() %>%
    bake(slack_new_data)
 
## End(Not run)

  # this will fail the check only at the upper end
## Not run: 
  recipe(slack_df) %>%
    check_range(x, slack_prop = c(0.1, 0.05)) %>%
    prep() %>%
    bake(slack_new_data)

## End(Not run)

  # give a warning instead of an error
## Not run: 
  recipe(slack_df) %>%
    check_range(x, warn = TRUE) %>%
    prep() %>%
    bake(slack_new_data)

## End(Not run)

[Package recipes version 0.1.15 Index]