check_type {recipes}	R Documentation

Quantitatively check on variables

Description

This internal function is to be used in the prep function to ensure that the type of the variables matches the expectation. Throws an error if check fails.

Usage

check_type(dat, quant = TRUE)

Arguments

dat
A data frame or tibble of the training data.

quant
A logical indicating whether the data is expected to be numeric (TRUE) or a factor/character (FALSE).

[Package recipes version 0.1.15 Index]