covers {recipes}	R Documentation

Raw Cover Type Data

Description

These data are raw data describing different types of forest cover-types from the UCI Machine Learning Database (see link below). There is one column in the data that has a few difference pieces of textual information (of variable lengths).

Value

covers
a data frame

Source

https://archive.ics.uci.edu/ml/machine-learning-databases/covtype/covtype.info

Examples


data(covers)
str(covers)

[Package recipes version 0.1.5 Index]