credit_data {recipes}	R Documentation

Credit Data

Description

These data are from the website of Dr. Lluís A. Belanche Muñoz by way of a github repository of Dr. Gaston Sanchez. One data point is a missing outcome was removed from the original data.

Value

credit_data
a data frame

Source

https://github.com/gastonstat/CreditScoring, http://bit.ly/2kkBFrk

Examples


data(credit_data)
str(credit_data)

[Package recipes version 0.1.5 Index]