fully_trained {recipes}	R Documentation

Check to see if a recipe is trained/prepared

Description

Check to see if a recipe is trained/prepared

Usage

fully_trained(x)

Arguments

x
A recipe

Value

A logical which is true if all of the recipe steps have been run through prep. If no steps have been added to the recipe, TRUE is returned only if the recipe has been prepped.

Examples

rec <- recipe(Species ~ ., data = iris) %>%
  step_center(all_numeric())

rec %>% fully_trained()

rec %>% prep(training = iris) %>% fully_trained()

[Package recipes version 0.1.15 Index]