Train a Data Recipe

For a recipe with at least one preprocessing step, estimate the required parameters from a training set that can be later applied to other data sets.

learn(x, ...)

# S3 method for recipe
learn(x, training = NULL, fresh = FALSE, verbose = TRUE,
  retain = FALSE, stringsAsFactors = TRUE, ...)

Arguments

x

an object

...

further arguments passed to or from other methods (not currently used).

training

A data frame or tibble that will be used to estimate parameters for preprocessing.

fresh

A logical indicating whether already trained steps should be re-trained. If TRUE, you should pass in a data set to the argument training.

verbose

A logical that controls wether progress is reported as steps are executed.

retain

A logical: should the processed training set be saved into the template slot of the recipe after training? This is a good idea if you want to add more steps later but want to avoid re-training the existing steps.

stringsAsFactors

A logical: should character columns be converted to factors? This affects the processed training set (when retain = TRUE) as well as the results of process.recipe.

Value

A recipe whose step objects have been updated with the required quantities (e.g. parameter estimates, model objects, etc). Also, the term_info object is likely to be modified as the steps are executed.

Details

Given a data set, this function estimates the required quantities and statistics required by any steps. learn returns an updated recipe with the estimates.

Note that missing data handling is handled in the steps; there is no global na.rm option at the recipe-level or in learn.

Also, if a recipe has been trained using learn and then steps are added, learn will only update the new steps. If fresh = TRUE, all of the steps will be (re)estimated.

As the steps are executed, the training set is updated. For example, if the first step is to center the data and the second is to scale the data, the step for scaling is given the centered data.