okc {recipes}	R Documentation

OkCupid Data

Description

These are a sample of columns of users of OkCupid dating website. The data are from Kim and Escobedo-Land (2015). Permission to use this data set was explicitly granted by OkCupid.

Value

okc
a data frame

Source

Kim, A. Y., and A. Escobedo-Land. 2015. "OkCupid Data for Introductory Statistics and Data Science Courses." Journal of Statistics Education: An International Journal on the Teaching and Learning of Statistics.

Examples


data(okc)
str(okc)

[Package recipes version 0.1.5 Index]