Parse a Step Function Formula.

parse_terms_formula takes the terms element of a step and translates it to a character vector of column names.

parse_terms_formula(f, info)

Arguments

f

A formula that can inlcude variable names and certain functions. See selections for a list of functions.

info

A tibble with columns variable, type, role, and source that represent the current state of the data. The function summary.recipe can be used to get this information from a recipe.

Value

A character string of column names.

See also

recipe summary.recipe learn.recipe

Examples

rec <- recipe( ~ ., data = USArrests)
rec <- step_pca(rec, terms = ~ is_numeric(), num = 3)
parse_terms_formula(~ predictors(), summary(rec))

#> [1] "Murder"   "Assault"  "UrbanPop" "Rape"    

rec <- learn(rec, training = USArrests)

#> step 1 pca training 

parse_terms_formula(~ predictors(), summary(rec))

#> [1] "PC1" "PC2" "PC3"