recipes_pkg_check {recipes}	R Documentation

Update packages

Description

This will check to see if all required packages are installed.

Usage

recipes_pkg_check(pkg = NULL, ...)

Arguments

pkg
A character string for the package being checked

...
Extra arguments to pass to utils::install.packages()

Value

Nothing is returned but a message is printed to the console about which packages (if any) should be installed along with code to do so.

Examples

## Not run: 
recipes_pkg_check("recipes")

## End(Not run)

[Package recipes version 0.1.15 Index]