required_pkgs.step_depth {recipes}	R Documentation

S3 methods for tracking which additional packages are needed for steps.

Description

S3 methods for tracking which additional packages are needed for steps.

Usage

## S3 method for class 'step_depth'
required_pkgs(x, ...)

## S3 method for class 'step_ica'
required_pkgs(x, ...)

## S3 method for class 'step_isomap'
required_pkgs(x, ...)

## S3 method for class 'step_kpca'
required_pkgs(x, ...)

## S3 method for class 'step_kpca_poly'
required_pkgs(x, ...)

## S3 method for class 'step_kpca_rbf'
required_pkgs(x, ...)

## S3 method for class 'step_nnmf'
required_pkgs(x, ...)

## S3 method for class 'step_pls'
required_pkgs(x, ...)

## S3 method for class 'step_window'
required_pkgs(x, ...)

Arguments

x
A recipe step

Value

A character vector

[Package recipes version 0.1.15 Index]