Select Terms in a Step Function.

This function processes the step function selectors and might be useful when creating custom steps.

select_terms(info, args)

Arguments

info

A tibble with columns variable, type, role, and source that represent the current state of the data. The function summary.recipe can be used to get this information from a recipe.

args

A list of formulas whose right-hand side contains quoted expressions. See tidy_quote for examples.

Value

A character string of column names or an error of there are no selectors or if no variables are selected.

See also

recipe summary.recipe learn.recipe

Examples

data(okc)
rec <- recipe(~ ., data = okc)
info <- summary(rec)
select_terms(info = info, tidy_quotes(all_predictors()))

#> Error in x$variable: $ operator is invalid for atomic vectors