step_bs {recipes}	R Documentation

B-Spline Basis Functions

Description

step_bs creates a specification of a recipe step that will create new columns that are basis expansions of variables using B-splines.

Usage

step_bs(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  deg_free = NULL,
  degree = 3,
  objects = NULL,
  options = list(),
  skip = FALSE,
  id = rand_id("bs")
)

## S3 method for class 'step_bs'
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details. For the tidy method, these are not currently used.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created from the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

deg_free
The degrees of freedom.

degree
Degree of polynomial spline (integer).

objects
A list of splines::bs() objects created once the step has been trained.

options
A list of options for splines::bs() which should not include x, degree, or df.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_bs object.

Details

step_bs can create new features from a single variable that enable fitting routines to model this variable in a nonlinear manner. The extent of the possible nonlinearity is determined by the df, degree, or knot arguments of splines::bs(). The original variables are removed from the data and new columns are added. The naming convention for the new variables is varname_bs_1 and so on.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms which is the columns that will be affected and holiday.

See Also

step_poly() recipe() step_ns() prep.recipe() bake.recipe()

Examples

library(modeldata)
data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)

with_splines <- rec %>%
  step_bs(carbon, hydrogen)
with_splines <- prep(with_splines, training = biomass_tr)

expanded <- bake(with_splines, biomass_te)
expanded

[Package recipes version 0.1.15 Index]