step_dummy {recipes}	R Documentation

Dummy Variables Creation

Description

step_dummy creates a specification of a recipe step that will convert nominal data (e.g. character or factors) into one or more numeric binary model terms for the levels of the original data.

Usage

step_dummy(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  one_hot = FALSE,
  preserve = FALSE,
  naming = dummy_names,
  levels = NULL,
  skip = FALSE,
  id = rand_id("dummy")
)

## S3 method for class 'step_dummy'
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which factor variables will be used to create the dummy variables. See selections() for more details. The selected variables must be factors. For the tidy method, these are not currently used.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the binary dummy variable columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

one_hot
A logical. For C levels, should C dummy variables be created rather than C-1?

preserve
A single logical; should the selected column(s) be retained (in addition to the new dummy variables).

naming
A function that defines the naming convention for new dummy columns. See Details below.

levels
A list that contains the information needed to create dummy variables for each variable contained in terms. This is NULL until the step is trained by prep.recipe().

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_dummy object.

Details

step_dummy will create a set of binary dummy variables from a factor variable. For example, if an unordered factor column in the data set has levels of "red", "green", "blue", the dummy variable bake will create two additional columns of 0/1 data for two of those three values (and remove the original column). For ordered factors, polynomial contrasts are used to encode the numeric values.

By default, the excluded dummy variable (i.e. the reference cell) will correspond to the first level of the unordered factor being converted.

The function allows for non-standard naming of the resulting variables. For an unordered factor named x, with levels "a" and "b", the default naming convention would be to create a new variable called x_b. Note that if the factor levels are not valid variable names (e.g. "some text with spaces"), it will be changed by base::make.names() to be valid (see the example below). The naming format can be changed using the naming argument and the function dummy_names() is the default. This function will also change the names of ordinal dummy variables. Instead of values such as ".L", ".Q", or "^4", ordinal dummy variables are given simple integer suffixes such as "_1", "_2", etc.

To change the type of contrast being used, change the global contrast option via options.

When the factor being converted has a missing value, all of the corresponding dummy variables are also missing.

When data to be processed contains novel levels (i.e., not contained in the training set), a missing value is assigned to the results. See step_other() for an alternative.

If no columns are selected (perhaps due to an earlier step_zv()), bake() will return the data as-is (e.g. with no dummy variables).

The package vignette for dummy variables and interactions has more information.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms (the selectors or original variables selected) and columns (the list of corresponding binary columns).

See Also

step_factor2string(), step_string2factor(), dummy_names(), step_regex(), step_count(), step_ordinalscore(), step_unorder(), step_other() step_novel()

Examples

library(modeldata)
data(okc)
okc <- okc[complete.cases(okc),]

rec <- recipe(~ diet + age + height, data = okc)

dummies <- rec %>% step_dummy(diet)
dummies <- prep(dummies, training = okc)

dummy_data <- bake(dummies, new_data = okc)

unique(okc$diet)
grep("^diet", names(dummy_data), value = TRUE)

# Obtain the full set of dummy variables using `one_hot` option
rec %>%
  step_dummy(diet, one_hot = TRUE) %>%
  prep(training = okc) %>%
  bake(new_data = NULL, starts_with("diet")) %>%
  names() %>%
  length()

length(unique(okc$diet))

# Without one_hot
length(grep("^diet", names(dummy_data), value = TRUE))

tidy(dummies, number = 1)

[Package recipes version 0.1.15 Index]