Multiple Choice Dummy Variables Creation

step_dummy_multi_choice() creates a specification of a recipe step that will convert multiple nominal data (e.g. character or factors) into one or more numeric binary model terms for the levels of the original data.

step_dummy_multi_choice(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  threshold = 0,
  levels = NULL,
  input = NULL,
  other = "other",
  naming = dummy_names,
  prefix = NULL,
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("dummy_multi_choice")
)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose variables for this step. See selections() for more details. The selected variables must be factors.

role
For model terms created by this step, what analysis role should they be assigned? By default, the new columns created by this step from the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

threshold
A numeric value between 0 and 1, or an integer greater or equal to one. If less than one, then factor levels with a rate of occurrence in the training set below threshold will be pooled to other. If greater or equal to one, then this value is treated as a frequency and factor levels that occur less than threshold times will be pooled to other.

levels
A list that contains the information needed to create dummy variables for each variable contained in terms. This is NULL until the step is trained by prep.recipe().

input
A character vector containing the names of the columns used. This is NULL until the step is trained by prep.recipe().

other
A single character value for the "other" category.

naming
A function that defines the naming convention for new dummy columns. See Details below.

prefix
A character string for the prefix of the resulting new variables. See notes below.

keep_original_cols
A logical to keep the original variables in the output. Defaults to FALSE.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id
A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Details

The overall proportion (or total counts) of the categories are computed. The "other" category is used in place of any categorical levels whose individual proportion (or frequency) in the training set is less than threshold.

This recipe step allows for flexible naming of the resulting variables. For an unordered factor named x, with levels "a" and "b", the default naming convention would be to create a new variable called x_b. The naming format can be changed using the naming argument; the function dummy_names() is the default.

See also

Examples

library(tibble)
languages <- tribble(
  ~lang_1,    ~lang_2,   ~lang_3,
  "English",  "Italian", NA,
  "Spanish",  NA,        "French",
  "Armenian", "English", "French",
  NA,         NA,        NA
)

dummy_multi_choice_rec <- recipe(~ ., data = languages) %>%
  step_dummy_multi_choice(starts_with("lang")) %>%
  prep()

bake(dummy_multi_choice_rec, new_data = NULL)
#> # A tibble: 4 × 5
#>   lang_1_Armenian lang_1_English lang_1_French lang_1_Italian lang_1_Spanish
#>             <int>          <int>         <int>          <int>          <int>
#> 1               0              1             0              1              0
#> 2               0              0             1              0              1
#> 3               1              1             1              0              0
#> 4               0              0             0              0              0
tidy(dummy_multi_choice_rec, number = 1)
#> # A tibble: 5 × 3
#>   terms  columns  id                      
#>   <chr>  <chr>    <chr>                   
#> 1 lang_1 Armenian dummy_multi_choice_ONQD4
#> 2 lang_1 English  dummy_multi_choice_ONQD4
#> 3 lang_1 French   dummy_multi_choice_ONQD4
#> 4 lang_1 Italian  dummy_multi_choice_ONQD4
#> 5 lang_1 Spanish  dummy_multi_choice_ONQD4

dummy_multi_choice_rec2 <- recipe(~ ., data = languages) %>%
  step_dummy_multi_choice(starts_with("lang"), prefix = "lang",
                         threshold = 0.2) %>%
  prep()

bake(dummy_multi_choice_rec2, new_data = NULL)
#> # A tibble: 4 × 3
#>   lang_English lang_French lang_other
#>          <int>       <int>      <int>
#> 1            1           0          1
#> 2            0           1          1
#> 3            1           1          1
#> 4            0           0          0
tidy(dummy_multi_choice_rec2, number = 1)
#> # A tibble: 3 × 3
#>   terms  columns id                      
#>   <chr>  <chr>   <chr>                   
#> 1 lang_1 English dummy_multi_choice_mbqAp
#> 2 lang_1 French  dummy_multi_choice_mbqAp
#> 3 lang_1 other   dummy_multi_choice_mbqAp