step_geodist {recipes}	R Documentation

Distance between two locations

Description

step_geodist creates a specification of a recipe step that will calculate the distance between points on a map to a reference location.

Usage

step_geodist(
  recipe,
  lat = NULL,
  lon = NULL,
  role = "predictor",
  trained = FALSE,
  ref_lat = NULL,
  ref_lon = NULL,
  log = FALSE,
  name = "geo_dist",
  columns = NULL,
  skip = FALSE,
  id = rand_id("geodist")
)

## S3 method for class 'step_geodist'
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

lon, lat
Selector functions to choose which variables are affected by the step. See selections() for more details.

role
or model term created by this step, what analysis role should be assigned?. By default, the function assumes that resulting distance will be used as a predictor in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

ref_lon, ref_lat
Single numeric values for the location of the reference point.

log
A logical: should the distance be transformed by the natural log function?

name
A single character value to use for the new predictor column. If a column exists with this name, an error is issued.

columns
A character string of variable names that will be populated (eventually) by the terms argument.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_geodist object.

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details. For the tidy method, these are not currently used.

Details

step_geodist will create a

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns echoing the values of lat, lon, ref_lat, ref_lon, name, and id.

Examples

library(modeldata)
data(Smithsonian)

# How close are the museums to Union Station?
near_station <- recipe( ~ ., data = Smithsonian) %>%
  update_role(name, new_role = "location") %>%
  step_geodist(lat = latitude, lon = longitude, log = FALSE,
               ref_lat = 38.8986312, ref_lon = -77.0062457) %>%
  prep(training = Smithsonian)

bake(near_station, new_data = NULL) %>%
  arrange(geo_dist)

tidy(near_station, number = 1)

[Package recipes version 0.1.15 Index]