step_hyperbolic {recipes}	R Documentation

Hyperbolic Transformations

Description

step_hyperbolic creates a specification of a recipe step that will transform data using a hyperbolic function.

Usage

step_hyperbolic(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  func = "sin",
  inverse = TRUE,
  columns = NULL,
  skip = FALSE,
  id = rand_id("hyperbolic")
)

## S3 method for class 'step_hyperbolic'
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details. For the tidy method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

func
A character value for the function. Valid values are "sin", "cos", or "tan".

inverse
A logical: should the inverse function be used?

columns
A character string of variable names that will be populated (eventually) by the terms argument.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_hyperbolic object.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms (the columns that will be affected), inverse, and func.

See Also

step_logit() step_invlogit() step_log() step_sqrt() recipe() prep.recipe() bake.recipe()

Examples

set.seed(313)
examples <- matrix(rnorm(40), ncol = 2)
examples <- as.data.frame(examples)

rec <- recipe(~ V1 + V2, data = examples)

cos_trans <- rec  %>%
  step_hyperbolic(all_predictors(),
                  func = "cos", inverse = FALSE)

cos_obj <- prep(cos_trans, training = examples)

transformed_te <- bake(cos_obj, examples)
plot(examples$V1, transformed_te$V1)

tidy(cos_trans, number = 1)
tidy(cos_obj, number = 1)

[Package recipes version 0.1.15 Index]