step_intercept {recipes}	R Documentation

Add intercept (or constant) column

Description

step_intercept creates a specification of a recipe step that will add an intercept or constant term in the first column of a data matrix. step_intercept has defaults to predictor role so that it is by default called in the bake step. Be careful to avoid unintentional transformations when calling steps with all_predictors.

Usage

step_intercept(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  name = "intercept",
  value = 1,
  skip = FALSE,
  id = rand_id("intercept")
)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
Argument ignored; included for consistency with other step specification functions.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created from the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated. Again included for consistency.

name
Character name for newly added column

value
A numeric constant to fill the intercept column. Defaults to 1.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

See Also

recipe() prep.recipe() bake.recipe()

Examples

library(modeldata)
data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)
rec_trans <- recipe(HHV ~ ., data = biomass_tr[, -(1:2)]) %>%
  step_intercept(value = 2) %>%
  step_scale(carbon)

rec_obj <- prep(rec_trans, training = biomass_tr)

with_intercept <- bake(rec_obj, biomass_te)
with_intercept

[Package recipes version 0.1.15 Index]