Impute Numeric Data Using the Mean

step_meanimpute creates a specification of a recipe step that will substitute missing values of numeric variables by the training set mean of those variables.

step_meanimpute(recipe, ..., role = NA, trained = FALSE, means = NULL,
  trim = 0)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections for more details.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

means
A named numeric vector of means. This is NULL until computed by prepare.recipe.

trim
The fraction (0 to 0.5) of observations to be trimmed from each end of the variables before the mean is computed. Values of trim outside that range are taken as the nearest endpoint.

Value

step_meanimpute returns an object of class step_meanimpute.

Details

step_meanimpute estimates the variable means from the data used in the training argument of prepare.recipe. bake.recipe then applies the new values to new data sets using these averages.

Examples

data("credit_data")

## missing data per column
vapply(credit_data, function(x) mean(is.na(x)), c(num = 0))

#>       Status    Seniority         Home         Time          Age      Marital 
#> 0.0000000000 0.0000000000 0.0013471037 0.0000000000 0.0000000000 0.0002245173 
#>      Records          Job     Expenses       Income       Assets         Debt 
#> 0.0000000000 0.0004490346 0.0000000000 0.0855410867 0.0105523125 0.0040413112 
#>       Amount        Price 
#> 0.0000000000 0.0000000000 

set.seed(342)
in_training <- sample(1:nrow(credit_data), 2000)

credit_tr <- credit_data[ in_training, ]
credit_te <- credit_data[-in_training, ]
missing_examples <- c(14, 394, 565)

rec <- recipe(Price ~ ., data = credit_tr)

impute_rec <- rec %>%
  step_meanimpute(Income, Assets, Debt)

imp_models <- prepare(impute_rec, training = credit_tr)

#> step 1 meanimpute training 

imputed_te <- bake(imp_models, newdata = credit_te, everything())

credit_te[missing_examples,]

#>      Status Seniority Home Time Age Marital Records   Job Expenses Income
#> 30      bad         0 <NA>   48  37  single      no  <NA>       35     NA
#> 736     bad         1 priv   48  36  single     yes fixed       35    173
#> 1057    bad        24 priv   60  46 married      no fixed       60    165
#>      Assets Debt Amount Price
#> 30       NA   NA   1500  1850
#> 736    7700    0   1000  2008
#> 1057   3500 1300   1400  1530

imputed_te[missing_examples, names(credit_te)]

#> # A tibble: 3 x 14
#>   Status Seniority   Home  Time   Age Marital Records    Job Expenses   Income
#>   <fctr>     <int> <fctr> <int> <int>  <fctr>  <fctr> <fctr>    <int>    <dbl>
#> 1    bad         0     NA    48    37  single      no     NA       35 138.9908
#> 2    bad         1   priv    48    36  single     yes  fixed       35 173.0000
#> 3    bad        24   priv    60    46 married      no  fixed       60 165.0000
#> # ... with 4 more variables: Assets <dbl>, Debt <dbl>, Amount <int>,
#> #   Price <int>