Mutate multiple columns

step_mutate_at creates a specification of a recipe step that will modify the selected variables using a common function.

step_mutate_at(recipe, ..., fn, role = "predictor", trained = FALSE,
  inputs = NULL, skip = FALSE, id = rand_id("mutate_at"))

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details. For the tidy method, these are not currently used.

fn
A function fun, a quosure style lambda `~ fun(.)`` or a list of either form. (see dplyr::mutate_at()). Note that this argument must be named.

role
For model terms created by this step, what analysis role should they be assigned? By default, the function assumes that the new dimension columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

inputs
A vector of column names populated by prep().

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms which contains the columns being transformed.

Examples

library(dplyr)
recipe(~ ., data = iris) %>%
  step_mutate_at(contains("Length"), fn = ~ 1/.) %>%
  prep() %>%
  juice() %>%
  slice(1:10)

#> # A tibble: 10 x 5
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>  
#>  1        0.196         3.5        0.714         0.2 setosa 
#>  2        0.204         3          0.714         0.2 setosa 
#>  3        0.213         3.2        0.769         0.2 setosa 
#>  4        0.217         3.1        0.667         0.2 setosa 
#>  5        0.2           3.6        0.714         0.2 setosa 
#>  6        0.185         3.9        0.588         0.4 setosa 
#>  7        0.217         3.4        0.714         0.3 setosa 
#>  8        0.2           3.4        0.667         0.2 setosa 
#>  9        0.227         2.9        0.714         0.2 setosa 
#> 10        0.204         3.1        0.667         0.1 setosa 

recipe(~ ., data = iris) %>%
  # leads to more columns being created.
  step_mutate_at(contains("Length"), fn = list(log = log, sqrt = sqrt)) %>%
  prep() %>%
  juice() %>%
  slice(1:10)

#> # A tibble: 10 x 9
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species Sepal.Length_log
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>              <dbl>
#>  1          5.1         3.5          1.4         0.2 setosa              1.63
#>  2          4.9         3            1.4         0.2 setosa              1.59
#>  3          4.7         3.2          1.3         0.2 setosa              1.55
#>  4          4.6         3.1          1.5         0.2 setosa              1.53
#>  5          5           3.6          1.4         0.2 setosa              1.61
#>  6          5.4         3.9          1.7         0.4 setosa              1.69
#>  7          4.6         3.4          1.4         0.3 setosa              1.53
#>  8          5           3.4          1.5         0.2 setosa              1.61
#>  9          4.4         2.9          1.4         0.2 setosa              1.48
#> 10          4.9         3.1          1.5         0.1 setosa              1.59
#> # … with 3 more variables: Petal.Length_log <dbl>, Sepal.Length_sqrt <dbl>,
#> #   Petal.Length_sqrt <dbl>