Nature Spline Basis Functions

step_ns creates a specification of a recipe step that will create new columns that are basis expansions of variables using natural spines.

step_ns(recipe, ..., role = "predictor", trained = FALSE, objects = NULL,
  options = list(df = 2))

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose which variables are affected by the step. See selections for more details.

role

For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created from the original variables will be used as all_predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

objects

A list of ns objects created once the step has been learned.

options

A list of options for ns which should not include x.

Value

step_ns returns an object of class step_ns.

Details

step_ns can new features from a single variable that enable fitting routines to model this variable in a nonlinear manner. The extent of the possible nonlinearity is determined by the df or knot arguments of ns. The original variables are removed from the data and new columns are added. The naming convention for the new variables is varname_ns_1 and so on.

See also

step_poly recipe learn.recipe process.recipe

Examples

data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)

with_splines <- rec %>%
  step_ns(carbon, hydrogen)
with_splines <- learn(with_splines, training = biomass_tr)

#> step 1 ns training 

#> Error in x$variable: $ operator is invalid for atomic vectors

expanded <- process(with_splines, biomass_te)
expanded

#> # A tibble: 80 × 6
#>    carbon hydrogen oxygen nitrogen sulfur    HHV
#> *   <dbl>    <dbl>  <dbl>    <dbl>  <dbl>  <dbl>
#> 1   46.35     5.67  47.20     0.30   0.22 18.275
#> 2   43.25     5.50  48.06     2.85   0.34 17.560
#> 3   42.70     5.50  49.10     2.40   0.30 17.173
#> 4   46.40     6.10  37.30     1.80   0.50 18.851
#> 5   48.76     6.32  42.77     0.20   0.00 20.547
#> 6   44.30     5.50  41.70     0.70   0.20 18.467
#> 7   38.94     5.23  54.13     1.19   0.51 15.095
#> 8   42.10     4.66  33.80     0.95   0.20 16.240
#> 9   29.20     4.40  31.10     0.14   4.90 11.147
#> 10  27.80     3.77  23.69     4.63   1.05 10.750
#> # ... with 70 more rows