step_rename_at {recipes}	R Documentation

Rename multiple columns

Description

step_rename_at creates a specification of a recipe step that will rename the selected variables using a common function.

Usage

step_rename_at(
  recipe,
  ...,
  fn,
  role = "predictor",
  trained = FALSE,
  inputs = NULL,
  skip = FALSE,
  id = rand_id("rename_at")
)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details. For the tidy method, these are not currently used.

fn
A function fun, a quosure style lambda '~ fun(.)“ or a list of either form (but containing only a single function, see dplyr::rename_at()). Note that this argument must be named.

role
For model terms created by this step, what analysis role should they be assigned? By default, the function assumes that the new dimension columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

inputs
A vector of column names populated by prep().

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms which contains the columns being transformed.

Examples

library(dplyr)
recipe(~ ., data = iris) %>%
  step_rename_at(everything(), fn = ~ gsub(".", "_", ., fixed = TRUE)) %>%
  prep() %>%
  bake(new_data = NULL) %>%
  slice(1:10)

[Package recipes version 0.1.15 Index]