Convert Ordered Factors to Unordered Factors

`step_unorder` creates a *specification* of a recipe step that will transform the data.

step_unorder(recipe, ..., role = NA, trained = FALSE, columns = NULL,
  skip = FALSE)

# S3 method for step_unorder
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See [selections()] for more details. For the `tidy` method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

columns
A character string of variable names that will be populated (eventually) by the `terms` argument.

skip
A logical. Should the step be skipped when the recipe is baked by [bake.recipe()]? While all operations are baked when [prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations

x
A `step_unorder` object.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any). For the `tidy` method, a tibble with columns `terms` (the columns that will be affected).

Details

The factors level order is preserved during the transformation.

See also

Examples

lmh <- c("Low", "Med", "High")

examples <- data.frame(X1 = factor(rep(letters[1:4], each = 3)),
                       X2 = ordered(rep(lmh, each = 4),
                                    levels = lmh))

rec <- recipe(~ X1 + X2, data = examples)

factor_trans <- rec  %>%
  step_unorder(all_predictors())

factor_obj <- prep(factor_trans, training = examples)

#> Warning: `step_unorder` requires ordered factors. Variables X1 will be ignored.

transformed_te <- bake(factor_obj, examples)
table(transformed_te$X2, examples$X2)

#>       
#>        Low Med High
#>   Low    4   0    0
#>   Med    0   4    0
#>   High   0   0    4

tidy(factor_trans, number = 1)

#> # A tibble: 1 x 1
#>   terms           
#>   <chr>           
#> 1 all_predictors()

tidy(factor_obj, number = 1)

#> # A tibble: 1 x 1
#>   terms
#>   <chr>
#> 1 X2