step_upsample {recipes}	R Documentation

Up-Sample a Data Set Based on a Factor Variable

Description

step_upsample is now available as themis::step_upsample(). This function creates a specification of a recipe step that will replicate rows of a data set to make the occurrence of levels in a specific factor level equal.

Usage

step_upsample(
  recipe,
  ...,
  over_ratio = 1,
  ratio = NA,
  role = NA,
  trained = FALSE,
  column = NULL,
  target = NA,
  skip = TRUE,
  seed = sample.int(10^5, 1),
  id = rand_id("upsample")
)

## S3 method for class 'step_upsample'
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variable is used to sample the data. See selections() for more details. The selection should result in single factor variable. For the tidy method, these are not currently used.

over_ratio
A numeric value for the ratio of the majority-to-minority frequencies. The default value (1) means that all other levels are sampled up to have the same frequency as the most occurring level. A value of 0.5 would mean that the minority levels will have (at most) (approximately) half as many rows than the majority level.

ratio
Deprecated argument; same as over_ratio.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

column
A character string of the variable name that will be populated (eventually) by the ... selectors.

target
An integer that will be used to subsample. This should not be set by the user and will be populated by prep.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

seed
An integer that will be used as the seed when upsampling.

id
A character string that is unique to this step to identify it.

x
A step_upsample object.

Details

Up-sampling is intended to be performed on the training set alone. For this reason, the default is skip = TRUE. It is advisable to use prep(recipe, retain = TRUE) when preparing the recipe; in this way bake(object, new_data = NULL) can be used to obtain the up-sampled version of the data.

If there are missing values in the factor variable that is used to define the sampling, missing data are selected at random in the same way that the other factor levels are sampled. Missing values are not used to determine the amount of data in the majority level (see example below).

For any data with factor levels occurring with the same frequency as the majority level, all data will be retained.

All columns in the data are sampled and returned by bake().

When used in modeling, users should strongly consider using the option skip = TRUE so that the extra sampling is not conducted outside of the training set.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms which is the variable used to sample.

Examples

library(modeldata)
data(okc)

orig <- table(okc$diet, useNA = "always")

sort(orig, decreasing = TRUE)

up_rec <- recipe( ~ ., data = okc) %>%
  # Bring the minority levels up to about 200 each
  # 200/16562 is approx 0.0121
  step_upsample(diet, over_ratio = 0.0121) %>%
  prep(training = okc)

training <- table(bake(up_rec, new_data = NULL)$diet, useNA = "always")

# Since `skip` defaults to TRUE, baking the step has no effect
baked_okc <- bake(up_rec, new_data = okc)
baked <- table(baked_okc$diet, useNA = "always")

# Note that if the original data contained more rows than the
# target n (= ratio * majority_n), the data are left alone:
data.frame(
  level = names(orig),
  orig_freq = as.vector(orig),
  train_freq = as.vector(training),
  baked_freq = as.vector(baked)
)

[Package recipes version 0.1.15 Index]