summary.recipe {recipes}	R Documentation

Summarize a Recipe

Description

This function prints the current set of variables/features and some of their characteristics.

Usage

## S3 method for class 'recipe'
summary(object, original = FALSE, ...)

Arguments

object
A recipe object

original
A logical: show the current set of variables or the original set when the recipe was defined.

...
further arguments passed to or from other methods (not currently used).

Details

Note that, until the recipe has been trained, the current and original variables are the same.

It is possible for variables to have multiple roles by adding them with add_role(). If a variable has multiple roles, it will have more than one row in the summary tibble.

Value

A tibble with columns variable, type, role, and source.

See Also

recipe() prep.recipe()

Examples

rec <- recipe( ~ ., data = USArrests)
summary(rec)
rec <- step_pca(rec, all_numeric(), num_comp = 3)
summary(rec) # still the same since not yet trained
rec <- prep(rec, training = USArrests)
summary(rec)

[Package recipes version 0.1.15 Index]