Manually Add Roles

`add_role` can add a role definition to an existing variable in the recipe.

add_role(recipe, ..., new_role = "predictor")

Arguments

recipe
An existing [recipe()].

...
One or more selector functions to choose which variables are being assigned a role. See [selections()] for more details.

new_role
A character string for a single role.

Value

An updated recipe object.

Details

If a variable is selected that currently has a role, the role is changed and a warning is issued.

Examples

data(biomass)

# Create the recipe manually
rec <- recipe(x = biomass)
rec

#> Data Recipe
#> 
#> Inputs:
#> 
#>   8 variables (no declared roles)

summary(rec)

#> # A tibble: 8 x 4
#>   variable type    role  source  
#>   <chr>    <chr>   <lgl> <chr>   
#> 1 sample   nominal NA    original
#> 2 dataset  nominal NA    original
#> 3 carbon   numeric NA    original
#> 4 hydrogen numeric NA    original
#> 5 oxygen   numeric NA    original
#> 6 nitrogen numeric NA    original
#> 7 sulfur   numeric NA    original
#> 8 HHV      numeric NA    original

rec <- rec %>%
  add_role(carbon, contains("gen"), sulfur, new_role = "predictor") %>%
  add_role(sample, new_role = "id variable") %>%
  add_role(dataset, new_role = "splitting variable") %>%
  add_role(HHV, new_role = "outcome")
rec

#> Data Recipe
#> 
#> Inputs:
#> 
#>                role #variables
#>         id variable          1
#>             outcome          1
#>           predictor          5
#>  splitting variable          1