Check for Missing Values

check_missing creates a specification of a recipe operation that will check if variables contain missing values.

check_missing(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  columns = NULL,
  skip = FALSE,
  id = rand_id("missing")
)

# S3 method for check_missing
tidy(x, ...)

Arguments

recipe
A recipe object. The check will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are checked in the check See selections() for more details. For the tidy method, these are not currently used.

role
Not used by this check since no new variables are created.

trained
A logical for whether the selectors in ... have been resolved by prep().

columns
A character string of variable names that will be populated (eventually) by the terms argument.

skip
A logical. Should the check be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id
A character string that is unique to this step to identify it.

x
A check_missing object.

Value

An updated version of recipe with the new check added to the sequence of existing operations (if any). For the tidy method, a tibble with columns terms (the selectors or variables selected).

Details

This check will break the bake function if any of the checked columns does contain NA values. If the check passes, nothing is changed to the data.

Examples

library(modeldata)
data(credit_data)
is.na(credit_data) %>% colSums()

#>    Status Seniority      Home      Time       Age   Marital   Records       Job 
#>         0         0         6         0         0         1         0         2 
#>  Expenses    Income    Assets      Debt    Amount     Price 
#>         0       381        47        18         0         0 

# If the test passes, `new_data` is returned unaltered
recipe(credit_data) %>%
  check_missing(Age, Expenses) %>%
  prep() %>%
  bake(credit_data)

#> # A tibble: 4,454 x 14
#>    Status Seniority Home   Time   Age Marital Records Job   Expenses Income
#>    <fct>      <int> <fct> <int> <int> <fct>   <fct>   <fct>    <int>  <int>
#>  1 good           9 rent     60    30 married no      free…       73    129
#>  2 good          17 rent     60    58 widow   no      fixed       48    131
#>  3 bad           10 owner    36    46 married yes     free…       90    200
#>  4 good           0 rent     60    24 single  no      fixed       63    182
#>  5 good           0 rent     36    26 single  no      fixed       46    107
#>  6 good           1 owner    60    36 married no      fixed       75    214
#>  7 good          29 owner    60    44 married no      fixed       75    125
#>  8 good           9 pare…    12    27 single  no      fixed       35     80
#>  9 good           0 owner    60    32 married no      free…       90    107
#> 10 bad            0 pare…    48    41 married no      part…       90     80
#> # … with 4,444 more rows, and 4 more variables: Assets <int>, Debt <int>,
#> #   Amount <int>, Price <int>

# If your training set doesn't pass, prep() will stop with an error

if (FALSE) {
recipe(credit_data)  %>%
  check_missing(Income) %>%
  prep()
}

# If `new_data` contain missing values, the check will stop bake()

train_data <- credit_data %>% dplyr::filter(Income > 150)
test_data  <- credit_data %>% dplyr::filter(Income <= 150 | is.na(Income))

rp <- recipe(train_data) %>%
  check_missing(Income) %>%
  prep()

bake(rp, train_data)

#> # A tibble: 1,338 x 14
#>    Status Seniority Home   Time   Age Marital Records Job   Expenses Income
#>    <fct>      <int> <fct> <int> <int> <fct>   <fct>   <fct>    <int>  <int>
#>  1 bad           10 owner    36    46 married yes     free…       90    200
#>  2 good           0 rent     60    24 single  no      fixed       63    182
#>  3 good           1 owner    60    36 married no      fixed       75    214
#>  4 good           8 owner    60    30 married no      fixed       75    199
#>  5 good          19 priv     36    37 married no      fixed       75    170
#>  6 good          15 priv     24    52 single  no      free…       35    330
#>  7 good          33 rent     24    68 married no      free…       65    200
#>  8 good           5 owner    60    22 single  no      fixed       45    324
#>  9 good          19 owner    60    43 single  no      fixed       75    180
#> 10 good          15 owner    36    43 married no      fixed       75    251
#> # … with 1,328 more rows, and 4 more variables: Assets <int>, Debt <int>,
#> #   Amount <int>, Price <int>

if (FALSE) {
bake(rp, test_data)
}