Check for Missing Values

check_new_values creates a a specification of a recipe operation that will check if variables contain new values.

check_new_values(recipe, ..., role = NA, trained = FALSE,
  columns = NULL, ignore_NA = TRUE, values = NULL, skip = FALSE,
  id = rand_id("new_values"))

Arguments

recipe
A recipe object. The check will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are checked in the check. See selections() for more details. For the tidy method, these are not currently used.

role
Not used by this check since no new variables are created.

trained
A logical for whether the selectors in ... have been resolved by prep().

columns
A character string of variable names that will be populated (eventually) by the terms argument.

ignore_NA
A logical that indicates if we should consider missing values as value or not. Defaults to TRUE.

values
A named list with the allowed values. This is NULL until computed by prep.recipe().

skip
A logical. Should the check be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id
A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new check added to the sequence of existing operations (if any). For the tidy method, a tibble with columns terms (the selectors or variables selected).

Details

This check will break the bake function if any of the checked columns does contain values it did not contain when prep was called on the recipe. If the check passes, nothing is changed to the data.

Examples

data(credit_data)

# If the test passes, `new_data` is returned unaltered
recipe(credit_data) %>%
  check_new_values(Home) %>%
  prep() %>%
  bake(new_data = credit_data)

#> # A tibble: 4,454 x 14
#>    Status Seniority Home   Time   Age Marital Records Job   Expenses Income
#>    <fct>      <int> <fct> <int> <int> <fct>   <fct>   <fct>    <int>  <int>
#>  1 good           9 rent     60    30 married no      free…       73    129
#>  2 good          17 rent     60    58 widow   no      fixed       48    131
#>  3 bad           10 owner    36    46 married yes     free…       90    200
#>  4 good           0 rent     60    24 single  no      fixed       63    182
#>  5 good           0 rent     36    26 single  no      fixed       46    107
#>  6 good           1 owner    60    36 married no      fixed       75    214
#>  7 good          29 owner    60    44 married no      fixed       75    125
#>  8 good           9 pare…    12    27 single  no      fixed       35     80
#>  9 good           0 owner    60    32 married no      free…       90    107
#> 10 bad            0 pare…    48    41 married no      part…       90     80
#> # … with 4,444 more rows, and 4 more variables: Assets <int>, Debt <int>,
#> #   Amount <int>, Price <int>

# If `new_data` contains values not in `x` at the `prep()` function,
# the `bake()` function will break.

# NOT RUN {
recipe(credit_data %>% dplyr::filter(Home != "rent")) %>%
  check_new_values(Home) %>%
  prep() %>%
  bake(new_data = credit_data)
# }

# By default missing values are ignored, so this passes.
recipe(credit_data %>% dplyr::filter(!is.na(Home))) %>%
  check_new_values(Home) %>%
  prep() %>%
  bake(credit_data)

#> # A tibble: 4,454 x 14
#>    Status Seniority Home   Time   Age Marital Records Job   Expenses Income
#>    <fct>      <int> <fct> <int> <int> <fct>   <fct>   <fct>    <int>  <int>
#>  1 good           9 rent     60    30 married no      free…       73    129
#>  2 good          17 rent     60    58 widow   no      fixed       48    131
#>  3 bad           10 owner    36    46 married yes     free…       90    200
#>  4 good           0 rent     60    24 single  no      fixed       63    182
#>  5 good           0 rent     36    26 single  no      fixed       46    107
#>  6 good           1 owner    60    36 married no      fixed       75    214
#>  7 good          29 owner    60    44 married no      fixed       75    125
#>  8 good           9 pare…    12    27 single  no      fixed       35     80
#>  9 good           0 owner    60    32 married no      free…       90    107
#> 10 bad            0 pare…    48    41 married no      part…       90     80
#> # … with 4,444 more rows, and 4 more variables: Assets <int>, Debt <int>,
#> #   Amount <int>, Price <int>

# Use `ignore_NA = FALSE` if you consider missing values  as a value,
# that should not occur when not observed in the train set.

# NOT RUN {
recipe(credit_data %>% dplyr::filter(!is.na(Home))) %>%
  check_new_values(Home, ignore_NA = FALSE) %>%
  prep() %>%
  bake(credit_data)
# }