Create a Formula from a Prepared Recipe

In case a model formula is required, the formula method can be used on a recipe to show what predictors and outcome(s) could be used.

# S3 method for recipe
formula(x, ...)

Arguments

x

A recipe object that has been prepared.

...

Note currently used.

Value

A formula.

Examples

formula(recipe(Species + Sepal.Length ~ ., data = iris) %>% prep())
#> Species + Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width
#> <environment: 0x7fb19d658df8>

iris_rec <- recipe(Species ~ ., data = iris) %>%
  step_center(all_numeric()) %>%
  prep()
formula(iris_rec)
#> Species ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width
#> <environment: 0x7fb1816ecde0>