Role Selection

has_role, predictors, and outcomes can be used to select variables in a formula that have certain roles. Similarly, has_type, numerics, and nominals are used to select columns based on their data type. See selections for more details.

has_role(x = "predictor", roles = NULL)

predictors(roles = NULL)

outcomes(roles = NULL)

has_type(x = "numeric", types = NULL)

numerics(types = NULL)

nominals(types = NULL)

Arguments

x

A single character string for the query.

roles

A character string of roles for the current set of terms.

types

A character string of roles for the current set of data types

Value

An integer vector.