OkCupid Data

These are a sample of columns of users of OkCupid dating website. The data are from Kim and Escobedo-Land (2015). Permission to use this data set was explicitly granted by OkCupid.

Source

Kim, A. Y., and A. Escobedo-Land. 2015. "OkCupid Data for Introductory Statistics and Data Science Courses." Journal of Statistics Education: An International Journal on the Teaching and Learning of Statistics.

Value

okc

a data frame

Examples

data(okc)
str(okc)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	59855 obs. of  6 variables:
#>  $ age     : int  22 35 38 23 29 29 32 31 24 37 ...
#>  $ diet    : chr  "strictly anything" "mostly other" "anything" "vegetarian" ...
#>  $ height  : int  75 70 68 71 66 67 65 65 67 65 ...
#>  $ location: chr  "south san francisco" "oakland" "san francisco" "berkeley" ...
#>  $ date    : Date, format: "2012-06-28" "2012-06-29" ...
#>  $ Class   : Factor w/ 2 levels "stem","other": 2 2 2 2 2 1 2 2 2 2 ...