Internal Functions

These are not to be used directly by the users.

This internal function is used for printing steps.

This internal function takes a list of selectors (e.g. terms in most steps) and returns a character vector version for printing.

yj_trans(x, lambda, eps = 0.001)

estimate_yj(dat, limits = c(-5, 5), num_unique = 5, na_rm = TRUE)

ellipse_check(...)

printer(tr_obj = NULL, untr_obj = NULL, trained = FALSE,
  width = max(20, options()$width - 30))

prepare(x, ...)

is_trained(x)

sel2char(x)

Arguments

x
a step object.

...
Arguments pass in from a call to step

tr_obj
A character vector of names that have been resolved during preparing the recipe (e.g. the columns object of step_log()).

untr_obj
An object of selectors prior to prepping the recipe (e.g. terms in most steps).

trained
A logical for whether the step has been trained.

width
An integer denoting where the output should be wrapped.

x
A list of selectors

Value

If not empty, a list of quosures. If empty, an error is thrown.

`NULL``, invisibly.

A logical

A character vector