Update packages

This will check to see if all required packages are installed.

recipes_pkg_check(pkg = NULL, ...)

Arguments

pkg
A character string for the package being checked

...
Extra arguments to pass to utils::install.packages()

Value

Nothing is returned but a message is printed to the console about which packages (if any) should be installed along with code to do so.

Examples

if (FALSE) {
recipes_pkg_check("recipes")
}