Methods for Select Variables in recipe Formulas

There are a few different methods for selecting variables or model terms using the terms argument for the step functions. The two main requirements are that 1) the value of the terms argument is a formula and 2) that does not contain functions beyond those supported (see below).

The formula is not processed until the learn function for the step is executed. Functions can be used inside of the formula that can select columns of the design matrix that may not currently exist. For example, when using step_pca, the number of columns created by feature extraction may not be known when subsequent steps are defined. In this case, using contains("^PC") will select all of the columns whose names start with "PC".

Standard formulas can be used where only the right-hand side is used (e.g. ~ x1 + x2 + x3). Alternatively, select helpers from the dplyr package, such as starts_with, ends_with, contains, matches, num_range, everything. As an example, ~ contains("x") + y + z is valid.

However, these are the only functions that can be used in the formula. Using other functions will cause an error, such as ~ contains("x") + y + log(z).

While plus signs between formula terms will add columns to the list, minus signs can also be used to exclude columns. For example, ~ contains("x") - x1 would keep all of the columns containing "x" but would exclude any called "x1".

Finally, there are sets of functions that can be used to select variables based on their role or type: has_role and has_type. For convenience, there are also functions that are more specific: numerics, nominals, predictors, and outcomes. These can be used in conjunction with the previous functions described for selecting variables using their names.