Sort rows using dplyr

step_arrange creates a specification of a recipe step that will sort rows using dplyr::arrange().

step_arrange(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  inputs = NULL,
  skip = FALSE,
  id = rand_id("arrange")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

Comma separated list of unquoted variable names. Use `desc()`` to sort a variable in descending order. See dplyr::arrange() for more details.

role

Not used by this step since no new variables are created.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

inputs

Quosure of values given by ....

skip

A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Details

When an object in the user's global environment is referenced in the expression defining the new variable(s), it is a good idea to use quasiquotation (e.g. !!!) to embed the value of the object in the expression (to be portable between sessions). See the examples.

When you tidy() this step, a tibble with column terms which contains the sorting variable(s) or expression(s) is returned. The expressions are text representations and are not parsable.

See also

Examples

rec <- recipe( ~ ., data = iris) %>%
  step_arrange(desc(Sepal.Length), 1/Petal.Length)

prepped <- prep(rec, training = iris %>% slice(1:75))
tidy(prepped, number = 1)
#> # A tibble: 2 × 2
#>   terms              id           
#>   <chr>              <chr>        
#> 1 desc(Sepal.Length) arrange_ut1BT
#> 2 1/Petal.Length     arrange_ut1BT

library(dplyr)

dplyr_train <-
  iris %>%
  as_tibble() %>%
  slice(1:75) %>%
  dplyr::arrange(desc(Sepal.Length), 1/Petal.Length)

rec_train <- bake(prepped, new_data = NULL)
all.equal(dplyr_train, rec_train)
#> [1] TRUE

dplyr_test <-
  iris %>%
  as_tibble() %>%
  slice(76:150) %>%
  dplyr::arrange(desc(Sepal.Length), 1/Petal.Length)
rec_test <- bake(prepped, iris %>% slice(76:150))
all.equal(dplyr_test, rec_test)
#> [1] TRUE

# When you have variables/expressions, you can create a
# list of symbols with `rlang::syms()`` and splice them in
# the call with `!!!`. See https://tidyeval.tidyverse.org

sort_vars <- c("Sepal.Length", "Petal.Length")

qq_rec <-
  recipe( ~ ., data = iris) %>%
  # Embed the `values` object in the call using !!!
  step_arrange(!!!syms(sort_vars)) %>%
  prep(training = iris)

tidy(qq_rec, number = 1)
#> # A tibble: 2 × 2
#>   terms        id           
#>   <chr>        <chr>        
#> 1 Sepal.Length arrange_syrb6
#> 2 Petal.Length arrange_syrb6