Imputation via Bagged Trees

step_bagimpute creates a specification of a recipe step that will create bagged tree models to impute missing data.

step_bagimpute(recipe, ..., role = NA, trained = FALSE,
  impute_with = imp_vars(all_predictors()), trees = 25,
  models = NULL, options = list(keepX = FALSE),
  seed_val = sample.int(10^4, 1), skip = FALSE,
  id = rand_id("bagimpute"))

imp_vars(...)

# S3 method for step_bagimpute
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose variables. For step_bagimpute, this indicates the variables to be imputed. When used with imp_vars, the dots indicates which variables are used to predict the missing data in each variable. See selections() for more details. For the tidy method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

impute_with
A call to imp_vars to specify which variables are used to impute the variables that can include specific variable names separated by commas or different selectors (see selections()). If a column is included in both lists to be imputed and to be an imputation predictor, it will be removed from the latter and not used to impute itself.

trees
An integer for the number bagged trees to use in each model.

models
The ipred::ipredbagg() objects are stored here once this bagged trees have be trained by prep.recipe().

options
A list of options to ipred::ipredbagg(). Defaults are set for the arguments nbagg and keepX but others can be passed in. Note that the arguments X and y should not be passed here.

seed_val
A integer used to create reproducible models. The same seed is used across all imputation models.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_bagimpute object.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms (the selectors or variables selected) and model (the bagged tree object).

Details

For each variables requiring imputation, a bagged tree is created where the outcome is the variable of interest and the predictors are any other variables listed in the impute_with formula. One advantage to the bagged tree is that is can accept predictors that have missing values themselves. This imputation method can be used when the variable of interest (and predictors) are numeric or categorical. Imputed categorical variables will remain categorical.

Note that if a variable that is to be imputed is also in impute_with, this variable will be ignored.

It is possible that missing values will still occur after imputation if a large majority (or all) of the imputing variables are also missing.

References

Kuhn, M. and Johnson, K. (2013). Applied Predictive Modeling. Springer Verlag.

Examples

data("credit_data")

## missing data per column
vapply(credit_data, function(x) mean(is.na(x)), c(num = 0))

#>       Status    Seniority         Home         Time          Age      Marital 
#> 0.0000000000 0.0000000000 0.0013471037 0.0000000000 0.0000000000 0.0002245173 
#>      Records          Job     Expenses       Income       Assets         Debt 
#> 0.0000000000 0.0004490346 0.0000000000 0.0855410867 0.0105523125 0.0040413112 
#>       Amount        Price 
#> 0.0000000000 0.0000000000 

set.seed(342)
in_training <- sample(1:nrow(credit_data), 2000)

credit_tr <- credit_data[ in_training, ]
credit_te <- credit_data[-in_training, ]
missing_examples <- c(14, 394, 565)

rec <- recipe(Price ~ ., data = credit_tr)

impute_rec <- rec %>%
  step_bagimpute(Status, Home, Marital, Job, Income, Assets, Debt)

imp_models <- prep(impute_rec, training = credit_tr)

imputed_te <- bake(imp_models, newdata = credit_te, everything())

credit_te[missing_examples,]

#>      Status Seniority Home Time Age Marital Records   Job Expenses Income
#> 30      bad         0 <NA>   48  37  single      no  <NA>       35     NA
#> 736     bad         1 priv   48  36  single     yes fixed       35    173
#> 1057    bad        24 priv   60  46 married      no fixed       60    165
#>      Assets Debt Amount Price
#> 30       NA   NA   1500  1850
#> 736    7700    0   1000  2008
#> 1057   3500 1300   1400  1530

imputed_te[missing_examples, names(credit_te)]

#> # A tibble: 3 x 14
#>   Status Seniority Home   Time   Age Marital Records Job   Expenses Income
#>   <fct>      <int> <fct> <int> <int> <fct>   <fct>   <fct>    <int>  <dbl>
#> 1 bad            0 other    48    37 single  no      fixed       35   115.
#> 2 bad            1 priv     48    36 single  yes     fixed       35   173 
#> 3 bad           24 priv     60    46 married no      fixed       60   165 
#> # ... with 4 more variables: Assets <dbl>, Debt <dbl>, Amount <int>,
#> #   Price <int>

tidy(impute_rec, number = 1)

#> # A tibble: 7 x 3
#>   terms   model id             
#>   <chr>   <lgl> <chr>          
#> 1 Status  NA    bagimpute_NKNDt
#> 2 Home    NA    bagimpute_NKNDt
#> 3 Marital NA    bagimpute_NKNDt
#> 4 Job     NA    bagimpute_NKNDt
#> 5 Income  NA    bagimpute_NKNDt
#> 6 Assets  NA    bagimpute_NKNDt
#> 7 Debt    NA    bagimpute_NKNDt

tidy(imp_models, number = 1)

#> # A tibble: 7 x 3
#>   terms   model           id             
#>   <chr>   <list>          <chr>          
#> 1 Status  <S3: classbagg> bagimpute_NKNDt
#> 2 Home    <S3: classbagg> bagimpute_NKNDt
#> 3 Marital <S3: classbagg> bagimpute_NKNDt
#> 4 Job     <S3: classbagg> bagimpute_NKNDt
#> 5 Income  <S3: regbagg>   bagimpute_NKNDt
#> 6 Assets  <S3: regbagg>   bagimpute_NKNDt
#> 7 Debt    <S3: regbagg>   bagimpute_NKNDt

## Specifying which variables to imputate with

 impute_rec <- rec %>%
  step_bagimpute(Status, Home, Marital, Job, Income, Assets, Debt,
                 impute_with = imp_vars(Time, Age, Expenses),
                 # for quick execution, nbagg lowered
                 options = list(nbagg = 5, keepX = FALSE))

imp_models <- prep(impute_rec, training = credit_tr)

imputed_te <- bake(imp_models, newdata = credit_te, everything())

credit_te[missing_examples,]

#>      Status Seniority Home Time Age Marital Records   Job Expenses Income
#> 30      bad         0 <NA>   48  37  single      no  <NA>       35     NA
#> 736     bad         1 priv   48  36  single     yes fixed       35    173
#> 1057    bad        24 priv   60  46 married      no fixed       60    165
#>      Assets Debt Amount Price
#> 30       NA   NA   1500  1850
#> 736    7700    0   1000  2008
#> 1057   3500 1300   1400  1530

imputed_te[missing_examples, names(credit_te)]

#> # A tibble: 3 x 14
#>   Status Seniority Home   Time   Age Marital Records Job   Expenses Income
#>   <fct>      <int> <fct> <int> <int> <fct>   <fct>   <fct>    <int>  <dbl>
#> 1 bad            0 pare…    48    37 single  no      fixed       35   129.
#> 2 bad            1 priv     48    36 single  yes     fixed       35   173 
#> 3 bad           24 priv     60    46 married no      fixed       60   165 
#> # ... with 4 more variables: Assets <dbl>, Debt <dbl>, Amount <int>,
#> #   Price <int>

tidy(impute_rec, number = 1)

#> # A tibble: 7 x 3
#>   terms   model id             
#>   <chr>   <lgl> <chr>          
#> 1 Status  NA    bagimpute_eCtKx
#> 2 Home    NA    bagimpute_eCtKx
#> 3 Marital NA    bagimpute_eCtKx
#> 4 Job     NA    bagimpute_eCtKx
#> 5 Income  NA    bagimpute_eCtKx
#> 6 Assets  NA    bagimpute_eCtKx
#> 7 Debt    NA    bagimpute_eCtKx

tidy(imp_models, number = 1)

#> # A tibble: 7 x 3
#>   terms   model           id             
#>   <chr>   <list>          <chr>          
#> 1 Status  <S3: classbagg> bagimpute_eCtKx
#> 2 Home    <S3: classbagg> bagimpute_eCtKx
#> 3 Marital <S3: classbagg> bagimpute_eCtKx
#> 4 Job     <S3: classbagg> bagimpute_eCtKx
#> 5 Income  <S3: regbagg>   bagimpute_eCtKx
#> 6 Assets  <S3: regbagg>   bagimpute_eCtKx
#> 7 Debt    <S3: regbagg>   bagimpute_eCtKx