Create a Factors from A Dummy Variable

step_bin2factor creates a specification of a recipe step that will create a two-level factor from a single dummy variable.

step_bin2factor(recipe, ..., role = NA, trained = FALSE, levels = c("yes",
  "no"), objects = NULL)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
Selector functions that choose which variables will be converted. See selections for more details.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

levels
A length 2 character string that indicate the factor levels for the 1's (in the first position) and the zeros (second)

objects
A vector with the selected variable names. This is NULL until computed by learn.recipe.

Value

step_bin2factor returns an object of class step_bin2factor. The variables selected will be replaced by their factor doppelgängers by process.recipe.

Details

This operation may be useful for situations where a binary piece of information may need to be represented as categorical instead of numeric. For example, naive Bayes models would do better to have factor predictors so that the binomial distribution is modeled in stead of a Gaussian probability density of numeric binary data. Note that the numeric data is only verified to be numeric (and does not count levels).

Examples

data(covers)

rec <- recipe(~ description, covers) %>%
 step_regex(description, pattern = "(rock|stony)", result = "rocks") %>%
 step_regex(description, pattern = "(rock|stony)", result = "more_rocks") %>%
 step_bin2factor(rocks)

rec <- learn(rec, training = covers)

#> step 1 regex training 
#> step 2 regex training 
#> step 3 bin2factor training 

results <- process(rec, newdata = covers)

table(results$rocks, results$more_rocks)

#>      
#>        0  1
#>   yes  0 29
#>   no  11  0