B-Spline Basis Functions

`step_ns` creates a *specification* of a recipe step that will create new columns that are basis expansions of variables using B-splines.

step_bs(recipe, ..., role = "predictor", trained = FALSE, objects = NULL,
  options = list(df = NULL, degree = 3), skip = FALSE)

# S3 method for step_bs
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See [selections()] for more details. For the `tidy` method, these are not currently used.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created from the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

objects
A list of [splines::bs()] objects created once the step has been trained.

options
A list of options for [splines::bs()] which should not include `x`.

skip
A logical. Should the step be skipped when the recipe is baked by [bake.recipe()]? While all operations are baked when [prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations

x
A `step_bs` object.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any). For the `tidy` method, a tibble with columns `terms` which is the columns that will be affected and `holiday`.

Details

`step_bs` can new features from a single variable that enable fitting routines to model this variable in a nonlinear manner. The extent of the possible nonlinearity is determined by the `df`, `degree`, or `knot` arguments of [splines::bs()]. The original variables are removed from the data and new columns are added. The naming convention for the new variables is `varname_bs_1` and so on.

See also

Examples

data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)

with_splines <- rec %>%
  step_bs(carbon, hydrogen)
with_splines <- prep(with_splines, training = biomass_tr)

expanded <- bake(with_splines, biomass_te)
expanded

#> # A tibble: 80 x 10
#>    oxygen nitrogen sulfur   HHV carbon_bs_1 carbon_bs_2 carbon_bs_3
#>     <dbl>    <dbl>  <dbl> <dbl>       <dbl>       <dbl>       <dbl>
#>  1   47.2    0.300  0.220  18.3       0.437      0.273      0.0568 
#>  2   48.1    2.85   0.340  17.6       0.444      0.236      0.0417 
#>  3   49.1    2.40   0.300  17.2       0.444      0.229      0.0394 
#>  4   37.3    1.80   0.500  18.9       0.437      0.273      0.0571 
#>  5   42.8    0.200  0.     20.5       0.427      0.301      0.0707 
#>  6   41.7    0.700  0.200  18.5       0.442      0.248      0.0465 
#>  7   54.1    1.19   0.510  15.1       0.440      0.184      0.0256 
#>  8   33.8    0.950  0.200  16.2       0.444      0.222      0.0369 
#>  9   31.1    0.140  4.90   11.1       0.359      0.0771     0.00552
#> 10   23.7    4.63   1.05   10.8       0.338      0.0643     0.00408
#> # ... with 70 more rows, and 3 more variables: hydrogen_bs_1 <dbl>,
#> #   hydrogen_bs_2 <dbl>, hydrogen_bs_3 <dbl>