Centering Numeric Data

step_center creates a specification of a recipe step that will normalize numeric data to have a mean of zero.

step_center(recipe, ..., role = NA, trained = FALSE, means = NULL,
  na.rm = TRUE)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections for more details.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

means
A named numeric vector of means. This is NULL until computed by learn.recipe.

na.rm
A logical value indicating whether NA values should be removed when averaging.

Value

step_center returns an object of class step_center.

Details

Centering data means that the average of a variable is subtracted from the data. step_center estimates the variable means from the data used in the training argument of learn.recipe. process.recipe then applies the centering to new data sets using these means.

See also

recipe learn.recipe process.recipe

Examples

data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)

center_trans <- rec %>%
  step_center(carbon, contains("gen"), -hydrogen)

center_obj <- learn(center_trans, training = biomass_tr)

#> step 1 center training 

transformed_te <- process(center_obj, biomass_te)

biomass_te[1:10, names(transformed_te)]

#>    carbon hydrogen oxygen nitrogen sulfur
#> 15  46.35     5.67  47.20     0.30   0.22
#> 20  43.25     5.50  48.06     2.85   0.34
#> 26  42.70     5.50  49.10     2.40   0.30
#> 31  46.40     6.10  37.30     1.80   0.50
#> 36  48.76     6.32  42.77     0.20   0.00
#> 41  44.30     5.50  41.70     0.70   0.20
#> 46  38.94     5.23  54.13     1.19   0.51
#> 51  42.10     4.66  33.80     0.95   0.20
#> 55  29.20     4.40  31.10     0.14   4.90
#> 65  27.80     3.77  23.69     4.63   1.05

transformed_te

#> # A tibble: 80 × 5
#>         carbon hydrogen     oxygen   nitrogen sulfur
#>          <dbl>    <dbl>      <dbl>      <dbl>  <dbl>
#>  1  -2.0041447     5.67   8.675329 -0.7745921   0.22
#>  2  -5.1041447     5.50   9.535329  1.7754079   0.34
#>  3  -5.6541447     5.50  10.575329  1.3254079   0.30
#>  4  -1.9541447     6.10  -1.224671  0.7254079   0.50
#>  5   0.4058553     6.32   4.245329 -0.8745921   0.00
#>  6  -4.0541447     5.50   3.175329 -0.3745921   0.20
#>  7  -9.4141447     5.23  15.605329  0.1154079   0.51
#>  8  -6.2541447     4.66  -4.724671 -0.1245921   0.20
#>  9 -19.1541447     4.40  -7.424671 -0.9345921   4.90
#> 10 -20.5541447     3.77 -14.834671  3.5554079   1.05
#> # ... with 70 more rows