Distances to Class Centroids

step_classdist creates a a specification of a recipe step that will convert numeric data into Mahalanobis distance measurements to the data centroid. This is done for each value of a categorical class variable.

step_classdist(recipe, terms, class, role = "predictor", trained = FALSE,
  mean_func = mean, cov_func = cov, pool = FALSE, log = TRUE,
  objects = NULL)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

terms

A representation of the variables or terms that will be used to create the new features.

class

A single character string or formula that that specifies a single categorical variable to be used as the class.

role

For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that resulting distances will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

mean_func

A function to compute the center of the distribution.

cov_func

A function that computes the covariance matrix

pool

A logical: should the covariance matrix be computed by pooling the data for all of the classes?

log

A logical: should the distances be transformed by the natural log function?

objects

Statistics are stored here once this step has been trained by learn.recipe.

Value

step_classdist returns an object of class step_classdist.

Details

step_classdist will create a

The function will create a new column for every unique value of the class variable. The resulting variables will not replace the original values and have the prefix classdist_.

Note that, by default, the default covariance function requires that each class should have at least as many rows as variables listed in the terms argument. If pool = TRUE, there must be at least as many data points are variables overall.

Examples

library(magrittr)

# in case of missing data...
mean2 <- function(x) mean(x, na.rm = TRUE)

rec <- recipe(Species ~ ., data = iris) %>%
  step_classdist(~ predictors(), class = ~ Species,
                 pool = FALSE, mean_func = mean2)

rec_dists <- learn(rec, training = iris)

#> step 1 classdist training 

dists_to_species <- process(rec_dists, newdata = iris)
## on log scale:
dists_to_species[, c("Species", grep("classdist", names(dists_to_species), value = TRUE))]

#> # A tibble: 150 × 4
#>    Species classdist_setosa classdist_versicolor classdist_virginica
#>     <fctr>            <dbl>                <dbl>               <dbl>
#> 1   setosa       -0.8004790             4.743231            5.209136
#> 2   setosa        0.7328938             4.422633            5.036790
#> 3   setosa        0.2502412             4.553039            5.078257
#> 4   setosa        0.5342728             4.416172            4.946213
#> 5   setosa       -0.2722217             4.791492            5.215137
#> 6   setosa        1.3117452             4.791484            5.211114
#> 7   setosa        1.2308667             4.564384            5.037074
#> 8   setosa       -1.0687451             4.642609            5.120650
#> 9   setosa        1.0974371             4.303348            4.888585
#> 10  setosa        1.1631777             4.537918            5.054580
#> # ... with 140 more rows