High Correlation Filter

step_corr creates a specification of a recipe step that will potentially remove variables that have large absolute correlations with other variables.

step_corr(recipe, terms, role = NA, trained = FALSE, threshold = 0.9,
  use = "pairwise.complete.obs", method = "pearson", removals = NULL)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

terms

A representation of the variables or terms that will evaluated by the filtering process.

role

Not used by this step since no new variables are created.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

threshold

A value for the threshold of absolute correlation values. The step will try to remove the minimum number of columns so that all the resulting absolute correlations are less than this value.

use

A character string for the use argument to the cor function.

method

A character string for the method argument to the cor function.

removals

A character string that contains the names of columns that should be removed. These values are not determined until learn.recipe is called.

Value

step_corr returns an object of class step_corr.

Details

This step attempts to remove variables to keep the largest absolute correlation between the variables less than threshold.

See also

step_nzv recipe learn.recipe process.recipe

Examples

data(biomass)

set.seed(3535)
biomass$duplicate <- biomass$carbon + rnorm(nrow(biomass))

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur + duplicate,
              data = biomass_tr)

library(magrittr)
corr_filter <- rec %>%
  step_corr(terms = ~ is_predictor(), threshold = .5)

filter_obj <- learn(corr_filter, training = biomass_tr)

#> step 1 corr training 

filtered_te <- process(filter_obj, biomass_te, role = "predictor")
round(abs(cor(biomass_tr[, c(3:7, 9)])), 2)

#>           carbon hydrogen oxygen nitrogen sulfur duplicate
#> carbon      1.00     0.32   0.63     0.15   0.09      1.00
#> hydrogen    0.32     1.00   0.54     0.07   0.19      0.31
#> oxygen      0.63     0.54   1.00     0.18   0.31      0.63
#> nitrogen    0.15     0.07   0.18     1.00   0.27      0.15
#> sulfur      0.09     0.19   0.31     0.27   1.00      0.10
#> duplicate   1.00     0.31   0.63     0.15   0.10      1.00

round(abs(cor(filtered_te)), 2)

#>           hydrogen nitrogen sulfur duplicate
#> hydrogen      1.00     0.11   0.26      0.20
#> nitrogen      0.11     1.00   0.16      0.13
#> sulfur        0.26     0.16   1.00      0.13
#> duplicate     0.20     0.13   0.13      1.00