High Correlation Filter

`step_corr` creates a *specification* of a recipe step that will potentially remove variables that have large absolute correlations with other variables.

step_corr(recipe, ..., role = NA, trained = FALSE, threshold = 0.9,
  use = "pairwise.complete.obs", method = "pearson", removals = NULL,
  skip = FALSE)

# S3 method for step_corr
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See [selections()] for more details. For the `tidy` method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

threshold
A value for the threshold of absolute correlation values. The step will try to remove the minimum number of columns so that all the resulting absolute correlations are less than this value.

use
A character string for the `use` argument to the [stats::cor()] function.

method
A character string for the `method` argument to the [stats::cor()] function.

removals
A character string that contains the names of columns that should be removed. These values are not determined until [prep.recipe()] is called.

skip
A logical. Should the step be skipped when the recipe is baked by [bake.recipe()]? While all operations are baked when [prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations

x
A `step_corr` object.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any). For the `tidy` method, a tibble with columns `terms` which is the columns that will be removed.

Details

This step attempts to remove variables to keep the largest absolute correlation between the variables less than `threshold`.

See also

Examples

data(biomass)

set.seed(3535)
biomass$duplicate <- biomass$carbon + rnorm(nrow(biomass))

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen +
                    sulfur + duplicate,
              data = biomass_tr)

corr_filter <- rec %>%
  step_corr(all_predictors(), threshold = .5)

filter_obj <- prep(corr_filter, training = biomass_tr)

filtered_te <- bake(filter_obj, biomass_te)
round(abs(cor(biomass_tr[, c(3:7, 9)])), 2)

#>           carbon hydrogen oxygen nitrogen sulfur duplicate
#> carbon      1.00     0.32   0.63     0.15   0.09      1.00
#> hydrogen    0.32     1.00   0.54     0.07   0.19      0.31
#> oxygen      0.63     0.54   1.00     0.18   0.31      0.63
#> nitrogen    0.15     0.07   0.18     1.00   0.27      0.15
#> sulfur      0.09     0.19   0.31     0.27   1.00      0.10
#> duplicate   1.00     0.31   0.63     0.15   0.10      1.00

round(abs(cor(filtered_te)), 2)

#>           hydrogen nitrogen sulfur  HHV duplicate
#> hydrogen      1.00     0.11   0.26 0.10      0.20
#> nitrogen      0.11     1.00   0.16 0.11      0.13
#> sulfur        0.26     0.16   1.00 0.08      0.13
#> HHV           0.10     0.11   0.08 1.00      0.94
#> duplicate     0.20     0.13   0.13 0.94      1.00

tidy(corr_filter, number = 1)

#> # A tibble: 1 x 1
#>   terms
#>   <chr>
#> 1 <NA> 

tidy(filter_obj, number = 1)

#> # A tibble: 2 x 1
#>   terms 
#>   <chr> 
#> 1 oxygen
#> 2 carbon