Create Counts of Patterns using Regular Expressions

step_count creates a specification of a recipe step that will create a variable that counts instances of a regular expression pattern in text.

step_count(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  pattern = ".",
  normalize = FALSE,
  options = list(),
  result = make.names(pattern),
  input = NULL,
  skip = FALSE,
  id = rand_id("count")
)

# S3 method for step_count
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
A single selector functions to choose which variable will be searched for the pattern. The selector should resolve into a single variable. See selections() for more details. For the tidy method, these are not currently used.

role
For a variable created by this step, what analysis role should they be assigned?. By default, the function assumes that the new dummy variable column created by the original variable will be used as a predictor in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

pattern
A character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector. Coerced by as.character to a character string if possible.

normalize
A logical; should the integer counts be divided by the total number of characters in the string?.

options
A list of options to gregexpr() that should not include x or pattern.

result
A single character value for the name of the new variable. It should be a valid column name.

input
A single character value for the name of the variable being searched. This is NULL until computed by prep.recipe().

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_count object.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms (the selectors or variables selected) and result (the new column name).

Examples

library(modeldata)
data(covers)

rec <- recipe(~ description, covers) %>%
  step_count(description, pattern = "(rock|stony)", result = "rocks") %>%
  step_count(description, pattern = "famil", normalize = TRUE)

rec2 <- prep(rec, training = covers)
rec2

#> Data Recipe
#> 
#> Inputs:
#> 
#>       role #variables
#>  predictor          1
#> 
#> Training data contained 40 data points and no missing data.
#> 
#> Operations:
#> 
#> Regular expression counts using `(rock|stony)` [trained]
#> Regular expression counts using `famil` [trained]

count_values <- bake(rec2, new_data = covers)
count_values

#> # A tibble: 40 x 3
#>    description                                                rocks  famil
#>    <fct>                                                      <int>  <dbl>
#>  1 1,cathedral family,rock outcrop complex,extremely stony        2 0.0182
#>  2 2,vanet,ratake families complex,very stony                     1 0.0238
#>  3 3,haploborolis,rock outcrop complex,rubbly                     1 0     
#>  4 4,ratake family,rock outcrop complex,rubbly                    1 0.0233
#>  5 5,vanet family,rock outcrop complex complex,rubbly             1 0.02  
#>  6 6,vanet,wetmore families,rock outcrop complex,stony            2 0.0196
#>  7 7,gothic family                                                0 0.0667
#>  8 8,supervisor,limber families complex                           0 0.0278
#>  9 9,troutville family,very stony                                 1 0.0333
#> 10 10,bullwark,catamount families,rock outcrop complex,rubbly     1 0.0172
#> # … with 30 more rows

tidy(rec, number = 1)

#> # A tibble: 1 x 3
#>   terms       result id         
#>   <chr>       <chr>  <chr>      
#> 1 description NA     count_HX7KJ

tidy(rec2, number = 1)

#> # A tibble: 1 x 3
#>   terms       result id         
#>   <chr>       <chr>  <chr>      
#> 1 description rocks  count_HX7KJ