Date Feature Generator

`step_date` creates a a *specification* of a recipe step that will convert date data into one or more factor or numeric variables.

step_date(recipe, ..., role = "predictor", trained = FALSE,
  features = c("dow", "month", "year"), abbr = TRUE, label = TRUE,
  ordinal = FALSE, columns = NULL, skip = FALSE)

# S3 method for step_date
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables that will be used to create the new variables. The selected variables should have class `Date` or `POSIXct`. See [selections()] for more details. For the `tidy` method, these are not currently used.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new variable columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

features
A character string that includes at least one of the following values: `month`, `dow` (day of week), `doy` (day of year), `week`, `month`, `decimal` (decimal date, e.g. 2002.197), `quarter`, `semester`, `year`.

abbr
A logical. Only available for features `month` or `dow`. `FALSE` will display the day of the week as an ordered factor of character strings, such as "Sunday". `TRUE` will display an abbreviated version of the label, such as "Sun". `abbr` is disregarded if `label = FALSE`.

label
A logical. Only available for features `month` or `dow`. `TRUE` will display the day of the week as an ordered factor of character strings, such as "Sunday." `FALSE` will display the day of the week as a number.

ordinal
A logical: should factors be ordered? Only available for features `month` or `dow`.

columns
A character string of variables that will be used as inputs. This field is a placeholder and will be populated once [prep.recipe()] is used.

skip
A logical. Should the step be skipped when the recipe is baked by [bake.recipe()]? While all operations are baked when [prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations

x
A `step_date` object.

Value

For `step_date`, an updated version of recipe with the new step added to the sequence of existing steps (if any). For the `tidy` method, a tibble with columns `terms` (the selectors or variables selected), `value` (the feature names), and `ordinal` (a logical).

Details

Unlike other steps, `step_date` does *not* remove the original date variables. [step_rm()] can be used for this purpose.

See also

Examples

library(lubridate)

#> Warning: package ‘lubridate’ was built under R version 3.4.4

#> 
#> Attaching package: ‘lubridate’

#> The following object is masked from ‘package:base’:
#> 
#>     date

examples <- data.frame(Dan = ymd("2002-03-04") + days(1:10),
                       Stefan = ymd("2006-01-13") + days(1:10))
date_rec <- recipe(~ Dan + Stefan, examples) %>%
   step_date(all_predictors())

tidy(date_rec, number = 1)

#> # A tibble: 3 x 3
#>   terms            value ordinal
#>   <fct>            <fct> <lgl>  
#> 1 all_predictors() dow   FALSE  
#> 2 all_predictors() month FALSE  
#> 3 all_predictors() year  FALSE  

date_rec <- prep(date_rec, training = examples)

date_values <- bake(date_rec, newdata = examples)
date_values

#> # A tibble: 10 x 8
#>    Dan        Stefan     Dan_dow Dan_month Dan_year Stefan_dow Stefan_month
#>    <date>     <date>     <fct>   <fct>        <dbl> <fct>      <fct>       
#>  1 2002-03-05 2006-01-14 Tue     Mar          2002. Sat        Jan         
#>  2 2002-03-06 2006-01-15 Wed     Mar          2002. Sun        Jan         
#>  3 2002-03-07 2006-01-16 Thu     Mar          2002. Mon        Jan         
#>  4 2002-03-08 2006-01-17 Fri     Mar          2002. Tue        Jan         
#>  5 2002-03-09 2006-01-18 Sat     Mar          2002. Wed        Jan         
#>  6 2002-03-10 2006-01-19 Sun     Mar          2002. Thu        Jan         
#>  7 2002-03-11 2006-01-20 Mon     Mar          2002. Fri        Jan         
#>  8 2002-03-12 2006-01-21 Tue     Mar          2002. Sat        Jan         
#>  9 2002-03-13 2006-01-22 Wed     Mar          2002. Sun        Jan         
#> 10 2002-03-14 2006-01-23 Thu     Mar          2002. Mon        Jan         
#> # ... with 1 more variable: Stefan_year <dbl>

tidy(date_rec, number = 1)

#> # A tibble: 6 x 3
#>   terms  value ordinal
#>   <fct>  <fct> <lgl>  
#> 1 Dan    dow   FALSE  
#> 2 Stefan dow   FALSE  
#> 3 Dan    month FALSE  
#> 4 Stefan month FALSE  
#> 5 Dan    year  FALSE  
#> 6 Stefan year  FALSE