Data Depths

step_depth creates a a specification of a recipe step that will convert numeric data into measurement of data depth. This is done for each value of a categorical class variable.

step_depth(recipe, terms, class, role = "predictor", trained = FALSE,
  metric = "halfspace", options = list(), data = NULL)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

terms

A representation of the variables or terms that will be used to create the new features.

class

A single character string or formula that that specifies a single categorical variable to be used as the class.

role

For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that resulting depth estimates will be used as all_predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

metric

A character string specifying the depth metric. Possible values are "potential", "halfspace", "Mahalanobis", "simplicialVolume", "spatial", and "zonoid".

options

A list of options to pass to the underlying depth functions. See depth.halfspace, depth.Mahalanobis, depth.potential, depth.projection, depth.simplicial, depth.simplicialVolume, depth.spatial, depth.zonoid.

data

The training data are stored here once after learn.recipe is executed.

Value

step_depth returns an object of class step_depth.

Details

step_depth will create a

Data depth metrics attempt to measure how close data a data point is to the center of its distribution. There are a number of methods for calculating death but a simple example is the inverse of the distance of a data point to the centroid of the distribution. Generally, small values indicate that a data point not close to the centroid. step_depth can compute a class-specific depth for a new data point based on the proximity of the new value to the training set distribution.

Note that the entire training set is saved to compute future depth values. The saved data have been learned and processed up to the point before the location that step_depth occupies in the recipe. Also, the data requirements for the different step methods may vary. For example, using metric = "Mahalanobis" requires that each class should have at least as many rows as variables listed in the terms argument.

The function will create a new column for every unique value of the class variable. The resulting variables will not replace the original values and have the prefix depth_.

Examples

library(magrittr)

# halfspace depth is the default
rec <- recipe(Species ~ ., data = iris) %>%
  step_depth(~ all_predictors(), class = ~ Species)

rec_dists <- learn(rec, training = iris)

#> step 1 depth training 

dists_to_species <- process(rec_dists, newdata = iris)
dists_to_species

#> # A tibble: 150 × 7
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width depth_setosa
#>           <dbl>       <dbl>        <dbl>       <dbl>        <dbl>
#> 1           5.1         3.5          1.4         0.2         0.20
#> 2           4.9         3.0          1.4         0.2         0.04
#> 3           4.7         3.2          1.3         0.2         0.10
#> 4           4.6         3.1          1.5         0.2         0.08
#> 5           5.0         3.6          1.4         0.2         0.14
#> 6           5.4         3.9          1.7         0.4         0.04
#> 7           4.6         3.4          1.4         0.3         0.02
#> 8           5.0         3.4          1.5         0.2         0.28
#> 9           4.4         2.9          1.4         0.2         0.02
#> 10          4.9         3.1          1.5         0.1         0.02
#> # ... with 140 more rows, and 2 more variables: depth_versicolor <dbl>,
#> #   depth_virginica <dbl>