Data Depths

`step_depth` creates a a *specification* of a recipe step that will convert numeric data into measurement of *data depth*. This is done for each value of a categorical class variable.

step_depth(recipe, ..., class, role = "predictor", trained = FALSE,
  metric = "halfspace", options = list(), data = NULL, skip = FALSE)

# S3 method for step_depth
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables that will be used to create the new features. See [selections()] for more details. For the `tidy` method, these are not currently used.

class
A single character string that specifies a single categorical variable to be used as the class.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that resulting depth estimates will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

metric
A character string specifying the depth metric. Possible values are "potential", "halfspace", "Mahalanobis", "simplicialVolume", "spatial", and "zonoid".

options
A list of options to pass to the underlying depth functions. See [ddalpha::depth.halfspace()], [ddalpha::depth.Mahalanobis()], [ddalpha::depth.potential()], [ddalpha::depth.projection()], [ddalpha::depth.simplicial()], [ddalpha::depth.simplicialVolume()], [ddalpha::depth.spatial()], [ddalpha::depth.zonoid()].

data
The training data are stored here once after [prep.recipe()] is executed.

skip
A logical. Should the step be skipped when the recipe is baked by [bake.recipe()]? While all operations are baked when [prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations

x
A `step_depth` object.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any). For the `tidy` method, a tibble with columns `terms` (the selectors or variables selected) and `class`.

Details

Data depth metrics attempt to measure how close data a data point is to the center of its distribution. There are a number of methods for calculating death but a simple example is the inverse of the distance of a data point to the centroid of the distribution. Generally, small values indicate that a data point not close to the centroid. `step_depth` can compute a class-specific depth for a new data point based on the proximity of the new value to the training set distribution.

Note that the entire training set is saved to compute future depth values. The saved data have been trained (i.e. prepared) and baked (i.e. processed) up to the point before the location that `step_depth` occupies in the recipe. Also, the data requirements for the different step methods may vary. For example, using `metric = "Mahalanobis"` requires that each class should have at least as many rows as variables listed in the `terms` argument.

The function will create a new column for every unique value of the `class` variable. The resulting variables will not replace the original values and have the prefix `depth_`.

Examples

# halfspace depth is the default
rec <- recipe(Species ~ ., data = iris) %>%
  step_depth(all_predictors(), class = "Species")

rec_dists <- prep(rec, training = iris)

dists_to_species <- bake(rec_dists, newdata = iris)
dists_to_species

#> # A tibble: 150 x 8
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species depth_setosa
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>          <dbl>
#>  1         5.10        3.50         1.40       0.200 setosa        0.180 
#>  2         4.90        3.00         1.40       0.200 setosa        0.0400
#>  3         4.70        3.20         1.30       0.200 setosa        0.100 
#>  4         4.60        3.10         1.50       0.200 setosa        0.0800
#>  5         5.00        3.60         1.40       0.200 setosa        0.140 
#>  6         5.40        3.90         1.70       0.400 setosa        0.0400
#>  7         4.60        3.40         1.40       0.300 setosa        0.0400
#>  8         5.00        3.40         1.50       0.200 setosa        0.260 
#>  9         4.40        2.90         1.40       0.200 setosa        0.0200
#> 10         4.90        3.10         1.50       0.100 setosa        0.0200
#> # ... with 140 more rows, and 2 more variables: depth_versicolor <dbl>,
#> #   depth_virginica <dbl>

tidy(rec, number = 1)

#> # A tibble: 1 x 2
#>   terms            class
#>   <chr>            <chr>
#> 1 all_predictors() <NA> 

tidy(rec_dists, number = 1)

#> # A tibble: 4 x 2
#>   terms        class  
#>   <chr>        <chr>  
#> 1 Sepal.Length Species
#> 2 Sepal.Width  Species
#> 3 Petal.Length Species
#> 4 Petal.Width  Species