Data Depths

step_depth creates a specification of a recipe step that will convert numeric data into measurement of data depth. This is done for each value of a categorical class variable.

step_depth(
  recipe,
  ...,
  class,
  role = "predictor",
  trained = FALSE,
  metric = "halfspace",
  options = list(),
  data = NULL,
  prefix = "depth_",
  skip = FALSE,
  id = rand_id("depth")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose variables for this step. See selections() for more details.

class

A single character string that specifies a single categorical variable to be used as the class.

role

For model terms created by this step, what analysis role should they be assigned? By default, the new columns created by this step from the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

metric

A character string specifying the depth metric. Possible values are "potential", "halfspace", "Mahalanobis", "simplicialVolume", "spatial", and "zonoid".

options

A list of options to pass to the underlying depth functions. See ddalpha::depth.halfspace(), ddalpha::depth.Mahalanobis(), ddalpha::depth.potential(), ddalpha::depth.projection(), ddalpha::depth.simplicial(), ddalpha::depth.simplicialVolume(), ddalpha::depth.spatial(), ddalpha::depth.zonoid().

data

The training data are stored here once after prep.recipe() is executed.

prefix

A character string for the prefix of the resulting new variables. See notes below.

skip

A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Details

Data depth metrics attempt to measure how close data a data point is to the center of its distribution. There are a number of methods for calculating depth but a simple example is the inverse of the distance of a data point to the centroid of the distribution. Generally, small values indicate that a data point not close to the centroid. step_depth can compute a class-specific depth for a new data point based on the proximity of the new value to the training set distribution.

This step requires the ddalpha package. If not installed, the step will stop with a note about installing the package.

Note that the entire training set is saved to compute future depth values. The saved data have been trained (i.e. prepared) and baked (i.e. processed) up to the point before the location that step_depth occupies in the recipe. Also, the data requirements for the different step methods may vary. For example, using metric = "Mahalanobis" requires that each class should have at least as many rows as variables listed in the terms argument.

The function will create a new column for every unique value of the class variable. The resulting variables will not replace the original values and by default have the prefix depth_. The naming format can be changed using the prefix argument.

When you tidy() this step, a tibble with columns terms (the selectors or variables selected) and class is returned.

See also

Examples

# halfspace depth is the default
rec <- recipe(Species ~ ., data = iris) %>%
  step_depth(all_numeric_predictors(), class = "Species")

# use zonoid metric instead
# also, define naming convention for new columns
rec <- recipe(Species ~ ., data = iris) %>%
  step_depth(all_numeric_predictors(), class = "Species",
             metric = "zonoid", prefix = "zonoid_")

rec_dists <- prep(rec, training = iris)

dists_to_species <- bake(rec_dists, new_data = iris)
dists_to_species
#> # A tibble: 150 × 8
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species zonoid_setosa
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>           <dbl>
#>  1          5.1         3.5          1.4         0.2 setosa         0.559 
#>  2          4.9         3            1.4         0.2 setosa         0.167 
#>  3          4.7         3.2          1.3         0.2 setosa         0.299 
#>  4          4.6         3.1          1.5         0.2 setosa         0.170 
#>  5          5           3.6          1.4         0.2 setosa         0.391 
#>  6          5.4         3.9          1.7         0.4 setosa         0.0838
#>  7          4.6         3.4          1.4         0.3 setosa         0.0200
#>  8          5           3.4          1.5         0.2 setosa         0.645 
#>  9          4.4         2.9          1.4         0.2 setosa         0.0200
#> 10          4.9         3.1          1.5         0.1 setosa         0.0200
#> # … with 140 more rows, and 2 more variables: zonoid_versicolor <dbl>,
#> #   zonoid_virginica <dbl>

tidy(rec, number = 1)
#> # A tibble: 1 × 3
#>   terms                    class id         
#>   <chr>                    <chr> <chr>      
#> 1 all_numeric_predictors() NA    depth_eCCL8
tidy(rec_dists, number = 1)
#> # A tibble: 4 × 3
#>   terms        class   id         
#>   <chr>        <chr>   <chr>      
#> 1 Sepal.Length Species depth_eCCL8
#> 2 Sepal.Width  Species depth_eCCL8
#> 3 Petal.Length Species depth_eCCL8
#> 4 Petal.Width  Species depth_eCCL8