Down-Sample a Data Set Based on a Factor Variable

`step_downsample` creates a *specification* of a recipe step that will remove rows of a data set to make the occurrence of levels in a specific factor level equal.

step_downsample(recipe, ..., ratio = 1, role = NA, trained = FALSE,
  column = NULL, target = NA, skip = TRUE)

# S3 method for step_downsample
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variable is used to sample the data. See [selections()] for more details. The selection should result in _single factor variable_. For the `tidy` method, these are not currently used.

ratio
A numeric value for the ratio of the minority-to-majority frequencies. The default value (1) means that all other levels are sampled down to have the same frequency as the least occurring level. A value of 2 would mean that the majority levels will have (at most) (approximately) twice as many rows than the minority level.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

column
A character string of the variable name that will be populated (eventually) by the `...` selectors.

target
An integer that will be used to subsample. This should not be set by the user and will be populated by `prep`.

skip
A logical. Should the step be skipped when the recipe is baked by [bake.recipe()]? While all operations are baked when [prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations

x
A `step_downsample` object.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any). For the `tidy` method, a tibble with columns `terms` which is the variable used to sample.

Details

Down-sampling is intended to be performed on the _training_ set alone. For this reason, the default is `skip = TRUE`. It is advisable to use `prep(recipe, retain = TRUE)` when preparing the recipe; in this way [juice()] can be used to obtain the down-sampled version of the data.

If there are missing values in the factor variable that is used to define the sampling, missing data are selected at random in the same way that the other factor levels are sampled. Missing values are not used to determine the amount of data in the minority level

For any data with factor levels occurring with the same frequency as the minority level, all data will be retained.

All columns in the data are sampled and returned by [juice()] and [bake()].

Keep in mind that the location of down-sampling in the step may have effects. For example, if centering and scaling, it is not clear whether those operations should be conducted _before_ or _after_ rows are removed.

Examples

data(okc)

sort(table(okc$diet, useNA = "always"))

#> 
#>               halal              kosher      strictly halal     strictly kosher 
#>                  11                  11                  18                  18 
#>        mostly halal       mostly kosher               vegan      strictly vegan 
#>                  48                  86                 136                 227 
#>               other        mostly vegan      strictly other          vegetarian 
#>                 331                 335                 450                 665 
#> strictly vegetarian        mostly other   mostly vegetarian   strictly anything 
#>                 874                1004                3438                5107 
#>            anything     mostly anything                <NA> 
#>                6174               16562               24360 

ds_rec <- recipe( ~ ., data = okc) %>%
  step_downsample(diet) %>%
  prep(training = okc, retain = TRUE)

table(juice(ds_rec)$diet, useNA = "always")

#> 
#>            anything               halal              kosher     mostly anything 
#>                  11                  11                  11                  11 
#>        mostly halal       mostly kosher        mostly other        mostly vegan 
#>                  11                  11                  11                  11 
#>   mostly vegetarian               other   strictly anything      strictly halal 
#>                  11                  11                  11                  11 
#>     strictly kosher      strictly other      strictly vegan strictly vegetarian 
#>                  11                  11                  11                  11 
#>               vegan          vegetarian                <NA> 
#>                  11                  11                  11 

# since `skip` defaults to TRUE, baking the step has no effect
baked_okc <- bake(ds_rec, newdata = okc)
table(baked_okc$diet, useNA = "always")

#> 
#>            anything               halal              kosher     mostly anything 
#>                6174                  11                  11               16562 
#>        mostly halal       mostly kosher        mostly other        mostly vegan 
#>                  48                  86                1004                 335 
#>   mostly vegetarian               other   strictly anything      strictly halal 
#>                3438                 331                5107                  18 
#>     strictly kosher      strictly other      strictly vegan strictly vegetarian 
#>                  18                 450                 227                 874 
#>               vegan          vegetarian                <NA> 
#>                 136                 665               24360