Dummy Variables Creation

step_dummy creates a a specification of a recipe step that will convert nominal data (e.g. character or factors) into one or more numeric binary model terms for the levels of the original data.

step_dummy(recipe, ..., role = "predictor", trained = FALSE,
  contrast = options("contrasts"), naming = function(var, lvl) paste(var,
  make.names(lvl), sep = "_"), levels = NULL)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose which variables will be used to create the dummy variables. See selections for more details.

role

For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the binary dummy variable columns created by the original variables will be used as all_predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

contrast

A specification for which type of contrast should be used to make a set of full rank dummy variables. See contrasts for more details. not currently working

naming

A function that defines the naming convention for new binary columns. See Details below.

levels

A list that contains the information needed to create dummy variables for each variable contained in terms. This is NULL until the step is trained by learn.recipe.

Value

step_dummy returns an object of class step_dummy.

Details

step_dummy will create a set of binary dummy variables from a factor variable. For example, if a factor column in the data set has levels of "red", "green", "blue", the dummy variable process will create two additional columns of 0/1 data for two of those three values (and remove the original column).

By default, the missing dummy variable will correspond to the first level of the factor being converted.

The function allows for non-standard naming of the resulting variables. For a factor named x, with levels "a" and "b", the default naming convention would be to create a new variable called x_b. Note that if the factor levels are not valid variable names (e.g. "some text with spaces"), it will be changed by make.names to be valid (see the example below). The naming format can be changed using the naming argument.

Examples

data(okc)
okc <- okc[complete.cases(okc),]

rec <- recipe(~ diet + age + height, data = okc)

dummies <- rec %>% step_dummy(diet)
dummies <- learn(dummies, training = okc)

#> step 1 dummy training 

#> Error in x$variable: $ operator is invalid for atomic vectors

dummy_data <- process(dummies, newdata = okc)

unique(okc$diet)

#>  [1] "strictly anything"   "mostly other"        "anything"           
#>  [4] "vegetarian"          "mostly anything"     "mostly vegetarian"  
#>  [7] "strictly vegan"      "strictly vegetarian" "mostly vegan"       
#> [10] "strictly other"      "mostly halal"        "other"              
#> [13] "vegan"               "mostly kosher"       "strictly halal"     
#> [16] "halal"               "strictly kosher"     "kosher"             

grep("^diet", names(dummy_data), value = TRUE)

#> [1] "diet"