Dummy Variables Creation

`step_dummy` creates a a *specification* of a recipe step that will convert nominal data (e.g. character or factors) into one or more numeric binary model terms for the levels of the original data.

step_dummy(recipe, ..., role = "predictor", trained = FALSE,
  contrast = options("contrasts"), naming = dummy_names, levels = NULL,
  skip = FALSE)

# S3 method for step_dummy
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables will be used to create the dummy variables. See [selections()] for more details. The selected variables must be factors. For the `tidy` method, these are not currently used.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the binary dummy variable columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

contrast
A specification for which type of contrast should be used to make a set of full rank dummy variables. See [stats::contrasts()] for more details. **not currently working**

naming
A function that defines the naming convention for new dummy columns. See Details below.

levels
A list that contains the information needed to create dummy variables for each variable contained in `terms`. This is `NULL` until the step is trained by [prep.recipe()].

skip
A logical. Should the step be skipped when the recipe is baked by [bake.recipe()]? While all operations are baked when [prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations

x
A `step_dummy` object.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any). For the `tidy` method, a tibble with columns `terms` (the selectors or variables selected).

Details

`step_dummy` will create a set of binary dummy variables from a factor variable. For example, if an unordered factor column in the data set has levels of "red", "green", "blue", the dummy variable bake will create two additional columns of 0/1 data for two of those three values (and remove the original column). For ordered factors, polynomial contrasts are used to encode the numeric values.

By default, the missing dummy variable (i.e. the reference cell) will correspond to the first level of the unordered factor being converted.

The function allows for non-standard naming of the resulting variables. For an unordered factor named `x`, with levels `"a"` and `"b"`, the default naming convention would be to create a new variable called `x_b`. Note that if the factor levels are not valid variable names (e.g. "some text with spaces"), it will be changed by [base::make.names()] to be valid (see the example below). The naming format can be changed using the `naming` argument and the function [dummy_names()] is the default. This function will also change the names of ordinal dummy variables. Instead of values such as "`.L`", "`.Q`", or "`^4`", ordinal dummy variables are given simple integer suffixes such as "`_1`", "`_2`", etc.

See also

Examples

data(okc)
okc <- okc[complete.cases(okc),]

rec <- recipe(~ diet + age + height, data = okc)

dummies <- rec %>% step_dummy(diet)
dummies <- prep(dummies, training = okc)

dummy_data <- bake(dummies, newdata = okc)

unique(okc$diet)

#>  [1] "strictly anything"   "mostly other"        "anything"           
#>  [4] "vegetarian"          "mostly anything"     "mostly vegetarian"  
#>  [7] "strictly vegan"      "strictly vegetarian" "mostly vegan"       
#> [10] "strictly other"      "mostly halal"        "other"              
#> [13] "vegan"               "mostly kosher"       "strictly halal"     
#> [16] "halal"               "strictly kosher"     "kosher"             

grep("^diet", names(dummy_data), value = TRUE)

#>  [1] "diet_halal"               "diet_kosher"             
#>  [3] "diet_mostly.anything"     "diet_mostly.halal"       
#>  [5] "diet_mostly.kosher"       "diet_mostly.other"       
#>  [7] "diet_mostly.vegan"        "diet_mostly.vegetarian"  
#>  [9] "diet_other"               "diet_strictly.anything"  
#> [11] "diet_strictly.halal"      "diet_strictly.kosher"    
#> [13] "diet_strictly.other"      "diet_strictly.vegan"     
#> [15] "diet_strictly.vegetarian" "diet_vegan"              
#> [17] "diet_vegetarian"         

tidy(dummies, number = 1)

#> # A tibble: 1 x 1
#>   terms
#>   <chr>
#> 1 diet