Create traditional dummy variables

step_dummy() creates a specification of a recipe step that will convert nominal data (e.g. character or factors) into one or more numeric binary model terms for the levels of the original data.

Usage

step_dummy(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  one_hot = FALSE,
  preserve = deprecated(),
  naming = dummy_names,
  levels = NULL,
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("dummy")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose variables for this step. See selections() for more details. The selected variables must be factors.

role

For model terms created by this step, what analysis role should they be assigned? By default, the new columns created by this step from the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

one_hot

A logical. For C levels, should C dummy variables be created rather than C-1?

preserve

Use keep_original_cols to specify whether the selected column(s) should be retained (in addition to the new dummy variables).

naming

A function that defines the naming convention for new dummy columns. See Details below.

levels

A list that contains the information needed to create dummy variables for each variable contained in terms. This is NULL until the step is trained by prep().

keep_original_cols

A logical to keep the original variables in the output. Defaults to FALSE.

skip

A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Details

step_dummy() will create a set of binary dummy variables from a factor variable. For example, if an unordered factor column in the data set has levels of "red", "green", "blue", the dummy variable bake will create two additional columns of 0/1 data for two of those three values (and remove the original column). For ordered factors, polynomial contrasts are used to encode the numeric values.

By default, the excluded dummy variable (i.e. the reference cell) will correspond to the first level of the unordered factor being converted.

This recipe step allows for flexible naming of the resulting variables. For an unordered factor named x, with levels "a" and "b", the default naming convention would be to create a new variable called x_b. The naming format can be changed using the naming argument; the function dummy_names() is the default.

To change the type of contrast being used, change the global contrast option via options.

When the factor being converted has a missing value, all of the corresponding dummy variables are also missing. See step_unknown() for a solution.

When data to be processed contains novel levels (i.e., not contained in the training set), a missing value is assigned to the results. See step_other() for an alternative.

If no columns are selected (perhaps due to an earlier step_zv()), bake() will return the data as-is (e.g. with no dummy variables).

Note that, by default, the new dummy variable column names obey the naming rules for columns. If there are levels such as "0", dummy_names() will put a leading "X" in front of the level (since it uses make.names()). This can be changed by passing in a different function to the naming argument for this step.

The package vignette for dummy variables and interactions has more information.

Tidying

When you tidy() this step, a tibble with columns terms (the selectors or original variables selected) and columns (the list of corresponding binary columns) is returned.

See also

Examples

library(modeldata)
data(okc)
okc <- okc[complete.cases(okc), ]

# Original data: diet has 18 levels
length(unique(okc$diet))
#> [1] 18
unique(okc$diet) %>% sort()
#>  [1] "anything"            "halal"               "kosher"             
#>  [4] "mostly anything"     "mostly halal"        "mostly kosher"      
#>  [7] "mostly other"        "mostly vegan"        "mostly vegetarian"  
#> [10] "other"               "strictly anything"   "strictly halal"     
#> [13] "strictly kosher"     "strictly other"      "strictly vegan"     
#> [16] "strictly vegetarian" "vegan"               "vegetarian"         

rec <- recipe(~ diet + age + height, data = okc)

# Default dummy coding: 17 dummy variables
dummies <- rec %>%
  step_dummy(diet) %>%
  prep(training = okc)

dummy_data <- bake(dummies, new_data = NULL)

dummy_data %>%
  select(starts_with("diet")) %>%
  names() # level "anything" is the reference level
#>  [1] "diet_halal"               "diet_kosher"             
#>  [3] "diet_mostly.anything"     "diet_mostly.halal"       
#>  [5] "diet_mostly.kosher"       "diet_mostly.other"       
#>  [7] "diet_mostly.vegan"        "diet_mostly.vegetarian"  
#>  [9] "diet_other"               "diet_strictly.anything"  
#> [11] "diet_strictly.halal"      "diet_strictly.kosher"    
#> [13] "diet_strictly.other"      "diet_strictly.vegan"     
#> [15] "diet_strictly.vegetarian" "diet_vegan"              
#> [17] "diet_vegetarian"         

# Obtain the full set of 18 dummy variables using `one_hot` option
dummies_one_hot <- rec %>%
  step_dummy(diet, one_hot = TRUE) %>%
  prep(training = okc)

dummy_data_one_hot <- bake(dummies_one_hot, new_data = NULL)

dummy_data_one_hot %>%
  select(starts_with("diet")) %>%
  names() # no reference level
#>  [1] "diet_anything"            "diet_halal"              
#>  [3] "diet_kosher"              "diet_mostly.anything"    
#>  [5] "diet_mostly.halal"        "diet_mostly.kosher"      
#>  [7] "diet_mostly.other"        "diet_mostly.vegan"       
#>  [9] "diet_mostly.vegetarian"   "diet_other"              
#> [11] "diet_strictly.anything"   "diet_strictly.halal"     
#> [13] "diet_strictly.kosher"     "diet_strictly.other"     
#> [15] "diet_strictly.vegan"      "diet_strictly.vegetarian"
#> [17] "diet_vegan"               "diet_vegetarian"         

tidy(dummies, number = 1)
#> # A tibble: 17 × 3
#>    terms columns             id         
#>    <chr> <chr>               <chr>      
#>  1 diet  halal               dummy_ypfCg
#>  2 diet  kosher              dummy_ypfCg
#>  3 diet  mostly anything     dummy_ypfCg
#>  4 diet  mostly halal        dummy_ypfCg
#>  5 diet  mostly kosher       dummy_ypfCg
#>  6 diet  mostly other        dummy_ypfCg
#>  7 diet  mostly vegan        dummy_ypfCg
#>  8 diet  mostly vegetarian   dummy_ypfCg
#>  9 diet  other               dummy_ypfCg
#> 10 diet  strictly anything   dummy_ypfCg
#> 11 diet  strictly halal      dummy_ypfCg
#> 12 diet  strictly kosher     dummy_ypfCg
#> 13 diet  strictly other      dummy_ypfCg
#> 14 diet  strictly vegan      dummy_ypfCg
#> 15 diet  strictly vegetarian dummy_ypfCg
#> 16 diet  vegan               dummy_ypfCg
#> 17 diet  vegetarian          dummy_ypfCg
tidy(dummies_one_hot, number = 1)
#> # A tibble: 18 × 3
#>    terms columns             id         
#>    <chr> <chr>               <chr>      
#>  1 diet  anything            dummy_ONQD4
#>  2 diet  halal               dummy_ONQD4
#>  3 diet  kosher              dummy_ONQD4
#>  4 diet  mostly anything     dummy_ONQD4
#>  5 diet  mostly halal        dummy_ONQD4
#>  6 diet  mostly kosher       dummy_ONQD4
#>  7 diet  mostly other        dummy_ONQD4
#>  8 diet  mostly vegan        dummy_ONQD4
#>  9 diet  mostly vegetarian   dummy_ONQD4
#> 10 diet  other               dummy_ONQD4
#> 11 diet  strictly anything   dummy_ONQD4
#> 12 diet  strictly halal      dummy_ONQD4
#> 13 diet  strictly kosher     dummy_ONQD4
#> 14 diet  strictly other      dummy_ONQD4
#> 15 diet  strictly vegan      dummy_ONQD4
#> 16 diet  strictly vegetarian dummy_ONQD4
#> 17 diet  vegan               dummy_ONQD4
#> 18 diet  vegetarian          dummy_ONQD4