Create traditional dummy variables

step_dummy() creates a specification of a recipe step that will convert nominal data (e.g. factors) into one or more numeric binary model terms corresponding to the levels of the original data.

Usage

step_dummy(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  one_hot = FALSE,
  preserve = deprecated(),
  naming = dummy_names,
  levels = NULL,
  sparse = "auto",
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("dummy")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose variables for this step. See selections() for more details. The selected variables must be factors.

role

For model terms created by this step, what analysis role should they be assigned? By default, the new columns created by this step from the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

one_hot

A logical. For C levels, should C dummy variables be created rather than C-1?

preserve

This argument has been deprecated. Please use keep_original_cols instead.

naming

A function that defines the naming convention for new dummy columns. See Details below.

levels

A list that contains the information needed to create dummy variables for each variable contained in terms. This is NULL until the step is trained by prep().

sparse

A single string. Should the columns produced be sparse vectors. Can take the values "yes", "no", and "auto". If sparse = "auto" then workflows can determine the best option. Defaults to "auto".

keep_original_cols

A logical to keep the original variables in the output. Defaults to FALSE.

skip

A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Details

step_dummy() will create a set of binary dummy variables from a factor variable. For example, if an unordered factor column in the data set has levels of "red", "green", "blue", the dummy variable bake will create two additional columns of 0/1 data for two of those three values (and remove the original column). For ordered factors, polynomial contrasts are used to encode the numeric values.

By default, the excluded dummy variable (i.e. the reference cell) will correspond to the first level of the unordered factor being converted. step_relevel() can be used to create a new reference level by setting the ref_level argument.

This recipe step allows for flexible naming of the resulting variables. For an unordered factor named x, with levels "a" and "b", the default naming convention would be to create a new variable called x_b. The naming format can be changed using the naming argument; the function dummy_names() is the default.

To change the type of contrast being used, change the global contrast option via options.

When the factor being converted has a missing value, all of the corresponding dummy variables are also missing. See step_unknown() for a solution.

When data to be processed contains novel levels (i.e., not contained in the training set), a missing value is assigned to the results. See step_other() for an alternative.

If no columns are selected (perhaps due to an earlier step_zv()), bake() will return the data as-is (e.g. with no dummy variables).

Note that, by default, the new dummy variable column names obey the naming rules for columns. If there are levels such as "0", dummy_names() will put a leading "X" in front of the level (since it uses make.names()). This can be changed by passing in a different function to the naming argument for this step.

Also, there are a number of contrast methods that return fractional values. The columns returned by this step are doubles (not integers) when sparse = FALSE. The columns returned when sparse = TRUE are integers.

The package vignette for dummy variables and interactions has more information.

Tidying

When you tidy() this step, a tibble is returned with columns terms, columns , and id:

terms

character, the selectors or variables selected

columns

character, names of resulting columns

id

character, id of this step

Sparse data

This step produces sparse columns if sparse = "yes" is being set. The default value "auto" won't trigger production fo sparse columns if a recipe is prep()ed, but allows for a workflow to toggle to "yes" or "no" depending on whether the model supports sparse_data and if the model is is expected to run faster with the data.

The mechanism for determining how much sparsity is produced isn't perfect, and there will be times when you want to manually overwrite by setting sparse = "yes" or sparse = "no".

Case weights

The underlying operation does not allow for case weights.

See also

Examples

data(Sacramento, package = "modeldata")

# Original data: city has 37 levels
length(unique(Sacramento$city))
#> [1] 37
unique(Sacramento$city) %>% sort()
#>  [1] ANTELOPE        AUBURN          CAMERON_PARK    CARMICHAEL     
#>  [5] CITRUS_HEIGHTS  COOL            DIAMOND_SPRINGS EL_DORADO      
#>  [9] EL_DORADO_HILLS ELK_GROVE       ELVERTA         FAIR_OAKS      
#> [13] FOLSOM          FORESTHILL      GALT            GARDEN_VALLEY  
#> [17] GOLD_RIVER      GRANITE_BAY     GREENWOOD       LINCOLN        
#> [21] LOOMIS          MATHER          MEADOW_VISTA    NORTH_HIGHLANDS
#> [25] ORANGEVALE      PENRYN          PLACERVILLE     POLLOCK_PINES  
#> [29] RANCHO_CORDOVA  RANCHO_MURIETA  RIO_LINDA       ROCKLIN        
#> [33] ROSEVILLE       SACRAMENTO      WALNUT_GROVE    WEST_SACRAMENTO
#> [37] WILTON         
#> 37 Levels: ANTELOPE AUBURN CAMERON_PARK CARMICHAEL ... WILTON

rec <- recipe(~ city + sqft + price, data = Sacramento)

# Default dummy coding: 36 dummy variables
dummies <- rec %>%
  step_dummy(city) %>%
  prep(training = Sacramento)

dummy_data <- bake(dummies, new_data = NULL)

dummy_data %>%
  select(starts_with("city")) %>%
  names() # level "anything" is the reference level
#>  [1] "city_AUBURN"          "city_CAMERON_PARK"   
#>  [3] "city_CARMICHAEL"      "city_CITRUS_HEIGHTS" 
#>  [5] "city_COOL"            "city_DIAMOND_SPRINGS"
#>  [7] "city_EL_DORADO"       "city_EL_DORADO_HILLS"
#>  [9] "city_ELK_GROVE"       "city_ELVERTA"        
#> [11] "city_FAIR_OAKS"       "city_FOLSOM"         
#> [13] "city_FORESTHILL"      "city_GALT"           
#> [15] "city_GARDEN_VALLEY"   "city_GOLD_RIVER"     
#> [17] "city_GRANITE_BAY"     "city_GREENWOOD"      
#> [19] "city_LINCOLN"         "city_LOOMIS"         
#> [21] "city_MATHER"          "city_MEADOW_VISTA"   
#> [23] "city_NORTH_HIGHLANDS" "city_ORANGEVALE"     
#> [25] "city_PENRYN"          "city_PLACERVILLE"    
#> [27] "city_POLLOCK_PINES"   "city_RANCHO_CORDOVA" 
#> [29] "city_RANCHO_MURIETA"  "city_RIO_LINDA"      
#> [31] "city_ROCKLIN"         "city_ROSEVILLE"      
#> [33] "city_SACRAMENTO"      "city_WALNUT_GROVE"   
#> [35] "city_WEST_SACRAMENTO" "city_WILTON"         

# Obtain the full set of 37 dummy variables using `one_hot` option
dummies_one_hot <- rec %>%
  step_dummy(city, one_hot = TRUE) %>%
  prep(training = Sacramento)

dummy_data_one_hot <- bake(dummies_one_hot, new_data = NULL)

dummy_data_one_hot %>%
  select(starts_with("city")) %>%
  names() # no reference level
#>  [1] "city_ANTELOPE"        "city_AUBURN"         
#>  [3] "city_CAMERON_PARK"    "city_CARMICHAEL"     
#>  [5] "city_CITRUS_HEIGHTS"  "city_COOL"           
#>  [7] "city_DIAMOND_SPRINGS" "city_EL_DORADO"      
#>  [9] "city_EL_DORADO_HILLS" "city_ELK_GROVE"      
#> [11] "city_ELVERTA"         "city_FAIR_OAKS"      
#> [13] "city_FOLSOM"          "city_FORESTHILL"     
#> [15] "city_GALT"            "city_GARDEN_VALLEY"  
#> [17] "city_GOLD_RIVER"      "city_GRANITE_BAY"    
#> [19] "city_GREENWOOD"       "city_LINCOLN"        
#> [21] "city_LOOMIS"          "city_MATHER"         
#> [23] "city_MEADOW_VISTA"    "city_NORTH_HIGHLANDS"
#> [25] "city_ORANGEVALE"      "city_PENRYN"         
#> [27] "city_PLACERVILLE"     "city_POLLOCK_PINES"  
#> [29] "city_RANCHO_CORDOVA"  "city_RANCHO_MURIETA" 
#> [31] "city_RIO_LINDA"       "city_ROCKLIN"        
#> [33] "city_ROSEVILLE"       "city_SACRAMENTO"     
#> [35] "city_WALNUT_GROVE"    "city_WEST_SACRAMENTO"
#> [37] "city_WILTON"         

tidy(dummies, number = 1)
#> # A tibble: 36 × 3
#>    terms columns         id         
#>    <chr> <chr>           <chr>      
#>  1 city  AUBURN          dummy_ypfCg
#>  2 city  CAMERON_PARK    dummy_ypfCg
#>  3 city  CARMICHAEL      dummy_ypfCg
#>  4 city  CITRUS_HEIGHTS  dummy_ypfCg
#>  5 city  COOL            dummy_ypfCg
#>  6 city  DIAMOND_SPRINGS dummy_ypfCg
#>  7 city  EL_DORADO       dummy_ypfCg
#>  8 city  EL_DORADO_HILLS dummy_ypfCg
#>  9 city  ELK_GROVE       dummy_ypfCg
#> 10 city  ELVERTA         dummy_ypfCg
#> # ℹ 26 more rows
tidy(dummies_one_hot, number = 1)
#> # A tibble: 37 × 3
#>    terms columns         id         
#>    <chr> <chr>           <chr>      
#>  1 city  ANTELOPE        dummy_ONQD4
#>  2 city  AUBURN          dummy_ONQD4
#>  3 city  CAMERON_PARK    dummy_ONQD4
#>  4 city  CARMICHAEL      dummy_ONQD4
#>  5 city  CITRUS_HEIGHTS  dummy_ONQD4
#>  6 city  COOL            dummy_ONQD4
#>  7 city  DIAMOND_SPRINGS dummy_ONQD4
#>  8 city  EL_DORADO       dummy_ONQD4
#>  9 city  EL_DORADO_HILLS dummy_ONQD4
#> 10 city  ELK_GROVE       dummy_ONQD4
#> # ℹ 27 more rows