Distance between two locations

step_geodist creates a a specification of a recipe step that will calculate the distance between points on a map to a reference location.

step_geodist(recipe, lat = NULL, lon = NULL, role = "predictor",
  trained = FALSE, ref_lat = NULL, ref_lon = NULL, log = FALSE,
  name = "geo_dist", columns = NULL, skip = FALSE,
  id = rand_id("geodist"))

# S3 method for step_geodist
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

lon, lat
Selector functions to choose which variables are affected by the step. See selections() for more details.

role
or model term created by this step, what analysis role should be assigned?. By default, the function assumes that resulting distance will be used as a predictor in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

ref_lon, ref_lat
Single numeric values for the location of the reference point.

log
A logical: should the distance be transformed by the natural log function?

name
A single character value to use for the new predictor column. If a column exists with this name, an error is issued.

columns
A character string of variable names that will be populated (eventually) by the terms argument.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_geodist object.

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details. For the tidy method, these are not currently used.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns echoing the values of lat, lon, ref_lat, ref_lon, name, and id.

Details

step_geodist will create a

Examples

data(Smithsonian)

# How close are the museums to Union Station?
near_station <- recipe( ~ ., data = Smithsonian) %>%
  update_role(name, new_role = "location") %>%
  step_geodist(lat = latitude, lon = longitude, log = FALSE,
               ref_lat = 38.8986312, ref_lon = -77.0062457) %>%
  prep(training = Smithsonian, retain = TRUE)

juice(near_station) %>%
  arrange(geo_dist)

#> [38;5;246m# A tibble: 20 x 4[39m
#>    name                                             latitude longitude geo_dist
#>    [3m[38;5;246m<fct>[39m[23m                                               [3m[38;5;246m<dbl>[39m[23m     [3m[38;5;246m<dbl>[39m[23m    [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m 1[39m National Postal Museum                               38.9     -[31m77[39m[31m.[39m[31m0[39m  0.004[4m2[24m[4m1[24m
#> [38;5;250m 2[39m Renwick Gallery                                      38.9     -[31m77[39m[31m.[39m[31m0[39m  0.010[4m8[24m 
#> [38;5;250m 3[39m National Museum of the American Indian               38.9     -[31m77[39m[31m.[39m[31m0[39m  0.016[4m1[24m 
#> [38;5;250m 4[39m Smithsonian American Art Museum                      38.9     -[31m77[39m[31m.[39m[31m0[39m  0.018[4m9[24m 
#> [38;5;250m 5[39m National Air and Space Museum                        38.9     -[31m77[39m[31m.[39m[31m0[39m  0.019[4m0[24m 
#> [38;5;250m 6[39m National Portrait Gallery                            38.9     -[31m77[39m[31m.[39m[31m0[39m  0.019[4m0[24m 
#> [38;5;250m 7[39m Hirshhorn Museum and Sculpture Garden                38.9     -[31m77[39m[31m.[39m[31m0[39m  0.021[4m6[24m 
#> [38;5;250m 8[39m Arthur M. Sackler Gallery                            38.9     -[31m77[39m[31m.[39m[31m0[39m  0.022[4m8[24m 
#> [38;5;250m 9[39m Arts and Industries Building                         38.9     -[31m77[39m[31m.[39m[31m0[39m  0.023[4m0[24m 
#> [38;5;250m10[39m National Museum of Natural History                   38.9     -[31m77[39m[31m.[39m[31m0[39m  0.023[4m2[24m 
#> [38;5;250m11[39m National Museum of African Art                       38.9     -[31m77[39m[31m.[39m[31m0[39m  0.023[4m9[24m 
#> [38;5;250m12[39m Smithsonian Institution Building                     38.9     -[31m77[39m[31m.[39m[31m0[39m  0.024[4m1[24m 
#> [38;5;250m13[39m Freer Gallery of Art                                 38.9     -[31m77[39m[31m.[39m[31m0[39m  0.024[4m7[24m 
#> [38;5;250m14[39m National Museum of American History                  38.9     -[31m77[39m[31m.[39m[31m0[39m  0.027[4m0[24m 
#> [38;5;250m15[39m National Museum of African American History and…     38.9     -[31m77[39m[31m.[39m[31m0[39m  0.029[4m5[24m 
#> [38;5;250m16[39m Anacostia Community Museum                           38.9     -[31m77[39m[31m.[39m[31m0[39m  0.051[4m4[24m 
#> [38;5;250m17[39m National Zoological Park                             38.9     -[31m77[39m[31m.[39m[31m1[39m  0.055[4m2[24m 
#> [38;5;250m18[39m Steven F. Udvar-Hazy Center                          38.9     -[31m77[39m[31m.[39m[31m4[39m  0.440  
#> [38;5;250m19[39m George Gustav Heye Center                            40.7     -[31m74[39m[31m.[39m[31m0[39m  3.49   
#> [38;5;250m20[39m Cooper Hewitt, Smithsonian Design Museum             40.8     -[31m74[39m[31m.[39m[31m0[39m  3.58   

tidy(near_station, number = 1)

#> [38;5;246m# A tibble: 1 x 6[39m
#>   latitude longitude ref_latitude ref_longitude name     id           
#>   [3m[38;5;246m<chr>[39m[23m    [3m[38;5;246m<chr>[39m[23m            [3m[38;5;246m<dbl>[39m[23m         [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<chr>[39m[23m    [3m[38;5;246m<chr>[39m[23m        
#> [38;5;250m1[39m latitude longitude         38.9         -[31m77[39m[31m.[39m[31m0[39m geo_dist geodist_GMlQC