Holiday feature generator

step_holiday() creates a specification of a recipe step that will convert date data into one or more binary indicator variables for common holidays.

Usage

step_holiday(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  holidays = c("LaborDay", "NewYearsDay", "ChristmasDay"),
  columns = NULL,
  sparse = "auto",
  keep_original_cols = TRUE,
  skip = FALSE,
  id = rand_id("holiday")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose variables for this step. The selected variables should have class Date or POSIXct. See selections() for more details.

role

For model terms created by this step, what analysis role should they be assigned? By default, the new columns created by this step from the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

holidays

A character string that includes at least one holiday supported by the timeDate package. See timeDate::listHolidays() for a complete list.

columns

A character string of the selected variable names. This field is a placeholder and will be populated once prep() is used.

sparse

A single string. Should the columns produced be sparse vectors. Can take the values "yes", "no", and "auto". If sparse = "auto" then workflows can determine the best option. Sparsity is only supported for "contr.treatment" contrasts. Defaults to "auto".

keep_original_cols

A logical to keep the original variables in the output. Defaults to TRUE.

skip

A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Details

Unlike some other steps, step_holiday does not remove the original date variables by default. Set keep_original_cols to FALSE to remove them.

Tidying

When you tidy() this step, a tibble is returned with columns terms, holiday , and id:

terms

character, the selectors or variables selected

holiday

character, name of holidays

id

character, id of this step

Case weights

The underlying operation does not allow for case weights.

See also

Examples

library(lubridate)

examples <- data.frame(someday = ymd("2000-12-20") + days(0:40))
holiday_rec <- recipe(~someday, examples) %>%
  step_holiday(all_predictors())

holiday_rec <- prep(holiday_rec, training = examples)
holiday_values <- bake(holiday_rec, new_data = examples)
holiday_values
#> # A tibble: 41 × 4
#>    someday    someday_LaborDay someday_NewYearsDay someday_ChristmasDay
#>    <date>                <int>               <int>                <int>
#>  1 2000-12-20                0                   0                    0
#>  2 2000-12-21                0                   0                    0
#>  3 2000-12-22                0                   0                    0
#>  4 2000-12-23                0                   0                    0
#>  5 2000-12-24                0                   0                    0
#>  6 2000-12-25                0                   0                    1
#>  7 2000-12-26                0                   0                    0
#>  8 2000-12-27                0                   0                    0
#>  9 2000-12-28                0                   0                    0
#> 10 2000-12-29                0                   0                    0
#> # ℹ 31 more rows