Convert values to predefined integers

step_integer creates a a specification of a recipe step that will convert new data into a set of integers based on the original data values.

step_integer(recipe, ..., role = "predictor", trained = FALSE,
  strict = FALSE, zero_based = FALSE, key = NULL, skip = FALSE,
  id = rand_id("integer"))

# S3 method for step_integer
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables will be used to create the integer variables. See selections() for more details. For the tidy method, these are not currently used.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

strict
A logical for whether the values should be returned as integers (as opposed to double).

zero_based
A logical for whether the integers should start at zero and new values be appended as the largest integer.

key
A list that contains the information needed to create integer variables for each variable contained in terms. This is NULL until the step is trained by prep.recipe().

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_integer object.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms (the selectors or variables selected) and value is a list column with the conversion key.

Details

step_integer will determine the unique values of each variable from the training set (excluding missing values), order them, and then assign integers to each value. When baked, each data point is translated to its corresponding integer or a value of zero for yet unseen data (although see the zero_based argument above). Missing values propagate.

Factor inputs are ordered by their levels. All others are ordered by sort.

Despite the name, the new values are returned as numeric unless strict = TRUE, which will coerce the results to integers.

See also

Examples

data(okc)

okc$location <- factor(okc$location)

okc_tr <- okc[1:100, ]
okc_tr$age[1] <- NA

okc_te <- okc[101:105, ]
okc_te$age[1] <- NA
okc_te$diet[1] <- "fast food"
okc_te$diet[2] <- NA

rec <- recipe(Class ~ ., data = okc_tr) %>%
  step_integer(all_predictors()) %>%
  prep(training = okc_tr, retain = TRUE)

bake(rec, okc_te, all_predictors())

#> [38;5;246m# A tibble: 5 x 5[39m
#>     age  diet height location  date
#>   [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<dbl>[39m[23m  [3m[38;5;246m<dbl>[39m[23m    [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m1[39m    [31mNA[39m     0     12      102    37
#> [38;5;250m2[39m    11    [31mNA[39m      9       25     0
#> [38;5;250m3[39m     0     2     10       31    36
#> [38;5;250m4[39m    20     3     11      102     0
#> [38;5;250m5[39m    12     1      4      102    33

tidy(rec, number = 1)

#> [38;5;246m# A tibble: 5 x 3[39m
#>   terms    value              id           
#>   [3m[38;5;246m<chr>[39m[23m    [3m[38;5;246m<list>[39m[23m             [3m[38;5;246m<chr>[39m[23m        
#> [38;5;250m1[39m age      [38;5;246m<tibble [23 × 2]>[39m  integer_ePkup
#> [38;5;250m2[39m diet     [38;5;246m<tibble [8 × 2]>[39m   integer_ePkup
#> [38;5;250m3[39m height   [38;5;246m<tibble [18 × 2]>[39m  integer_ePkup
#> [38;5;250m4[39m location [38;5;246m<tibble [135 × 2]>[39m integer_ePkup
#> [38;5;250m5[39m date     [38;5;246m<tibble [38 × 2]>[39m  integer_ePkup