Convert values to predefined integers

step_integer creates a specification of a recipe step that will convert new data into a set of integers based on the original data values.

Usage

step_integer(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  strict = FALSE,
  zero_based = FALSE,
  key = NULL,
  skip = FALSE,
  id = rand_id("integer")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose variables for this step. See selections() for more details.

role

For model terms created by this step, what analysis role should they be assigned? By default, the new columns created by this step from the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

strict

A logical for whether the values should be returned as integers (as opposed to double).

zero_based

A logical for whether the integers should start at zero and new values be appended as the largest integer.

key

A list that contains the information needed to create integer variables for each variable contained in terms. This is NULL until the step is trained by prep().

skip

A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Details

step_integer will determine the unique values of each variable from the training set (excluding missing values), order them, and then assign integers to each value. When baked, each data point is translated to its corresponding integer or a value of zero for yet unseen data (although see the zero_based argument above). Missing values propagate.

Factor inputs are ordered by their levels. All others are ordered by sort.

Despite the name, the new values are returned as numeric unless strict = TRUE, which will coerce the results to integers.

Tidying

When you tidy() this step, a tibble with columns terms (the selectors or variables selected) and value (a list column with the conversion key) is returned.

Case weights

The underlying operation does not allow for case weights.

See also

Examples

data(Sacramento, package = "modeldata")

sacr_tr <- Sacramento[1:100, ]
sacr_tr$sqft[1] <- NA

sacr_te <- Sacramento[101:105, ]
sacr_te$sqft[1] <- NA
sacr_te$city[1] <- "whoville"
#> Warning: invalid factor level, NA generated
sacr_te$city[2] <- NA

rec <- recipe(type ~ ., data = sacr_tr) %>%
  step_integer(all_predictors()) %>%
  prep(training = sacr_tr)

bake(rec, sacr_te, all_predictors())
#> # A tibble: 5 × 8
#>    city   zip  beds baths  sqft price latitude longitude
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>    <dbl>     <dbl>
#> 1    NA    35     4     2    NA     0        0         0
#> 2    NA    62     3     2     0     0        0         0
#> 3    28    34     3     2    56     0        0         0
#> 4    34    51     3     1     0     0        0         0
#> 5    34    58     4     3     0     0        0         0
tidy(rec, number = 1)
#> # A tibble: 8 × 3
#>   terms     value             id           
#>   <chr>     <list>            <chr>        
#> 1 city      <tibble [37 × 2]> integer_3IckW
#> 2 zip       <tibble [68 × 2]> integer_3IckW
#> 3 beds      <tibble [5 × 2]>  integer_3IckW
#> 4 baths     <tibble [4 × 2]>  integer_3IckW
#> 5 sqft      <tibble [94 × 2]> integer_3IckW
#> 6 price     <tibble [95 × 2]> integer_3IckW
#> 7 latitude  <tibble [99 × 2]> integer_3IckW
#> 8 longitude <tibble [99 × 2]> integer_3IckW