Create Interaction Variables

step_interact creates a specification of a recipe step that will create new columns that are interaction terms between two or more variables.

step_interact(recipe, terms, role = "predictor", trained = FALSE,
  objects = NULL, sep = "_x_")

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

terms

A traditional R formula that contains interaction terms.

role

For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created from the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

objects

A list of terms objects for each individual interation.

sep

A character value used to delinate variables in an interaction (e.g. var1_x_var2 instead of the more traditional var1:var2).

Value

step_interact returns an object of class step_interact.

Details

step_interact can create interactions between variables. It is primarily intended for numeric data; categorical variables should probably be converted to dummy variables using step_dummy prior to being used for interactions.

Unlike other step functions, the terms argument should be a traditional R model formula but should contain no inline functions (e.g. log). For example, for predictors A, B, and C, a formula such as ~A:B:C can be used to make a three way interaction between the variables. If the formula contains terms other than interactions (e.g. (A+B+C)^3) only the interaction terms are retained for the design matrix.

The separator between the variables defaults to "_x_" so that the three way interaction shown previously would generate a column named A_x_B_x_C. This can be changed using the sep argument.

Examples

data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)

library(magrittr)
int_mod_1 <- rec %>%
  step_interact(terms = ~ carbon:hydrogen)

int_mod_2 <- int_mod_1 %>%
  step_interact(terms = ~ (oxygen + nitrogen + sulfur)^3)

int_mod_1 <- learn(int_mod_1, training = biomass_tr)

#> step 1 interact training 

int_mod_2 <- learn(int_mod_2, training = biomass_tr)

#> step 1 interact training 
#> step 2 interact training 

dat_1 <- process(int_mod_1, biomass_te)
dat_2 <- process(int_mod_2, biomass_te)

names(dat_1)

#> [1] "carbon"            "hydrogen"          "oxygen"           
#> [4] "nitrogen"          "sulfur"            "HHV"              
#> [7] "carbon_x_hydrogen"

names(dat_2)

#>  [1] "carbon"                     "hydrogen"                  
#>  [3] "oxygen"                     "nitrogen"                  
#>  [5] "sulfur"                     "HHV"                       
#>  [7] "carbon_x_hydrogen"          "oxygen_x_nitrogen"         
#>  [9] "oxygen_x_sulfur"            "nitrogen_x_sulfur"         
#> [11] "oxygen_x_nitrogen_x_sulfur"