Create Interaction Variables

step_interact creates a specification of a recipe step that will create new columns that are interaction terms between two or more variables.

step_interact(recipe, terms, role = "predictor", trained = FALSE,
  objects = NULL, sep = "_x_", skip = FALSE)

# S3 method for step_interact
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

terms
A traditional R formula that contains interaction terms. This can include . and selectors.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created from the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

objects
A list of terms objects for each individual interaction.

sep
A character value used to delineate variables in an interaction (e.g. var1_x_var2 instead of the more traditional var1:var2).

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

x
A step_interact object

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details. For the tidy method, these are not currently used.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms which is the interaction effects.

Details

step_interact can create interactions between variables. It is primarily intended for numeric data; categorical variables should probably be converted to dummy variables using step_dummy() prior to being used for interactions.

Unlike other step functions, the terms argument should be a traditional R model formula but should contain no inline functions (e.g. log). For example, for predictors A, B, and C, a formula such as ~A:B:C can be used to make a three way interaction between the variables. If the formula contains terms other than interactions (e.g. (A+B+C)^3) only the interaction terms are retained for the design matrix.

The separator between the variables defaults to "_x_" so that the three way interaction shown previously would generate a column named A_x_B_x_C. This can be changed using the sep argument.

When dummy variables are created and are used in interactions, selectors can help specify the interactions succinctly. For example, suppose a factor column X gets converted to dummy variables x_2, x_3, ..., x_6 using step_dummy(). If you wanted an interaction with numeric column z, you could create a set of specific interaction effects (e.g. x_2:z + x_3:z and so on) or you could use starts_with("z_"):z. When prep() evaluates this step, starts_with("z_") resolves to (x_2 + x_3 + x_4 + x_5 + x6) so that the formula is now (x_2 + x_3 + x_4 + x_5 + x6):z and all two-way interactions are created.

Examples

data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)

int_mod_1 <- rec %>%
  step_interact(terms = ~ carbon:hydrogen)

int_mod_2 <- rec %>%
  step_interact(terms = ~ (matches("gen$") + sulfur)^2)

int_mod_1 <- prep(int_mod_1, training = biomass_tr)
int_mod_2 <- prep(int_mod_2, training = biomass_tr)

dat_1 <- bake(int_mod_1, biomass_te)
dat_2 <- bake(int_mod_2, biomass_te)

names(dat_1)

#> [1] "carbon"            "hydrogen"          "oxygen"           
#> [4] "nitrogen"          "sulfur"            "HHV"              
#> [7] "carbon_x_hydrogen"

names(dat_2)

#>  [1] "carbon"              "hydrogen"            "oxygen"             
#>  [4] "nitrogen"            "sulfur"              "HHV"                
#>  [7] "hydrogen_x_oxygen"   "hydrogen_x_nitrogen" "hydrogen_x_sulfur"  
#> [10] "oxygen_x_nitrogen"   "oxygen_x_sulfur"     "nitrogen_x_sulfur"  

tidy(int_mod_1, number = 1)

#> # A tibble: 1 x 1
#>   terms          
#>   <chr>          
#> 1 carbon:hydrogen

tidy(int_mod_2, number = 1)

#> # A tibble: 6 x 1
#>   terms            
#>   <chr>            
#> 1 hydrogen:oxygen  
#> 2 hydrogen:nitrogen
#> 3 hydrogen:sulfur  
#> 4 oxygen:nitrogen  
#> 5 oxygen:sulfur    
#> 6 nitrogen:sulfur