Add intercept (or constant) column

step_intercept creates a specification of a recipe step that will add an intercept or constant term in the first column of a data matrix. step_intercept has defaults to predictor role so that it is by default called in the bake step. Be careful to avoid unintentional transformations when calling steps with all_predictors.

step_intercept(recipe, ..., role = "predictor", trained = FALSE,
  name = "intercept", value = 1)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
Argument ignored; included for consistency with other step specification functions.

role
Defaults to "predictor"

trained
A logical to indicate if the quantities for preprocessing have been estimated. Again included for consistency.

name
Character name for new added column

value
A numeric constant to fill the intercept column. Defaults to 1.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

See also

recipe() prep.recipe() bake.recipe()

Examples

data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)
rec_trans <- recipe(HHV ~ ., data = biomass_tr[, -(1:2)]) %>%
  step_intercept(value = 2)

rec_obj <- prep(rec_trans, training = biomass_tr)

with_intercept <- bake(rec_obj, biomass_te)
with_intercept

#> # A tibble: 80 x 7
#>    intercept carbon hydrogen oxygen nitrogen sulfur    HHV
#>        <dbl>  <dbl>    <dbl>  <dbl>    <dbl>  <dbl>  <dbl>
#>  1         2  46.35     5.67  47.20     0.30   0.22 18.275
#>  2         2  43.25     5.50  48.06     2.85   0.34 17.560
#>  3         2  42.70     5.50  49.10     2.40   0.30 17.173
#>  4         2  46.40     6.10  37.30     1.80   0.50 18.851
#>  5         2  48.76     6.32  42.77     0.20   0.00 20.547
#>  6         2  44.30     5.50  41.70     0.70   0.20 18.467
#>  7         2  38.94     5.23  54.13     1.19   0.51 15.095
#>  8         2  42.10     4.66  33.80     0.95   0.20 16.240
#>  9         2  29.20     4.40  31.10     0.14   4.90 11.147
#> 10         2  27.80     3.77  23.69     4.63   1.05 10.750
#> # ... with 70 more rows