Add intercept (or constant) column

`step_intercept` creates a *specification* of a recipe step that will add an intercept or constant term in the first column of a data matrix. `step_intercept` has defaults to *predictor* role so that it is by default called in the bake step. Be careful to avoid unintentional transformations when calling steps with `all_predictors`.

step_intercept(recipe, ..., role = "predictor", trained = FALSE,
  name = "intercept", value = 1, skip = FALSE)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
Argument ignored; included for consistency with other step specification functions.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created from the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated. Again included for consistency.

name
Character name for new added column

value
A numeric constant to fill the intercept column. Defaults to 1.

skip
A logical. Should the step be skipped when the recipe is baked by [bake.recipe()]? While all operations are baked when [prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any).

See also

Examples

data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)
rec_trans <- recipe(HHV ~ ., data = biomass_tr[, -(1:2)]) %>%
  step_intercept(value = 2)

rec_obj <- prep(rec_trans, training = biomass_tr)

with_intercept <- bake(rec_obj, biomass_te)
with_intercept

#> # A tibble: 80 x 7
#>    intercept carbon hydrogen oxygen nitrogen sulfur   HHV
#>        <dbl>  <dbl>    <dbl>  <dbl>    <dbl>  <dbl> <dbl>
#>  1        2.   46.4     5.67   47.2    0.300  0.220  18.3
#>  2        2.   43.2     5.50   48.1    2.85   0.340  17.6
#>  3        2.   42.7     5.50   49.1    2.40   0.300  17.2
#>  4        2.   46.4     6.10   37.3    1.80   0.500  18.9
#>  5        2.   48.8     6.32   42.8    0.200  0.     20.5
#>  6        2.   44.3     5.50   41.7    0.700  0.200  18.5
#>  7        2.   38.9     5.23   54.1    1.19   0.510  15.1
#>  8        2.   42.1     4.66   33.8    0.950  0.200  16.2
#>  9        2.   29.2     4.40   31.1    0.140  4.90   11.1
#> 10        2.   27.8     3.77   23.7    4.63   1.05   10.8
#> # ... with 70 more rows