Create a lagged predictor

step_lag creates a specification of a recipe step that will add new columns of lagged data. Lagged data will by default include NA values where the lag was induced. These can be removed with step_naomit(), or you may specify an alternative filler value with the default argument.

step_lag(recipe, ..., role = "predictor", trained = FALSE, lag = 1,
  prefix = "lag_", default = NA, skip = FALSE)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details.

role
Defaults to "predictor"

trained
A logical to indicate if the quantities for preprocessing have been estimated.

lag
A vector of positive integers. Each specified column will be lagged for each value in the vector.

prefix
A prefix for generated column names, default to "lag_".

default
Passed to dplyr::lag, determines what fills empty rows left by lagging (defaults to NA).

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

Details

The step assumes that the data are already in the proper sequential order for lagging.

See also

Examples

n <- 10
start <- as.Date('1999/01/01')
end <- as.Date('1999/01/10')

df <- data.frame(x = runif(n),
                 index = 1:n,
                 day = seq(start, end, by = "day"))

recipe(~ ., data = df) %>%
  step_lag(index, day, lag = 2:3) %>%
  prep(df) %>%
  bake(df)

#> # A tibble: 10 x 7
#>         x index day        lag_2_index lag_2_day  lag_3_index lag_3_day 
#>     <dbl> <int> <date>           <int> <date>           <int> <date>    
#>  1 0.700      1 1999-01-01          NA NA                  NA NA        
#>  2 0.636      2 1999-01-02          NA NA                  NA NA        
#>  3 0.950      3 1999-01-03           1 1999-01-01          NA NA        
#>  4 0.0814     4 1999-01-04           2 1999-01-02           1 1999-01-01
#>  5 0.839      5 1999-01-05           3 1999-01-03           2 1999-01-02
#>  6 0.193      6 1999-01-06           4 1999-01-04           3 1999-01-03
#>  7 0.130      7 1999-01-07           5 1999-01-05           4 1999-01-04
#>  8 0.574      8 1999-01-08           6 1999-01-06           5 1999-01-05
#>  9 0.923      9 1999-01-09           7 1999-01-07           6 1999-01-06
#> 10 0.0992    10 1999-01-10           8 1999-01-08           7 1999-01-07