Create a lagged predictor

step_lag() creates a specification of a recipe step that will add new columns of lagged data. Lagged data will by default include NA values where the lag was induced. These can be removed with step_naomit(), or you may specify an alternative filler value with the default argument.

Usage

step_lag(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  lag = 1,
  prefix = "lag_",
  default = NA,
  columns = NULL,
  keep_original_cols = TRUE,
  skip = FALSE,
  id = rand_id("lag")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose variables for this step. See selections() for more details.

role

For model terms created by this step, what analysis role should they be assigned? By default, the new columns created by this step from the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

lag

A vector of positive integers. Each specified column will be lagged for each value in the vector.

prefix

A prefix for generated column names, default to "lag_".

default

Passed to dplyr::lag, determines what fills empty rows left by lagging (defaults to NA).

columns

A character string of the selected variable names. This field is a placeholder and will be populated once prep() is used.

keep_original_cols

A logical to keep the original variables in the output. Defaults to TRUE.

skip

A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Details

The step assumes that the data are already in the proper sequential order for lagging.

Tidying

When you tidy() this step, a tibble is returned with columns terms and id:

terms

character, the selectors or variables selected

id

character, id of this step

Sparse data

This step can be applied to sparse_data such that it is preserved. Nothing needs to be done for this to happen as it is done automatically.

Case weights

The underlying operation does not allow for case weights.

See also

Examples

n <- 10
start <- as.Date("1999/01/01")
end <- as.Date("1999/01/10")

df <- data.frame(
  x = runif(n),
  index = 1:n,
  day = seq(start, end, by = "day")
)

recipe(~., data = df) %>%
  step_lag(index, day, lag = 2:3) %>%
  prep(df) %>%
  bake(df)
#> # A tibble: 10 × 7
#>         x index day        lag_2_index lag_3_index lag_2_day 
#>     <dbl> <int> <date>           <int>       <int> <date>    
#>  1 0.189      1 1999-01-01          NA          NA NA        
#>  2 0.0579     2 1999-01-02          NA          NA NA        
#>  3 0.0469     3 1999-01-03           1          NA 1999-01-01
#>  4 0.356      4 1999-01-04           2           1 1999-01-02
#>  5 0.563      5 1999-01-05           3           2 1999-01-03
#>  6 0.757      6 1999-01-06           4           3 1999-01-04
#>  7 0.687      7 1999-01-07           5           4 1999-01-05
#>  8 0.966      8 1999-01-08           6           5 1999-01-06
#>  9 0.977      9 1999-01-09           7           6 1999-01-07
#> 10 0.0903    10 1999-01-10           8           7 1999-01-08
#> # ℹ 1 more variable: lag_3_day <date>