Impute Numeric Data Using the Median

step_medianimpute creates a specification of a recipe step that will substitute missing values of numeric variables by the training set median of those variables.

step_medianimpute(recipe, ..., role = NA, trained = FALSE,
  medians = NULL, skip = FALSE, id = rand_id("medianimpute"))

# S3 method for step_medianimpute
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details. For the tidy method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

medians
A named numeric vector of medians. This is NULL until computed by prep.recipe().

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_medianimpute object.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms (the selectors or variables selected) and model (the median value).

Details

step_medianimpute estimates the variable medians from the data used in the training argument of prep.recipe. bake.recipe then applies the new values to new data sets using these medians

Examples

data("credit_data")

## missing data per column
vapply(credit_data, function(x) mean(is.na(x)), c(num = 0))

#>       Status    Seniority         Home         Time          Age      Marital 
#> 0.0000000000 0.0000000000 0.0013471037 0.0000000000 0.0000000000 0.0002245173 
#>      Records          Job     Expenses       Income       Assets         Debt 
#> 0.0000000000 0.0004490346 0.0000000000 0.0855410867 0.0105523125 0.0040413112 
#>       Amount        Price 
#> 0.0000000000 0.0000000000 

set.seed(342)
in_training <- sample(1:nrow(credit_data), 2000)

credit_tr <- credit_data[ in_training, ]
credit_te <- credit_data[-in_training, ]
missing_examples <- c(14, 394, 565)

rec <- recipe(Price ~ ., data = credit_tr)

impute_rec <- rec %>%
  step_medianimpute(Income, Assets, Debt)

imp_models <- prep(impute_rec, training = credit_tr)

imputed_te <- bake(imp_models, new_data = credit_te, everything())

credit_te[missing_examples,]

#>      Status Seniority Home Time Age Marital Records   Job Expenses Income
#> 30      bad         0 <NA>   48  37  single      no  <NA>       35     NA
#> 736     bad         1 priv   48  36  single     yes fixed       35    173
#> 1057    bad        24 priv   60  46 married      no fixed       60    165
#>      Assets Debt Amount Price
#> 30       NA   NA   1500  1850
#> 736    7700    0   1000  2008
#> 1057   3500 1300   1400  1530

imputed_te[missing_examples, names(credit_te)]

#> [38;5;246m# A tibble: 3 x 14[39m
#>   Status Seniority Home   Time   Age Marital Records Job   Expenses Income
#>   [3m[38;5;246m<fct>[39m[23m      [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<fct>[39m[23m [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<int>[39m[23m [3m[38;5;246m<fct>[39m[23m   [3m[38;5;246m<fct>[39m[23m   [3m[38;5;246m<fct>[39m[23m    [3m[38;5;246m<int>[39m[23m  [3m[38;5;246m<dbl>[39m[23m
#> [38;5;250m1[39m bad            0 [31mNA[39m       48    37 single  no      [31mNA[39m          35    125
#> [38;5;250m2[39m bad            1 priv     48    36 single  yes     fixed       35    173
#> [38;5;250m3[39m bad           24 priv     60    46 married no      fixed       60    165
#> [38;5;246m# ... with 4 more variables: Assets [3m[38;5;246m<dbl>[38;5;246m[23m, Debt [3m[38;5;246m<dbl>[38;5;246m[23m, Amount [3m[38;5;246m<int>[38;5;246m[23m,
#> #   Price [3m[38;5;246m<int>[38;5;246m[23m[39m

tidy(impute_rec, number = 1)

#> [38;5;246m# A tibble: 3 x 3[39m
#>   terms  model id                
#>   [3m[38;5;246m<chr>[39m[23m  [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<chr>[39m[23m             
#> [38;5;250m1[39m Income    [31mNA[39m medianimpute_iNKND
#> [38;5;250m2[39m Assets    [31mNA[39m medianimpute_iNKND
#> [38;5;250m3[39m Debt      [31mNA[39m medianimpute_iNKND

tidy(imp_models, number = 1)

#> [38;5;246m# A tibble: 3 x 3[39m
#>   terms  model id                
#>   [3m[38;5;246m<chr>[39m[23m  [3m[38;5;246m<dbl>[39m[23m [3m[38;5;246m<chr>[39m[23m             
#> [38;5;250m1[39m Income   125 medianimpute_iNKND
#> [38;5;250m2[39m Assets  [4m3[24m500 medianimpute_iNKND
#> [38;5;250m3[39m Debt       0 medianimpute_iNKND