Impute Numeric Data Using the Median

step_medianimpute creates a specification of a recipe step that will substitute missing values of numeric variables by the training set median of those variables.

step_medianimpute(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  medians = NULL,
  skip = FALSE,
  id = rand_id("medianimpute")
)

# S3 method for step_medianimpute
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details. For the tidy method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

medians
A named numeric vector of medians. This is NULL until computed by prep.recipe(). Note that, if the original data are integers, the median will be converted to an integer to maintain the same data type.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_medianimpute object.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms (the selectors or variables selected) and model (the median value).

Details

step_medianimpute estimates the variable medians from the data used in the training argument of prep.recipe. bake.recipe then applies the new values to new data sets using these medians.

Examples

library(modeldata)
data("credit_data")

## missing data per column
vapply(credit_data, function(x) mean(is.na(x)), c(num = 0))

#>       Status    Seniority         Home         Time          Age      Marital 
#> 0.0000000000 0.0000000000 0.0013471037 0.0000000000 0.0000000000 0.0002245173 
#>      Records          Job     Expenses       Income       Assets         Debt 
#> 0.0000000000 0.0004490346 0.0000000000 0.0855410867 0.0105523125 0.0040413112 
#>       Amount        Price 
#> 0.0000000000 0.0000000000 

set.seed(342)
in_training <- sample(1:nrow(credit_data), 2000)

credit_tr <- credit_data[ in_training, ]
credit_te <- credit_data[-in_training, ]
missing_examples <- c(14, 394, 565)

rec <- recipe(Price ~ ., data = credit_tr)

impute_rec <- rec %>%
  step_medianimpute(Income, Assets, Debt)

imp_models <- prep(impute_rec, training = credit_tr)

imputed_te <- bake(imp_models, new_data = credit_te, everything())

credit_te[missing_examples,]

#>      Status Seniority  Home Time Age Marital Records     Job Expenses Income
#> 28     good        15 owner   36  43 married      no   fixed       75    251
#> 688    good         2  rent   60  32 married      no partime       87    115
#> 1002   good        21  rent   60  39 married      no   fixed      124    191
#>      Assets Debt Amount Price
#> 28     4000    0   1800  2557
#> 688    2000    0   1250  1517
#> 1002   2000    0   2000  2536

imputed_te[missing_examples, names(credit_te)]

#> # A tibble: 3 x 14
#>   Status Seniority Home   Time   Age Marital Records Job   Expenses Income
#>   <fct>      <int> <fct> <int> <int> <fct>   <fct>   <fct>    <int>  <int>
#> 1 good          15 owner    36    43 married no      fixed       75    251
#> 2 good           2 rent     60    32 married no      part…       87    115
#> 3 good          21 rent     60    39 married no      fixed      124    191
#> # … with 4 more variables: Assets <int>, Debt <int>, Amount <int>, Price <int>

tidy(impute_rec, number = 1)

#> # A tibble: 3 x 3
#>   terms  model id                
#>   <chr>  <dbl> <chr>             
#> 1 Income    NA medianimpute_Hlm2y
#> 2 Assets    NA medianimpute_Hlm2y
#> 3 Debt      NA medianimpute_Hlm2y

tidy(imp_models, number = 1)

#> # A tibble: 3 x 3
#>   terms  model id                
#>   <chr>  <int> <chr>             
#> 1 Income   125 medianimpute_Hlm2y
#> 2 Assets  3000 medianimpute_Hlm2y
#> 3 Debt       0 medianimpute_Hlm2y