Impute Nominal Data Using the Most Common Value

`step_modeimpute` creates a *specification* of a recipe step that will substitute missing values of nominal variables by the training set mode of those variables.

step_modeimpute(recipe, ..., role = NA, trained = FALSE, modes = NULL,
  skip = FALSE)

# S3 method for step_modeimpute
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See [selections()] for more details. For the `tidy` method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

modes
A named character vector of modes. This is `NULL` until computed by [prep.recipe()].

skip
A logical. Should the step be skipped when the recipe is baked by [bake.recipe()]? While all operations are baked when [prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations

x
A `step_modeimpute` object.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any). For the `tidy` method, a tibble with columns `terms` (the selectors or variables selected) and `model` (the mode value).

Details

`step_modeimpute` estimates the variable modes from the data used in the `training` argument of `prep.recipe`. `bake.recipe` then applies the new values to new data sets using these values. If the training set data has more than one mode, one is selected at random.

Examples

data("credit_data")

## missing data per column
vapply(credit_data, function(x) mean(is.na(x)), c(num = 0))

#>       Status    Seniority         Home         Time          Age      Marital 
#> 0.0000000000 0.0000000000 0.0013471037 0.0000000000 0.0000000000 0.0002245173 
#>      Records          Job     Expenses       Income       Assets         Debt 
#> 0.0000000000 0.0004490346 0.0000000000 0.0855410867 0.0105523125 0.0040413112 
#>       Amount        Price 
#> 0.0000000000 0.0000000000 

set.seed(342)
in_training <- sample(1:nrow(credit_data), 2000)

credit_tr <- credit_data[ in_training, ]
credit_te <- credit_data[-in_training, ]
missing_examples <- c(14, 394, 565)

rec <- recipe(Price ~ ., data = credit_tr)

impute_rec <- rec %>%
  step_modeimpute(Status, Home, Marital)

imp_models <- prep(impute_rec, training = credit_tr)

imputed_te <- bake(imp_models, newdata = credit_te, everything())

table(credit_te$Home, imputed_te$Home, useNA = "always")

#>          
#>           ignore other owner parents priv rent <NA>
#>   ignore      14     0     0       0    0    0    0
#>   other        0   175     0       0    0    0    0
#>   owner        0     0  1132       0    0    0    0
#>   parents      0     0     0     441    0    0    0
#>   priv         0     0     0       0  134    0    0
#>   rent         0     0     0       0    0  556    0
#>   <NA>         0     0     2       0    0    0    0

tidy(impute_rec, number = 1)

#> # A tibble: 3 x 2
#>   terms   model
#>   <chr>   <chr>
#> 1 Status  <NA> 
#> 2 Home    <NA> 
#> 3 Marital <NA> 

tidy(imp_models, number = 1)

#> # A tibble: 3 x 2
#>   terms   model  
#>   <chr>   <chr>  
#> 1 Status  good   
#> 2 Home    owner  
#> 3 Marital married