Mutate multiple columns using dplyr

step_mutate_at creates a specification of a recipe step that will modify the selected variables using a common function via dplyr::mutate_at().

Usage

step_mutate_at(
  recipe,
  ...,
  fn,
  role = "predictor",
  trained = FALSE,
  inputs = NULL,
  skip = FALSE,
  id = rand_id("mutate_at")
)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose variables for this step. See selections() for more details.

fn

A function fun, a quosure style lambda `~ fun(.)`` or a list of either form. (see dplyr::mutate_at()). Note that this argument must be named.

role

For model terms created by this step, what analysis role should they be assigned? By default, the new columns created by this step from the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

inputs

A vector of column names populated by prep().

skip

A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Details

When using this flexible step, use extra care to avoid data leakage in your preprocessing. Consider, for example, the transformation x = w > mean(w). When applied to new data or testing data, this transformation would use the mean of w from the new data, not the mean of w from the training data.

Tidying

When you tidy() this step, a tibble with column terms which contains the columns being transformed is returned.

Case weights

The underlying operation does not allow for case weights.

See also

Examples

library(dplyr)
recipe(~., data = iris) %>%
  step_mutate_at(contains("Length"), fn = ~ 1 / .) %>%
  prep() %>%
  bake(new_data = NULL) %>%
  slice(1:10)
#> # A tibble: 10 × 5
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>  
#>  1        0.196         3.5        0.714         0.2 setosa 
#>  2        0.204         3          0.714         0.2 setosa 
#>  3        0.213         3.2        0.769         0.2 setosa 
#>  4        0.217         3.1        0.667         0.2 setosa 
#>  5        0.2           3.6        0.714         0.2 setosa 
#>  6        0.185         3.9        0.588         0.4 setosa 
#>  7        0.217         3.4        0.714         0.3 setosa 
#>  8        0.2           3.4        0.667         0.2 setosa 
#>  9        0.227         2.9        0.714         0.2 setosa 
#> 10        0.204         3.1        0.667         0.1 setosa 

recipe(~., data = iris) %>%
  # leads to more columns being created.
  step_mutate_at(contains("Length"), fn = list(log = log, sqrt = sqrt)) %>%
  prep() %>%
  bake(new_data = NULL) %>%
  slice(1:10)
#> # A tibble: 10 × 9
#>    Sepal…¹ Sepal…² Petal…³ Petal…⁴ Species Sepal…⁵ Petal…⁶ Sepal…⁷ Petal…⁸
#>      <dbl>   <dbl>   <dbl>   <dbl> <fct>     <dbl>   <dbl>   <dbl>   <dbl>
#>  1     5.1     3.5     1.4     0.2 setosa     1.63   0.336    2.26    1.18
#>  2     4.9     3       1.4     0.2 setosa     1.59   0.336    2.21    1.18
#>  3     4.7     3.2     1.3     0.2 setosa     1.55   0.262    2.17    1.14
#>  4     4.6     3.1     1.5     0.2 setosa     1.53   0.405    2.14    1.22
#>  5     5       3.6     1.4     0.2 setosa     1.61   0.336    2.24    1.18
#>  6     5.4     3.9     1.7     0.4 setosa     1.69   0.531    2.32    1.30
#>  7     4.6     3.4     1.4     0.3 setosa     1.53   0.336    2.14    1.18
#>  8     5       3.4     1.5     0.2 setosa     1.61   0.405    2.24    1.22
#>  9     4.4     2.9     1.4     0.2 setosa     1.48   0.336    2.10    1.18
#> 10     4.9     3.1     1.5     0.1 setosa     1.59   0.405    2.21    1.22
#> # … with abbreviated variable names ¹​Sepal.Length, ²​Sepal.Width,
#> #   ³​Petal.Length, ⁴​Petal.Width, ⁵​Sepal.Length_log, ⁶​Petal.Length_log,
#> #   ⁷​Sepal.Length_sqrt, ⁸​Petal.Length_sqrt