Remove observations with missing values

step_naomit creates a specification of a recipe step that will add remove observations (rows of data) if they contain NA or NaN values.

step_naomit(recipe, ..., role = NA, trained = FALSE, skip = FALSE)

# S3 method for step_naomit
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables will be used to create the dummy variables. See selections() for more details. The selected variables must be factors.

role
Unused, include for consistency with other steps.

trained
A logical to indicate if the quantities for preprocessing have been estimated. Again included for consistency.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

x
A step_naomit object.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

See also

Examples

recipe(Ozone ~ ., data = airquality) %>%
  step_naomit(Solar.R) %>%
  prep(airquality, verbose = FALSE, retain = TRUE) %>%
  juice()

#> # A tibble: 146 x 6
#>    Solar.R  Wind  Temp Month   Day Ozone
#>      <int> <dbl> <int> <int> <int> <int>
#>  1     190   7.4    67     5     1    41
#>  2     118   8      72     5     2    36
#>  3     149  12.6    74     5     3    12
#>  4     313  11.5    62     5     4    18
#>  5     299   8.6    65     5     7    23
#>  6      99  13.8    59     5     8    19
#>  7      19  20.1    61     5     9     8
#>  8     194   8.6    69     5    10    NA
#>  9     256   9.7    69     5    12    16
#> 10     290   9.2    66     5    13    11
#> # ... with 136 more rows