NNMF Signal Extraction

step_nnmf creates a specification of a recipe step that will convert numeric data into one or more non-negative components.

step_nnmf(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  num_comp = 2,
  num_run = 30,
  options = list(),
  res = NULL,
  prefix = "NNMF",
  seed = sample.int(10^5, 1),
  skip = FALSE,
  id = rand_id("nnmf")
)

# S3 method for step_nnmf
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables will be used to compute the components. See selections() for more details. For the tidy method, these are not currently used.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new component columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

num_comp
The number of components to retain as new predictors. If num_comp is greater than the number of columns or the number of possible components, a smaller value will be used.

num_run
A positive integer for the number of computations runs used to obtain a consensus projection.

options
A list of options to nmf() in the NMF package by way of the NNMF() function in the dimRed package. Note that the arguments data and ndim should not be passed here.

res
The NNMF() object is stored here once this preprocessing step has been trained by prep.recipe().

prefix
A character string that will be the prefix to the resulting new variables. See notes below.

seed
An integer that will be used to set the seed in isolation when computing the factorization.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_nnmf object.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms (the selectors or variables selected) and the number of components.

Details

Non-negative matrix factorization computes latent components that have non-negative values and take into account that the original data have non-negative values.

The argument num_comp controls the number of components that will be retained (the original variables that are used to derive the components are removed from the data). The new components will have names that begin with prefix and a sequence of numbers. The variable names are padded with zeros. For example, if num < 10, their names will be NNMF1 - NNMF9. If num = 101, the names would be NNMF001 - NNMF101.

See also

Examples

library(modeldata)
data(biomass)

# rec <- recipe(HHV ~ ., data = biomass) %>%
#   update_role(sample, new_role = "id var") %>%
#   update_role(dataset, new_role = "split variable") %>%
#   step_nnmf(all_predictors(), num_comp = 2, seed = 473, num_run = 2) %>%
#   prep(training = biomass)
#
# bake(rec, new_data = NULL)
#
# library(ggplot2)
# bake(rec, new_data = NULL) %>%
#  ggplot(aes(x = NNMF2, y = NNMF1, col = HHV)) + geom_point()