Simple Value Assignments for Novel Factor Levels

step_novel creates a specification of a recipe step that will assign a previously unseen factor level to a new value.

step_novel(recipe, ..., role = NA, trained = FALSE,
  new_level = "new", objects = NULL, skip = FALSE,
  id = rand_id("novel"))

# S3 method for step_novel
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables that will be affected by the step. These variables should be character or factor types. See selections() for more details. For the tidy method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

new_level
A single character value that will be assigned to new factor levels.

objects
A list of objects that contain the information on factor levels that will be determined by prep.recipe().

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_novel object.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms (the columns that will be affected) and value (the factor levels that is used for the new value)

Details

The selected variables are adjusted to have a new level (given by new_level) that is placed in the last position. During preparation there will be no data points associated with this new level since all of the data have been seen.

Note that if the original columns are character, they will be converted to factors by this step.

Missing values will remain missing.

If new_level is already in the data given to prep, an error is thrown.

See also

Examples

data(okc)

okc_tr <- okc[1:30000,]
okc_te <- okc[30001:30006,]
okc_te$diet[3] <- "cannibalism"
okc_te$diet[4] <- "vampirism"

rec <- recipe(~ diet + location, data = okc_tr)

rec <- rec %>%
  step_novel(diet, location)
rec <- prep(rec, training = okc_tr)

processed <- bake(rec, okc_te)
tibble(old = okc_te$diet, new = processed$diet)

#> # A tibble: 6 x 2
#>   old               new              
#>   <chr>             <fct>            
#> 1 vegetarian        vegetarian       
#> 2 strictly anything strictly anything
#> 3 cannibalism       new              
#> 4 vampirism         new              
#> 5 <NA>              <NA>             
#> 6 <NA>              <NA>             

tidy(rec, number = 1)

#> # A tibble: 2 x 3
#>   terms    value id         
#>   <chr>    <chr> <chr>      
#> 1 diet     new   novel_XMukl
#> 2 location new   novel_XMukl