Nature Spline Basis Functions

step_ns creates a specification of a recipe step that will create new columns that are basis expansions of variables using natural spines.

step_ns(recipe, terms, role = "predictor", trained = FALSE,
  objects = NULL, options = list(df = 2))

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

terms

A formula that represents the variables or terms that will be processed. The raw variable names can be used or dplyr selection tools. See selections for more details.

role

For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created from the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

objects

A list of ns objects created once the step has been learned.

options

A list of options for ns which should not include x.

Value

step_ns returns an object of class step_ns.

Details

step_ns can new features from a single variable that enable fitting routines to model this variable in a nonlinear manner. The extent of the possible nonlinearity is determined by the df or knot arguments of ns. The original variables are removed from the data and new columns are added. The naming convention for the new variables is varname_ns_1 and so on.

See also

step_poly recipe learn.recipe process.recipe

Examples

data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)

library(magrittr)
with_splines <- rec %>%
  step_ns(terms = ~ carbon + hydrogen)
with_splines <- learn(with_splines, training = biomass_tr)

#> step 1 ns training 

expanded <- process(with_splines, biomass_te)
expanded

#> # A tibble: 80 × 8
#>    oxygen nitrogen sulfur    HHV carbon_ns_1 carbon_ns_2 hydrogen_ns_1
#> *   <dbl>    <dbl>  <dbl>  <dbl>       <dbl>       <dbl>         <dbl>
#> 1   47.20     0.30   0.22 18.275   0.5235017  -0.2360067     0.5626398
#> 2   48.06     2.85   0.34 17.560   0.4929257  -0.2408145     0.5563744
#> 3   49.10     2.40   0.30 17.173   0.4868199  -0.2407445     0.5563744
#> 4   37.30     1.80   0.50 18.851   0.5239392  -0.2358537     0.5739704
#> 5   42.77     0.20   0.00 20.547   0.5424882  -0.2257850     0.5773303
#> 6   41.70     0.70   0.20 18.467   0.5040254  -0.2401936     0.5563744
#> 7   54.13     1.19   0.51 15.095   0.4400424  -0.2334398     0.5444179
#> 8   33.80     0.95   0.20 16.240   0.4799363  -0.2403662     0.5115938
#> 9   31.10     0.14   4.90 11.147   0.2850862  -0.1687288     0.4934496
#> 10  23.69     4.63   1.05 10.750   0.2596973  -0.1552040     0.4420907
#> # ... with 70 more rows, and 1 more variables: hydrogen_ns_2 <dbl>