Collapse Some Categorical Levels

`step_other` creates a *specification* of a recipe step that will potentially pool infrequently occurring values into an "other" category.

step_other(recipe, ..., role = NA, trained = FALSE, threshold = 0.05,
  other = "other", objects = NULL, skip = FALSE)

# S3 method for step_other
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables that will potentially be reduced. See [selections()] for more details. For the `tidy` method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

threshold
A single numeric value in (0, 1) for pooling.

other
A single character value for the "other" category.

objects
A list of objects that contain the information to pool infrequent levels that is determined by [prep.recipe()].

skip
A logical. Should the step be skipped when the recipe is baked by [bake.recipe()]? While all operations are baked when [prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations

x
A `step_other` object.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any). For the `tidy` method, a tibble with columns `terms` (the columns that will be affected) and `retained` (the factor levels that were not pulled into "other")

Details

The overall proportion of the categories are computed. The "other" category is used in place of any categorical levels whose individual proportion in the training set is less than `threshold`.

If no pooling is done the data are unmodified (although character data may be changed to factors based on the value of `stringsAsFactors` in [prep.recipe()]). Otherwise, a factor is always returned with different factor levels.

If `threshold` is less than the largest category proportion, all levels except for the most frequent are collapsed to the `other` level.

If the retained categories include the value of `other`, an error is thrown. If `other` is in the list of discarded levels, no error occurs.

See also

Examples

data(okc)

set.seed(19)
in_train <- sample(1:nrow(okc), size = 30000)

okc_tr <- okc[ in_train,]
okc_te <- okc[-in_train,]

rec <- recipe(~ diet + location, data = okc_tr)

rec <- rec %>%
  step_other(diet, location, threshold = .1, other = "other values")
rec <- prep(rec, training = okc_tr)

collapsed <- bake(rec, okc_te)
table(okc_te$diet, collapsed$diet, useNA = "always")

#>                      
#>                       anything mostly anything strictly anything other values
#>   anything                3013               0                 0            0
#>   halal                      0               0                 0            5
#>   kosher                     0               0                 0            6
#>   mostly anything            0            8254                 0            0
#>   mostly halal               0               0                 0           27
#>   mostly kosher              0               0                 0           39
#>   mostly other               0               0                 0          517
#>   mostly vegan               0               0                 0          186
#>   mostly vegetarian          0               0                 0         1719
#>   other                      0               0                 0          162
#>   strictly anything          0               0              2497            0
#>   strictly halal             0               0                 0            7
#>   strictly kosher            0               0                 0           11
#>   strictly other             0               0                 0          223
#>   strictly vegan             0               0                 0          121
#>   strictly vegetarian        0               0                 0          437
#>   vegan                      0               0                 0           71
#>   vegetarian                 0               0                 0          336
#>   <NA>                       0               0                 0            0
#>                      
#>                        <NA>
#>   anything                0
#>   halal                   0
#>   kosher                  0
#>   mostly anything         0
#>   mostly halal            0
#>   mostly kosher           0
#>   mostly other            0
#>   mostly vegan            0
#>   mostly vegetarian       0
#>   other                   0
#>   strictly anything       0
#>   strictly halal          0
#>   strictly kosher         0
#>   strictly other          0
#>   strictly vegan          0
#>   strictly vegetarian     0
#>   vegan                   0
#>   vegetarian              0
#>   <NA>                12224

tidy(rec, number = 1)

#> # A tibble: 5 x 2
#>   terms    retained                 
#>   <chr>    <chr>                    
#> 1 diet     anything                 
#> 2 diet     mostly anything          
#> 3 diet     strictly anything        
#> 4 location oakland, california      
#> 5 location san francisco, california