Partial Least Squares Feature Extraction

step_pls creates a specification of a recipe step that will convert numeric data into one or more new dimensions.

step_pls(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  num_comp = 2,
  predictor_prop = 1,
  outcome = NULL,
  options = list(scale = TRUE),
  preserve = FALSE,
  res = NULL,
  prefix = "PLS",
  skip = FALSE,
  id = rand_id("pls")
)

# S3 method for step_pls
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables will be used to compute the dimensions. See selections() for more details. For the tidy method, these are not currently used.

role
For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new dimension columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

num_comp
The number of pls dimensions to retain as new predictors. If num_comp is greater than the number of columns or the number of possible dimensions, a smaller value will be used.

predictor_prop
The maximum number of original predictors that can have non-zero coefficients for each PLS component (via regularization).

outcome
When a single outcome is available, character string or call to dplyr::vars() can be used to specify a single outcome variable.

options
A list of options to mixOmics::pls(), mixOmics::spls(), mixOmics::plsda(), or mixOmics::splsda() (depending on the data and arguments).

preserve
A single logical: should the original predictor data be retained along with the new features?

res
A list of results are stored here once this preprocessing step has been trained by prep.recipe().

prefix
A character string that will be the prefix to the resulting new variables. See notes below.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_pls object

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms (the selectors or variables selected), components, and values.

Details

PLS is a supervised version of principal component analysis that requires the outcome data to compute the new features.

This step requires the Bioconductor mixOmics package. If not installed, the step will stop with a note about installing the package.

The argument num_comp controls the number of components that will be retained (the original variables that are used to derive the components are removed from the data). The new components will have names that begin with prefix and a sequence of numbers. The variable names are padded with zeros. For example, if num_comp < 10, their names will be PLS1 - PLS9. If num_comp = 101, the names would be PLS001 - PLS101.

Sparsity can be encouraged using the predictor_prop parameter. This affects each PLS component, and indicates the maximum proportion of predictors with non-zero coefficients in each component. step_pls() converts this proportion to determine the keepX parameter in mixOmics::spls() and mixOmics::splsda(). See the references in mixOmics::spls() for details.

The tidy() method returns the coefficients that are usually defined as

$$W(P'W)^{-1}$$

(See the Wikipedia article below)

When applied to data, these values are usually scaled by a column-specific norm. The tidy() method applies this same norm to the coefficients shown above.

References

https://en.wikipedia.org/wiki/Partial_least_squares_regression

Rohart F, Gautier B, Singh A, Lê Cao K-A (2017) mixOmics: An R package for 'omics feature selection and multiple data integration. PLoS Comput Biol 13(11): e1005752. https://doi.org/10.1371/journal.pcbi.1005752

See also

Examples

# requires the Bioconductor mixOmics package
data(biomass, package = "modeldata")

biom_tr <-
  biomass %>%
  dplyr::filter(dataset == "Training") %>%
  dplyr::select(-dataset,-sample)
biom_te <-
  biomass %>%
  dplyr::filter(dataset == "Testing")  %>%
  dplyr::select(-dataset,-sample,-HHV)

dense_pls <-
  recipe(HHV ~ ., data = biom_tr) %>%
  step_pls(all_predictors(), outcome = "HHV", num_comp = 3)

sparse_pls <-
  recipe(HHV ~ ., data = biom_tr) %>%
  step_pls(all_predictors(), outcome = "HHV", num_comp = 3, predictor_prop = 4/5)

## -----------------------------------------------------------------------------
## PLS discriminant analysis

data(cells, package = "modeldata")

cell_tr <-
  cells %>%
  dplyr::filter(case == "Train") %>%
  dplyr::select(-case)
cell_te <-
  cells %>%
  dplyr::filter(case == "Test")  %>%
  dplyr::select(-case,-class)

dense_plsda <-
  recipe(class ~ ., data = cell_tr) %>%
  step_pls(all_predictors(), outcome = "class", num_comp = 5)

sparse_plsda <-
  recipe(class ~ ., data = cell_tr) %>%
  step_pls(all_predictors(), outcome = "class", num_comp = 5, predictor_prop = 1/4)