Orthogonal Polynomial Basis Functions

step_poly creates a specification of a recipe step that will create new columns that are basis expansions of variables using orthogonal polynomials.

step_poly(recipe, terms, role = "predictor", trained = FALSE,
  objects = NULL, options = list(degree = 2))

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

terms

A formula that represents the variables or terms that will be processed. The raw variable names can be used or dplyr selection tools. See selections for more details.

role

For model terms created by this step, what analysis role should they be assigned?. By default, the function assumes that the new columns created from the original variables will be used as all_predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

objects

A list of poly objects created once the step has been learned.

options

A list of options for poly which should not include x or simple. Note that the option raw = TRUE will produce the regular polynomial values (not orthogonalized).

Value

step_poly returns an object of class step_poly.

Details

step_poly can new features from a single variable that enable fitting routines to model this variable in a nonlinear manner. The extent of the possible nonlinearity is determined by the degree argument of poly. The original variables are removed from the data and new columns are added. The naming convention for the new variables is varname_poly_1 and so on.

See also

step_ns recipe learn.recipe process.recipe

Examples

data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)

library(magrittr)
quadratic <- rec %>%
  step_poly(terms = ~ carbon + hydrogen)
quadratic <- learn(quadratic, training = biomass_tr)

#> step 1 poly training 

expanded <- process(quadratic, biomass_te)
expanded

#> # A tibble: 80 × 8
#>    oxygen nitrogen sulfur    HHV carbon_poly_1 carbon_poly_2 hydrogen_poly_1
#> *   <dbl>    <dbl>  <dbl>  <dbl>         <dbl>         <dbl>           <dbl>
#> 1   47.20     0.30   0.22 18.275  -0.009027012  -0.012392984     0.008261592
#> 2   48.06     2.85   0.34 17.560  -0.022989944   0.004033492     0.001602842
#> 3   49.10     2.40   0.30 17.173  -0.025467239   0.007338302     0.001602842
#> 4   37.30     1.80   0.50 18.851  -0.008801803  -0.012627295     0.025104314
#> 5   42.77     0.20   0.00 20.547   0.001828042  -0.022580618     0.033721520
#> 6   41.70     0.70   0.20 18.467  -0.018260564  -0.001948950     0.001602842
#> 7   54.13     1.19   0.51 15.095  -0.042402924   0.033082977    -0.008972821
#> 8   33.80     0.95   0.20 16.240  -0.028169742   0.011077745    -0.031299219
#> 9   31.10     0.14   4.90 11.147  -0.086273557   0.125345872    -0.041483191
#> 10  23.69     4.63   1.05 10.750  -0.092579397   0.141640714    -0.066159736
#> # ... with 70 more rows, and 1 more variables: hydrogen_poly_2 <dbl>