Scaling Numeric Data to a Specific Range

step_range creates a specification of a recipe step that will normalize numeric data to have a standard deviation of one.

step_range(recipe, ..., role = NA, trained = FALSE, min = 0, max = 1,
  ranges = NULL)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables will be scaled. See selections for more details.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

min
A single numeric value for the smallest value in the range

max
A single numeric value for the largest value in the range

ranges
A character vector of variables that will be normalized. Note that this is ignored until the values are determined by learn.recipe. Setting this value will be ineffective.

Value

step_range returns an object of class step_range.

Details

Scaling data means that the standard deviation of a variable is divided out of the data. step_range estimates the variable standard deviations from the data used in the training argument of learn.recipe. process.recipe then applies the scaling to new data sets using these standard deviations.

Examples

data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)

ranged_trans <- rec %>%
  step_range(carbon, hydrogen)

ranged_obj <- learn(ranged_trans, training = biomass_tr)

#> step 1 range training 

transformed_te <- process(ranged_obj, biomass_te)

biomass_te[1:10, names(transformed_te)]

#>    carbon hydrogen oxygen nitrogen sulfur
#> 15  46.35     5.67  47.20     0.30   0.22
#> 20  43.25     5.50  48.06     2.85   0.34
#> 26  42.70     5.50  49.10     2.40   0.30
#> 31  46.40     6.10  37.30     1.80   0.50
#> 36  48.76     6.32  42.77     0.20   0.00
#> 41  44.30     5.50  41.70     0.70   0.20
#> 46  38.94     5.23  54.13     1.19   0.51
#> 51  42.10     4.66  33.80     0.95   0.20
#> 55  29.20     4.40  31.10     0.14   4.90
#> 65  27.80     3.77  23.69     4.63   1.05

transformed_te

#> # A tibble: 80 x 5
#>       carbon  hydrogen oxygen nitrogen sulfur
#>        <dbl>     <dbl>  <dbl>    <dbl>  <dbl>
#>  1 0.3844011 0.4895833  47.20     0.30   0.22
#>  2 0.3468572 0.4748264  48.06     2.85   0.34
#>  3 0.3401962 0.4748264  49.10     2.40   0.30
#>  4 0.3850067 0.5269097  37.30     1.80   0.50
#>  5 0.4135885 0.5460069  42.77     0.20   0.00
#>  6 0.3595737 0.4748264  41.70     0.70   0.20
#>  7 0.2946591 0.4513889  54.13     1.19   0.51
#>  8 0.3329296 0.4019097  33.80     0.95   0.20
#>  9 0.1766986 0.3793403  31.10     0.14   4.90
#> 10 0.1597432 0.3246528  23.69     4.63   1.05
#> # ... with 70 more rows