Ratio Variable Creation

step_ratio creates a specification of a recipe step that will create one or more ratios out of numeric variables.

Usage

step_ratio(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  denom = denom_vars(),
  naming = function(numer, denom) {
     make.names(paste(numer, denom, sep = "_o_"))

    },
  columns = NULL,
  keep_original_cols = TRUE,
  skip = FALSE,
  id = rand_id("ratio")
)

denom_vars(...)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose which variables will be used in the numerator of the ratio. When used with denom_vars, the dots indicate which variables are used in the denominator. See selections() for more details.

role

For model terms created by this step, what analysis role should they be assigned? By default, the new columns created by this step from the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

denom

A call to denom_vars to specify which variables are used in the denominator that can include specific variable names separated by commas or different selectors (see selections()). If a column is included in both lists to be numerator and denominator, it will be removed from the listing.

naming

A function that defines the naming convention for new ratio columns.

columns

The column names used in the ratios. This argument is not populated until prep() is executed.

keep_original_cols

A logical to keep the original variables in the output. Defaults to TRUE.

skip

A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Tidying

When you tidy() this step, a tibble with columns terms (the selectors or variables selected) and denom is returned.

Case weights

The underlying operation does not allow for case weights.

See also

Examples

library(recipes)
data(biomass, package = "modeldata")

biomass$total <- apply(biomass[, 3:7], 1, sum)
biomass_tr <- biomass[biomass$dataset == "Training", ]
biomass_te <- biomass[biomass$dataset == "Testing", ]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen +
  sulfur + total,
data = biomass_tr
)

ratio_recipe <- rec %>%
  # all predictors over total
  step_ratio(all_numeric_predictors(), denom = denom_vars(total)) %>%
  # get rid of the original predictors
  step_rm(all_predictors(), -ends_with("total"))

ratio_recipe <- prep(ratio_recipe, training = biomass_tr)

ratio_data <- bake(ratio_recipe, biomass_te)
ratio_data
#> # A tibble: 80 × 7
#>    total   HHV carbon_o_total hydrogen_o_total oxygen_o_…¹ nitro…² sulfu…³
#>    <dbl> <dbl>          <dbl>            <dbl>       <dbl>   <dbl>   <dbl>
#>  1  99.7  18.3          0.465           0.0568       0.473 0.00301 0.00221
#>  2 100    17.6          0.432           0.055        0.481 0.0285  0.0034 
#>  3 100    17.2          0.427           0.055        0.491 0.024   0.003  
#>  4  92.1  18.9          0.504           0.0662       0.405 0.0195  0.00543
#>  5  98.0  20.5          0.497           0.0645       0.436 0.00204 0      
#>  6  92.4  18.5          0.479           0.0595       0.451 0.00758 0.00216
#>  7 100    15.1          0.389           0.0523       0.541 0.0119  0.0051 
#>  8  81.7  16.2          0.515           0.0570       0.414 0.0116  0.00245
#>  9  69.7  11.1          0.419           0.0631       0.446 0.00201 0.0703 
#> 10  60.9  10.8          0.456           0.0619       0.389 0.0760  0.0172 
#> # … with 70 more rows, and abbreviated variable names ¹​oxygen_o_total,
#> #   ²​nitrogen_o_total, ³​sulfur_o_total