Create Dummy Variables using Regular Expressions

step_regex creates a specification of a recipe step that will create a new dummy variable based on a regular expression.

step_regex(recipe, ..., role = "predictor", trained = FALSE,
  pattern = ".", options = list(), result = make.names(pattern),
  input = NULL)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
A single selector functions to choose which variable will be searched for the pattern. The selector should resolve into a single variable. See selections for more details.

role
For a variable created by this step, what analysis role should they be assigned?. By default, the function assumes that the new dummy variable column created by the original variable will be used as a predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

pattern
A character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector. Coerced by as.character to a character string if possible.

options
A list of options to grepl that should not include x or pattern.

result
A single character value for the name of the new variable. It should be a valid column name.

input
A single character value for the name of the variable being searched. This is NULL until computed by prepare.recipe.

Value

step_regex returns an object of class step_regex. A new variable will be created in the data set by bake.recipe.

Examples

data(covers)

rec <- recipe(~ description, covers) %>%
  step_regex(description, pattern = "(rock|stony)", result = "rocks") %>%
  step_regex(description, pattern = "ratake families")

rec2 <- prepare(rec, training = covers)

#> step 1 regex training 
#> step 2 regex training 

rec2

#> Data Recipe
#> 
#> Inputs:
#> 
#>       role #variables
#>  predictor          1
#> 
#> Training data contained 40 data points and no missing data.
#> 
#> Steps:
#> 
#> Regular expression dummy variable using `(rock|stony)` [trained]
#> Regular expression dummy variable using `ratake families` [trained]

with_dummies <- bake(rec2, newdata = covers)
with_dummies

#> # A tibble: 40 x 3
#>                                                   description rocks
#>                                                        <fctr> <dbl>
#>  1    1,cathedral family,rock outcrop complex,extremely stony     1
#>  2                 2,vanet,ratake families complex,very stony     1
#>  3                 3,haploborolis,rock outcrop complex,rubbly     1
#>  4                4,ratake family,rock outcrop complex,rubbly     1
#>  5         5,vanet family,rock outcrop complex complex,rubbly     1
#>  6        6,vanet,wetmore families,rock outcrop complex,stony     1
#>  7                                            7,gothic family     0
#>  8                       8,supervisor,limber families complex     0
#>  9                             9,troutville family,very stony     1
#> 10 10,bullwark,catamount families,rock outcrop complex,rubbly     1
#> # ... with 30 more rows, and 1 more variables: ratake.families <dbl>