Create Dummy Variables using Regular Expressions

`step_regex` creates a *specification* of a recipe step that will create a new dummy variable based on a regular expression.

step_regex(recipe, ..., role = "predictor", trained = FALSE,
  pattern = ".", options = list(), result = make.names(pattern),
  input = NULL, skip = FALSE)

# S3 method for step_regex
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
A single selector functions to choose which variable will be searched for the pattern. The selector should resolve into a single variable. See [selections()] for more details. For the `tidy` method, these are not currently used.

role
For a variable created by this step, what analysis role should they be assigned?. By default, the function assumes that the new dummy variable column created by the original variable will be used as a predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

pattern
A character string containing a regular expression (or character string for `fixed = TRUE`) to be matched in the given character vector. Coerced by `as.character` to a character string if possible.

options
A list of options to [grepl()] that should not include `x` or `pattern`.

result
A single character value for the name of the new variable. It should be a valid column name.

input
A single character value for the name of the variable being searched. This is `NULL` until computed by [prep.recipe()].

skip
A logical. Should the step be skipped when the recipe is baked by [bake.recipe()]? While all operations are baked when [prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations

x
A `step_regex` object.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any). For the `tidy` method, a tibble with columns `terms` (the selectors or variables selected) and `result` (the new column name).

Examples

data(covers)

rec <- recipe(~ description, covers) %>%
  step_regex(description, pattern = "(rock|stony)", result = "rocks") %>%
  step_regex(description, pattern = "ratake families")

rec2 <- prep(rec, training = covers)
rec2

#> Data Recipe
#> 
#> Inputs:
#> 
#>       role #variables
#>  predictor          1
#> 
#> Training data contained 40 data points and no missing data.
#> 
#> Operations:
#> 
#> Regular expression dummy variable using `(rock|stony)` [trained]
#> Regular expression dummy variable using `ratake families` [trained]

with_dummies <- bake(rec2, newdata = covers)
with_dummies

#> # A tibble: 40 x 3
#>    description                                            rocks ratake.families
#>    <fct>                                                  <dbl>           <dbl>
#>  1 1,cathedral family,rock outcrop complex,extremely sto…    1.              0.
#>  2 2,vanet,ratake families complex,very stony                1.              1.
#>  3 3,haploborolis,rock outcrop complex,rubbly                1.              0.
#>  4 4,ratake family,rock outcrop complex,rubbly               1.              0.
#>  5 5,vanet family,rock outcrop complex complex,rubbly        1.              0.
#>  6 6,vanet,wetmore families,rock outcrop complex,stony       1.              0.
#>  7 7,gothic family                                           0.              0.
#>  8 8,supervisor,limber families complex                      0.              0.
#>  9 9,troutville family,very stony                            1.              0.
#> 10 10,bullwark,catamount families,rock outcrop complex,r…    1.              0.
#> # ... with 30 more rows

tidy(rec, number = 1)

#> # A tibble: 1 x 2
#>   terms       result
#>   <chr>       <chr> 
#> 1 description <NA>  

tidy(rec2, number = 1)

#> # A tibble: 1 x 2
#>   terms       result
#>   <chr>       <chr> 
#> 1 description rocks