Apply (Smoothed) Rectified Linear Transformation

`step_relu` creates a *specification* of a recipe step that will apply the rectified linear or softplus transformations to numeric data. The transformed data is added as new columns to the data matrix.

step_relu(recipe, ..., role = "predictor", trained = FALSE, shift = 0,
  reverse = FALSE, smooth = FALSE, prefix = "right_relu_", skip = FALSE)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See [selections()] for more details.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

shift
A numeric value dictating a translation to apply to the data.

reverse
A logical to indicate if the left hinge should be used as opposed to the right hinge.

smooth
A logical indicating if the softplus function, a smooth appromixation to the rectified linear transformation, should be used.

prefix
A prefix for generated column names, default to "right_relu_" when right hinge transformation and "left_relu_" for reversed/left hinge transformations.

skip
A logical. Should the step be skipped when the recipe is baked by [bake.recipe()]? While all operations are baked when [prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any).

Details

The rectified linear transformation is calculated as $$max(0, x - c)$$ and is also known as the ReLu or right hinge function. If `reverse` is true, then the transformation is reflected about the y-axis, like so: $$max(0, c - x)$$ Setting the `smooth` option to true will instead calculate a smooth approximation to ReLu according to $$ln(1 + e^(x - c)$$ The `reverse` argument may also be applied to this transformation.

Connection to MARS

The rectified linear transformation is used in the Multivariate Adaptive Regression Splines as a basis function to fit piecewise linear functions to data in a strategy similar to that employeed in tree based models. The transformation is a popular choice as an activation function in many neural networks, which could then be seen as a stacked generalization of MARS when making use of ReLu activations. The hinge function also appears in the loss function of Support Vector Machines, where it penalizes residuals only if they are within a certain margin of the decision boundary.

See also

Examples

data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)

transformed_te <- rec %>%
  step_relu(carbon, shift = 40) %>%
  prep(biomass_tr) %>%
  bake(biomass_te)

transformed_te

#> # A tibble: 80 x 7
#>    carbon hydrogen oxygen nitrogen sulfur   HHV right_relu_carbon
#>     <dbl>    <dbl>  <dbl>    <dbl>  <dbl> <dbl>             <dbl>
#>  1   46.4     5.67   47.2    0.300  0.220  18.3              6.35
#>  2   43.2     5.50   48.1    2.85   0.340  17.6              3.25
#>  3   42.7     5.50   49.1    2.40   0.300  17.2              2.70
#>  4   46.4     6.10   37.3    1.80   0.500  18.9              6.40
#>  5   48.8     6.32   42.8    0.200  0.     20.5              8.76
#>  6   44.3     5.50   41.7    0.700  0.200  18.5              4.30
#>  7   38.9     5.23   54.1    1.19   0.510  15.1              0.  
#>  8   42.1     4.66   33.8    0.950  0.200  16.2              2.10
#>  9   29.2     4.40   31.1    0.140  4.90   11.1              0.  
#> 10   27.8     3.77   23.7    4.63   1.05   10.8              0.  
#> # ... with 70 more rows