Apply (Smoothed) Rectified Linear Transformation

step_relu creates a specification of a recipe step that will apply the rectified linear or softplus transformations to numeric data. The transformed data is added as new columns to the data matrix.

step_relu(recipe, ..., role = "predictor", trained = FALSE, shift = 0,
  reverse = FALSE, smooth = FALSE, prefix = "right_relu_", skip = FALSE)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

shift
A numeric value dictating a translation to apply to the data.

reverse
A logical to indicate if the left hinge should be used as opposed to the right hinge.

smooth
A logical indicating if the softplus function, a smooth appromixation to the rectified linear transformation, should be used.

prefix
A prefix for generated column names, default to "right_relu_" when right hinge transformation and "left_relu_" for reversed/left hinge transformations.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While are steps are baked when prep.recipe() is run, some steps may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent steps.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any).

Details

The rectified linear transformation is calculated as $$max(0, x - c)$$ and is also known as the ReLu or right hinge function. If reverse is true, then the transformation is reflected about the y-axis, like so: $$max(0, c - x)$$ Setting the smooth option to true will instead calculate a smooth approximation to ReLu according to $$ln(1 + e^(x - c)$$ The reverse argument may also be applied to this transformation.

Connection to MARS

The rectified linear transformation is used in the Multivariate Adaptive Regression Splines as a basis function to fit piecewise linear functions to data in a strategy similar to that employeed in tree based models. The transformation is a popular choice as an activation function in many neural networks, which could then be seen as a stacked generalization of MARS when making use of ReLu activations. The hinge function also appears in the loss function of Support Vector Machines, where it penalizes residuals only if they are within a certain margin of the decision boundary.

See also

recipe() prep.recipe() bake.recipe()

Examples

data(biomass)

biomass_tr <- biomass[biomass$dataset == "Training",]
biomass_te <- biomass[biomass$dataset == "Testing",]

rec <- recipe(HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur,
              data = biomass_tr)

transformed_te <- rec %>%
  step_relu(carbon, shift = 40) %>%
  prep(biomass_tr) %>%
  bake(biomass_te)

transformed_te

#> # A tibble: 80 x 7
#>    carbon hydrogen oxygen nitrogen sulfur    HHV right_relu_carbon
#>     <dbl>    <dbl>  <dbl>    <dbl>  <dbl>  <dbl>             <dbl>
#>  1  46.35     5.67  47.20     0.30   0.22 18.275              6.35
#>  2  43.25     5.50  48.06     2.85   0.34 17.560              3.25
#>  3  42.70     5.50  49.10     2.40   0.30 17.173              2.70
#>  4  46.40     6.10  37.30     1.80   0.50 18.851              6.40
#>  5  48.76     6.32  42.77     0.20   0.00 20.547              8.76
#>  6  44.30     5.50  41.70     0.70   0.20 18.467              4.30
#>  7  38.94     5.23  54.13     1.19   0.51 15.095              0.00
#>  8  42.10     4.66  33.80     0.95   0.20 16.240              2.10
#>  9  29.20     4.40  31.10     0.14   4.90 11.147              0.00
#> 10  27.80     3.77  23.69     4.63   1.05 10.750              0.00
#> # ... with 70 more rows