Rename variables by name

step_rename creates a specification of a recipe step that will add variables using dplyr::rename().

step_rename(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  inputs = NULL,
  skip = FALSE,
  id = rand_id("rename")
)

# S3 method for step_rename
tidy(x, ...)

# S3 method for step_rename_at
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more unquoted expressions separated by commas. See dplyr::rename() where the convention is new_name = old_name.

role
For model terms created by this step, what analysis role should they be assigned? By default, the function assumes that the new dimension columns created by the original variables will be used as predictors in a model.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

inputs
Quosure(s) of ....

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_rename object

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns values which contains the rename expressions as character strings (and are not reparsable).

Details

When an object in the user's global environment is referenced in the expression defining the new variable(s), it is a good idea to use quasiquotation (e.g. !!) to embed the value of the object in the expression (to be portable between sessions).

Examples

recipe( ~ ., data = iris) %>%
  step_rename(Sepal_Width = Sepal.Width) %>%
  prep() %>%
  bake(new_data = NULL) %>%
  slice(1:5)

#> # A tibble: 5 x 5
#>   Sepal.Length Sepal_Width Petal.Length Petal.Width Species
#>          <dbl>       <dbl>        <dbl>       <dbl> <fct>  
#> 1          5.1         3.5          1.4         0.2 setosa 
#> 2          4.9         3            1.4         0.2 setosa 
#> 3          4.7         3.2          1.3         0.2 setosa 
#> 4          4.6         3.1          1.5         0.2 setosa 
#> 5          5           3.6          1.4         0.2 setosa 

vars <- c(var1 = "cyl", var2 = "am")
car_rec <-
  recipe(~ ., data = mtcars) %>%
  step_rename(!!vars)

car_rec %>%
  prep() %>%
  bake(new_data = NULL)

#> # A tibble: 32 x 11
#>      mpg `<chr>...var1`  disp    hp  drat    wt  qsec    vs `<chr>...var2`  gear
#>    <dbl>          <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>          <dbl> <dbl>
#>  1  21                6  160    110  3.9   2.62  16.5     0              1     4
#>  2  21                6  160    110  3.9   2.88  17.0     0              1     4
#>  3  22.8              4  108     93  3.85  2.32  18.6     1              1     4
#>  4  21.4              6  258    110  3.08  3.22  19.4     1              0     3
#>  5  18.7              8  360    175  3.15  3.44  17.0     0              0     3
#>  6  18.1              6  225    105  2.76  3.46  20.2     1              0     3
#>  7  14.3              8  360    245  3.21  3.57  15.8     0              0     3
#>  8  24.4              4  147.    62  3.69  3.19  20       1              0     4
#>  9  22.8              4  141.    95  3.92  3.15  22.9     1              0     4
#> 10  19.2              6  168.   123  3.92  3.44  18.3     1              0     4
#> # … with 22 more rows, and 1 more variable: carb <dbl>

car_rec %>%
  tidy(number = 1)

#> # A tibble: 1 x 3
#>   terms value                              id          
#>   <chr> <chr>                              <chr>       
#> 1 <chr> "c(var1 = \"cyl\", var2 = \"am\")" rename_xzrZ1