Impute Numeric Data Using a Rolling Window Statistic

step_rollimpute creates a specification of a recipe step that will substitute missing values of numeric variables by the a measure of location (e.g. median) within a moving window.

step_rollimpute(recipe, ..., role = NA, trained = FALSE, columns = NULL,
  statistic = median, window = 5, skip = FALSE)

# S3 method for step_rollimpute
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables are affected by the step. See selections() for more details. These columns should be non-integer numerics (i.e., double precision). For the tidy method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

columns
A named numeric vector of columns. This is NULL until computed by prep.recipe().

statistic
A function with a single argument for the data to compute the imputed value. Only complete values will be passed to the function and it should return a double precision value.

window
The size of the window around a point to be imputed. Should be an odd integer greater than one. See Details below for a discussion of points at the ends of the series.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

x
A step_rollimpute object.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms (the selectors or variables selected) and window (the window size).

Details

On the tails, the window is shifted towards the ends. For example, for a 5-point window, the windows for the first four points are 1:5, 1:5, 1:5, and then 2:6.

When missing data are in the window, they are not passed to the function. If all of the data in the window are missing, a missing value is returned.

The statistics are calculated on the training set values before imputation. This means that if previous data within the window are missing, their imputed values are not included in the window data used for imputation. In other words, each imputation does not know anything about previous imputations in the series prior to the current point.

Examples

library(lubridate)

set.seed(145)
example_data <-
  data.frame(
    day = ymd("2012-06-07") + days(1:12),
    x1 = round(runif(12), 2),
    x2 = round(runif(12), 2),
    x3 = round(runif(12), 2)
  )
example_data$x1[c(1, 5, 6)] <- NA
example_data$x2[c(1:4, 10)] <- NA

library(recipes)
seven_pt <- recipe(~ . , data = example_data) %>%
  add_role(day, new_role = "time_index") %>%
  step_rollimpute(all_predictors(), window = 7) %>%
  prep(training = example_data, retain = TRUE)

#> Warning: Changing role(s) for day

juice(seven_pt)

#> # A tibble: 12 x 4
#>    day            x1    x2     x3
#>    <date>      <dbl> <dbl>  <dbl>
#>  1 2012-06-08 0.890  0.790 0.580 
#>  2 2012-06-09 0.530  0.790 0.450 
#>  3 2012-06-10 0.860  0.790 0.670 
#>  4 2012-06-11 0.920  0.790 0.0500
#>  5 2012-06-12 0.860  0.880 0.270 
#>  6 2012-06-13 0.860  0.600 0.130 
#>  7 2012-06-14 0.970  0.790 0.670 
#>  8 2012-06-15 0.850  0.270 0.160 
#>  9 2012-06-16 0.860  0.110 0.360 
#> 10 2012-06-17 0.960  0.435 0.0900
#> 11 2012-06-18 0.910  0.160 0.760 
#> 12 2012-06-19 0.0700 0.860 0.740