Shuffle Variables

step_shuffle creates a specification of a recipe step that will randomly change the order of rows for selected variables.

step_shuffle(recipe, ..., role = NA, trained = FALSE, columns = NULL,
  skip = FALSE, id = rand_id("shuffle"))

# S3 method for step_shuffle
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables will permuted. See selections() for more details. For the tidy method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

columns
A character string that contains the names of columns that should be shuffled. These values are not determined until prep.recipe() is called.

skip
A logical. Should the step be skipped when the recipe is baked by bake.recipe()? While all operations are baked when prep.recipe() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations

id
A character string that is unique to this step to identify it.

x
A step_shuffle object.

Value

An updated version of recipe with the new step added to the sequence of existing steps (if any). For the tidy method, a tibble with columns terms which is the columns that will be affected.

Examples

integers <- data.frame(A = 1:12, B = 13:24, C = 25:36)

library(dplyr)
rec <- recipe(~ A + B + C, data = integers) %>%
  step_shuffle(A, B)

rand_set <- prep(rec, training = integers)

set.seed(5377)
bake(rand_set, integers)

#> # A tibble: 12 x 3
#>        A     B     C
#>    <int> <int> <int>
#>  1     5    13    25
#>  2     4    24    26
#>  3     9    23    27
#>  4     1    18    28
#>  5    11    14    29
#>  6    10    21    30
#>  7     6    19    31
#>  8     3    22    32
#>  9    12    16    33
#> 10     7    15    34
#> 11     2    17    35
#> 12     8    20    36

tidy(rec, number = 1)

#> # A tibble: 2 x 2
#>   terms id           
#>   <chr> <chr>        
#> 1 A     shuffle_mjSqT
#> 2 B     shuffle_mjSqT

tidy(rand_set, number = 1)

#> # A tibble: 2 x 2
#>   terms id           
#>   <chr> <chr>        
#> 1 A     shuffle_mjSqT
#> 2 B     shuffle_mjSqT