Convert Strings to Factors

`step_string2factor` will convert one or more character vectors to factors (ordered or unordered).

step_string2factor(recipe, ..., role = NA, trained = FALSE, levels = NULL,
  ordered = FALSE, skip = FALSE)

# S3 method for step_string2factor
tidy(x, ...)

Arguments

recipe
A recipe object. The step will be added to the sequence of operations for this recipe.

...
One or more selector functions to choose which variables will converted to factors. See [selections()] for more details. For the `tidy` method, these are not currently used.

role
Not used by this step since no new variables are created.

trained
A logical to indicate if the quantities for preprocessing have been estimated.

levels
An options specification of the levels to be used for the new factor. If left `NULL`, the sorted unique values present when `bake` is called will be used.

ordered
A single logical value; should the factor(s) be ordered?

skip
A logical. Should the step be skipped when the recipe is baked by [bake.recipe()]? While all operations are baked when [prep.recipe()] is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using `skip = TRUE` as it may affect the computations for subsequent operations

x
A `step_string2factor` object.

Value

An updated version of `recipe` with the new step added to the sequence of existing steps (if any). For the `tidy` method, a tibble with columns `terms` (the selectors or variables selected) and `ordered`.

Details

If `levels` is given, `step_string2factor` will convert all factors to have the same levels. Also, note that `prep` has an option `stringsAsFactors` that defaults to `TRUE`. This should be changed so that raw character data will be applied to `step_string2factor`.

See also

Examples

data(okc)

rec <- recipe(~ diet + location, data = okc)

make_factor <- rec %>%
  step_string2factor(diet)
make_factor <- prep(make_factor,
                    training = okc,
                    stringsAsFactors = FALSE,
                    retain = TRUE)

# note that `diet` is a factor
juice(make_factor) %>% head

#> # A tibble: 6 x 2
#>   diet              location                       
#>   <fct>             <chr>                          
#> 1 strictly anything south san francisco, california
#> 2 mostly other      oakland, california            
#> 3 anything          san francisco, california      
#> 4 vegetarian        berkeley, california           
#> 5 <NA>              san francisco, california      
#> 6 mostly anything   san francisco, california      

okc %>% head

#> # A tibble: 6 x 5
#>     age diet              height location                        date      
#>   <int> <chr>              <int> <chr>                           <date>    
#> 1    22 strictly anything     75 south san francisco, california 2012-06-28
#> 2    35 mostly other          70 oakland, california             2012-06-29
#> 3    38 anything              68 san francisco, california       2012-06-27
#> 4    23 vegetarian            71 berkeley, california            2012-06-28
#> 5    29 <NA>                  66 san francisco, california       2012-06-27
#> 6    29 mostly anything       67 san francisco, california       2012-06-29

tidy(make_factor, number = 1)

#> # A tibble: 1 x 2
#>   terms ordered
#>   <chr> <lgl>  
#> 1 diet  FALSE