Summarize a Recipe

This function prints the current set of variables/features and some of their characteristics.

# S3 method for recipe
summary(object, original = FALSE, ...)

Arguments

object

A recipe object

original

A logical: show the current set of variables or the original set when the recipe was defined.

...

further arguments passed to or from other methods (not currently used).

Value

A tibble with columns variable, type, role, and source.

Details

Note that, until the recipe has been learned, the currrent and original variables are the same.

See also

recipe parse_terms_formula learn.recipe

Examples

rec <- recipe( ~ ., data = USArrests)
summary(rec)

#> # A tibble: 4 × 4
#>   variable    type      role   source
#>      <chr>   <chr>     <chr>    <chr>
#> 1   Murder numeric predictor original
#> 2  Assault numeric predictor original
#> 3 UrbanPop numeric predictor original
#> 4     Rape numeric predictor original

rec <- step_pca(rec, terms = ~ all_numeric(), num = 3)
summary(rec) # still the same since not yet learned

#> # A tibble: 4 × 4
#>   variable    type      role   source
#>      <chr>   <chr>     <chr>    <chr>
#> 1   Murder numeric predictor original
#> 2  Assault numeric predictor original
#> 3 UrbanPop numeric predictor original
#> 4     Rape numeric predictor original

rec <- learn(rec, training = USArrests)

#> step 1 pca training 

#> Error in x$variable: $ operator is invalid for atomic vectors

summary(rec)

#> # A tibble: 4 × 4
#>   variable    type      role   source
#>      <chr>   <chr>     <chr>    <chr>
#> 1   Murder numeric predictor original
#> 2  Assault numeric predictor original
#> 3 UrbanPop numeric predictor original
#> 4     Rape numeric predictor original