Summarize a Recipe

This function prints the current set of variables/features and some of their characteristics.

# S3 method for recipe
summary(object, original = FALSE, ...)

Arguments

object
A recipe object

original
A logical: show the current set of variables or the original set when the recipe was defined.

...
further arguments passed to or from other methods (not currently used).

Value

A tibble with columns variable, type, role, and source.

Details

Note that, until the recipe has been trained, the current and original variables are the same.

See also

Examples

rec <- recipe( ~ ., data = USArrests)
summary(rec)

#> [38;5;246m# A tibble: 4 x 4[39m
#>   variable type    role      source  
#>   [3m[38;5;246m<chr>[39m[23m    [3m[38;5;246m<chr>[39m[23m   [3m[38;5;246m<chr>[39m[23m     [3m[38;5;246m<chr>[39m[23m   
#> [38;5;250m1[39m Murder   numeric predictor original
#> [38;5;250m2[39m Assault  numeric predictor original
#> [38;5;250m3[39m UrbanPop numeric predictor original
#> [38;5;250m4[39m Rape     numeric predictor original

rec <- step_pca(rec, all_numeric(), num = 3)

#> The argument `num` is deprecated in factor of `num_comp`. `num` will be removed in next version.FALSE

summary(rec) # still the same since not yet trained

#> [38;5;246m# A tibble: 4 x 4[39m
#>   variable type    role      source  
#>   [3m[38;5;246m<chr>[39m[23m    [3m[38;5;246m<chr>[39m[23m   [3m[38;5;246m<chr>[39m[23m     [3m[38;5;246m<chr>[39m[23m   
#> [38;5;250m1[39m Murder   numeric predictor original
#> [38;5;250m2[39m Assault  numeric predictor original
#> [38;5;250m3[39m UrbanPop numeric predictor original
#> [38;5;250m4[39m Rape     numeric predictor original

rec <- prep(rec, training = USArrests)
summary(rec)

#> [38;5;246m# A tibble: 4 x 4[39m
#>   variable type    role      source 
#>   [3m[38;5;246m<chr>[39m[23m    [3m[38;5;246m<chr>[39m[23m   [3m[38;5;246m<chr>[39m[23m     [3m[38;5;246m<chr>[39m[23m  
#> [38;5;250m1[39m PC1      numeric predictor derived
#> [38;5;250m2[39m PC2      numeric predictor derived
#> [38;5;250m3[39m PC3      numeric predictor derived
#> [38;5;250m4[39m PC4      numeric predictor derived