Summarize a Recipe

This function prints the current set of variables/features and some of their characteristics.

# S3 method for recipe
summary(object, original = FALSE, ...)

Arguments

object
A recipe object

original
A logical: show the current set of variables or the original set when the recipe was defined.

...
further arguments passed to or from other methods (not currently used).

Value

A tibble with columns variable, type, role, and source.

Details

Note that, until the recipe has been trained, the current and original variables are the same.

It is possible for variables to have multiple roles by adding them with add_role(). If a variable has multiple roles, it will have more than one row in the summary tibble.

See also

Examples

rec <- recipe( ~ ., data = USArrests)
summary(rec)

#> # A tibble: 4 x 4
#>   variable type    role      source  
#>   <chr>    <chr>   <chr>     <chr>   
#> 1 Murder   numeric predictor original
#> 2 Assault  numeric predictor original
#> 3 UrbanPop numeric predictor original
#> 4 Rape     numeric predictor original

rec <- step_pca(rec, all_numeric(), num_comp = 3)
summary(rec) # still the same since not yet trained

#> # A tibble: 4 x 4
#>   variable type    role      source  
#>   <chr>    <chr>   <chr>     <chr>   
#> 1 Murder   numeric predictor original
#> 2 Assault  numeric predictor original
#> 3 UrbanPop numeric predictor original
#> 4 Rape     numeric predictor original

rec <- prep(rec, training = USArrests)
summary(rec)

#> # A tibble: 3 x 4
#>   variable type    role      source 
#>   <chr>    <chr>   <chr>     <chr>  
#> 1 PC1      numeric predictor derived
#> 2 PC2      numeric predictor derived
#> 3 PC3      numeric predictor derived