Select Terms in a Step Function.

terms_select() has been deprecated in favor of recipes_eval_select().

This function bakes the step function selectors and might be useful when creating custom steps.

Usage

terms_select(terms, info, empty_fun = abort_selection)

Arguments

terms

A list of formulas whose right-hand side contains quoted expressions. See rlang::quos() for examples.

info

A tibble with columns variable, type, role, and source that represent the current state of the data. The function summary.recipe() can be used to get this information from a recipe.

empty_fun

A function to execute when no terms are selected by the step. The default function throws an error with a message.

Value

A character string of column names or an error of there are no selectors or if no variables are selected.

See also

Examples

library(rlang)
data(Sacramento, package = "modeldata")
rec <- recipe(~., data = Sacramento)
info <- summary(rec)
terms_select(info = info, quos(all_predictors()))
#> Warning: `terms_select()` was deprecated in recipes 0.1.17.
#> Please use `recipes_eval_select()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
#> [1] "city"      "zip"       "beds"      "baths"     "sqft"     
#> [6] "type"      "price"     "latitude"  "longitude"