Tidy the Result of a Recipe

`tidy` will return a data frame that contains information regarding a recipe or operation within the recipe (when a `tidy` method for the operation exists).

# S3 method for recipe
tidy(x, number = NA, ...)

Arguments

x
A `recipe` object (trained or otherwise).

number
An integer or `NA`. If missing, the return value is a list of the operation in the recipe. If a number is given, a `tidy` method is executed for that operation in the recipe (if it exists).

...
Not currently used.

Value

A tibble with columns that would vary depending on what `tidy` method is executed. When `x` is `NA`, a tibble with columns `number` (the operation iteration), `operation` (either "step" or "check"), `type` (the method, e.g. "nzv", "center"), a logical column called `trained` for whether the operation has been estimated using `prep`, and a logical for `skip`.

Examples

data(okc)

okc_rec <- recipe(~ ., data = okc) %>%
  step_other(all_nominal(), threshold = 0.05) %>%
  step_date(date, features = "dow") %>%
  step_center(all_numeric()) %>%
  step_dummy(all_nominal()) %>%
  check_cols(starts_with("date"), age, height)

tidy(okc_rec)

#> # A tibble: 5 x 5
#>   number operation type   trained skip 
#>    <int> <chr>     <chr>  <lgl>   <lgl>
#> 1      1 step      other  FALSE   FALSE
#> 2      2 step      date   FALSE   FALSE
#> 3      3 step      center FALSE   FALSE
#> 4      4 step      dummy  FALSE   FALSE
#> 5      5 check     cols   FALSE   FALSE

tidy(okc_rec, number = 2)

#> # A tibble: 1 x 3
#>   terms value ordinal
#>   <fct> <fct> <lgl>  
#> 1 date  dow   FALSE  

tidy(okc_rec, number = 3)

#> # A tibble: 1 x 2
#>   terms         value
#>   <chr>         <dbl>
#> 1 all_numeric()    NA

okc_rec_trained <- prep(okc_rec, training = okc)

tidy(okc_rec_trained)

#> # A tibble: 5 x 5
#>   number operation type   trained skip 
#>    <int> <chr>     <chr>  <lgl>   <lgl>
#> 1      1 step      other  TRUE    FALSE
#> 2      2 step      date   TRUE    FALSE
#> 3      3 step      center TRUE    FALSE
#> 4      4 step      dummy  TRUE    FALSE
#> 5      5 check     cols   TRUE    FALSE

tidy(okc_rec_trained, number = 3)

#> # A tibble: 2 x 2
#>   terms  value
#>   <chr>  <dbl>
#> 1 age     32.3
#> 2 height  68.3