Tidy the Result of a Recipe

tidy will return a data frame that contains information regarding a recipe or operation within the recipe (when a tidy method for the operation exists).

# S3 method for recipe
tidy(x, number = NA, ...)

# S3 method for step
tidy(x, ...)

# S3 method for check
tidy(x, ...)

Arguments

x
A recipe object (trained or otherwise).

number
An integer or NA. If missing, the return value is a list of the operation in the recipe. If a number is given, a tidy method is executed for that operation in the recipe (if it exists).

...
Not currently used.

Value

A tibble with columns that would vary depending on what tidy method is executed. When x is NA, a tibble with columns number (the operation iteration), operation (either "step" or "check"), type (the method, e.g. "nzv", "center"), a logical column called trained for whether the operation has been estimated using prep, a logical for skip, and a character column id.

Examples

data(okc)

okc_rec <- recipe(~ ., data = okc) %>%
  step_other(all_nominal(), threshold = 0.05, other = "another") %>%
  step_date(date, features = "dow") %>%
  step_center(all_numeric()) %>%
  step_dummy(all_nominal()) %>%
  check_cols(starts_with("date"), age, height)

tidy(okc_rec)

#> # A tibble: 5 x 6
#>   number operation type   trained skip  id          
#>    <int> <chr>     <chr>  <lgl>   <lgl> <chr>       
#> 1      1 step      other  FALSE   FALSE other_g6xzQ 
#> 2      2 step      date   FALSE   FALSE date_2bArV  
#> 3      3 step      center FALSE   FALSE center_ccO0Y
#> 4      4 step      dummy  FALSE   FALSE dummy_ILC16 
#> 5      5 check     cols   FALSE   FALSE cols_dWoJD  

tidy(okc_rec, number = 2)

#> # A tibble: 1 x 4
#>   terms value ordinal id        
#>   <fct> <fct> <lgl>   <chr>     
#> 1 date  dow   FALSE   date_2bArV

tidy(okc_rec, number = 3)

#> # A tibble: 1 x 3
#>   terms         value id          
#>   <chr>         <dbl> <chr>       
#> 1 all_numeric()    NA center_ccO0Y

okc_rec_trained <- prep(okc_rec, training = okc)

#> Warning: `overscope_eval_next()` is soft-deprecated as of rlang 0.2.0.
#> Please use `eval_tidy()` with a data mask instead
#> This warning is displayed once per session.

#> Warning: There are new levels in a factor: NA

tidy(okc_rec_trained)

#> # A tibble: 5 x 6
#>   number operation type   trained skip  id          
#>    <int> <chr>     <chr>  <lgl>   <lgl> <chr>       
#> 1      1 step      other  TRUE    FALSE other_g6xzQ 
#> 2      2 step      date   TRUE    FALSE date_2bArV  
#> 3      3 step      center TRUE    FALSE center_ccO0Y
#> 4      4 step      dummy  TRUE    FALSE dummy_ILC16 
#> 5      5 check     cols   TRUE    FALSE cols_dWoJD  

tidy(okc_rec_trained, number = 3)

#> # A tibble: 2 x 3
#>   terms  value id          
#>   <chr>  <dbl> <chr>       
#> 1 age     32.3 center_ccO0Y
#> 2 height  68.3 center_ccO0Y