Find recommended methods for generating parameter values

tunable() determines which parameters in an object can be tuned along with information about the parameters.

# S3 method for recipe
tunable(x, ...)

# S3 method for step
tunable(x, ...)

# S3 method for check
tunable(x, ...)

Arguments

x
A recipe, recipe step, or recipe check object

...
Not currently used.

Value

A tibble with a column for the parameter name, information on the default method for generating a corresponding parameter object, the source of the parameter (e.g. "recipe", etc.), and the component within the source. For the component column, a little more specificity is given about the location of the parameter (e.g. "step_normalize" for recipes). The component_id column contains the unique step id field.

Details

These functions define what parameters can be tuned for specific steps. They also define the recommended objects from the dials package that can be used to generate new parameter values and other characteristics.

If the object has no tunable parameters, a tibble with no rows is returned.

The information about the default parameter object takes the form of a named list with an element for the function call and an optional element for the source of the function (e.g. the dials package).

Examples

# \donttest{
library(recipes)

recipe(mpg ~ ., data = mtcars) %>%
  step_knnimpute(all_predictors()) %>%
  step_pca(all_predictors()) %>%
  tunable()

#> # A tibble: 2 x 5
#>   name      call_info        source component      component_id   
#>   <chr>     <list>           <chr>  <chr>          <chr>          
#> 1 neighbors <named list [3]> recipe step_knnimpute knnimpute_hw1Dj
#> 2 num_comp  <named list [3]> recipe step_pca       pca_nADLd      

recipe(mpg ~ ., data = mtcars) %>%
  step_normalize(all_predictors()) %>%
  tunable()

#> # A tibble: 0 x 5
#> # … with 5 variables: name <chr>, call_info <list>, source <chr>,
#> #   component <chr>, component_id <chr>

# }